@@ -1,14 +1,12 @@
 module github.com/TarsCloud/TarsGo
 
-go 1.13
+go 1.14
 
 require (
-	github.com/gin-gonic/gin v1.8.1
-	github.com/google/go-cmp v0.5.9 // indirect
-	github.com/google/uuid v1.3.0
-	github.com/opentracing/opentracing-go v1.1.0
-	github.com/openzipkin-contrib/zipkin-go-opentracing v0.4.5
-	github.com/openzipkin/zipkin-go v0.4.0
-	github.com/stretchr/testify v1.8.2
-	go.uber.org/automaxprocs v1.5.1
+	github.com/google/uuid v1.6.0
+	github.com/kr/pretty v0.3.0 // indirect
+	github.com/rogpeppe/go-internal v1.8.0 // indirect
+	github.com/stretchr/testify v1.8.4
+	go.uber.org/automaxprocs v1.5.2
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 )
