@@ -17,12 +17,15 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/metric/instrument"
 	"go.opentelemetry.io/otel/propagation"
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 	"go.opentelemetry.io/otel/trace"
 )
 
 const (
 	instrumentationName = "github.com/TarsCloud/TarsGo/tars/middleware/opentelemetry"
+	masterName          = "TARS_MASTER_NAME"
 	TarsRpcRetKey       = attribute.Key("tars.rpc.ret")
 )
 
@@ -30,6 +33,9 @@ type Opentelemetry struct {
 	Propagators    propagation.TextMapPropagator
 	TracerProvider trace.TracerProvider
 	MeterProvider  metric.MeterProvider
+
+	meter             metric.Meter
+	rpcServerDuration instrument.Int64Histogram
 }
 
 type Option func(*Opentelemetry)
@@ -46,14 +52,32 @@ func WithPropagators(p propagation.TextMapPropagator) Option {
 	}
 }
 
+func WithMeterProvider(mp metric.MeterProvider) Option {
+	return func(o *Opentelemetry) {
+		o.MeterProvider = mp
+	}
+}
+
 func New(opts ...Option) *Opentelemetry {
 	o := &Opentelemetry{
 		TracerProvider: otel.GetTracerProvider(),
 		Propagators:    otel.GetTextMapPropagator(),
+		MeterProvider:  otel.GetMeterProvider(),
 	}
 	for _, opt := range opts {
 		opt(o)
 	}
+
+	o.meter = o.MeterProvider.Meter(
+		instrumentationName,
+		metric.WithInstrumentationVersion(SemVersion()),
+		metric.WithSchemaURL(semconv.SchemaURL),
+	)
+	var err error
+	if o.rpcServerDuration, err = o.meter.Int64Histogram("tars.server.duration", instrument.WithUnit("ms")); err != nil {
+		otel.Handle(err)
+	}
+
 	return o
 }
 
@@ -66,26 +90,39 @@ func (o *Opentelemetry) BuildServerFilter() tars.ServerFilterMiddleware {
 			port, _ := current.GetClientPortFromContext(ctx)
 			var span trace.Span
 			ctx = o.extract(ctx, req)
-			servants := strings.Split(req.SServantName, ".")
-			ctx, span = tracer.Start(
-				ctx,
-				fmt.Sprintf("%s.%s", servants[2], req.SFuncName),
-				trace.WithSpanKind(trace.SpanKindServer),
-				trace.WithAttributes(
-					attribute.String("tars.interface", req.SServantName),
-					attribute.String("tars.method", req.SFuncName),
-					attribute.String("tars.local_ip", localIp),
-					attribute.Int("tars.request_id", int(req.IRequestId)),
-					attribute.String("tars.client.ipv4", ip),
-					attribute.String("tars.client.port", port),
-					attribute.String("tars.server.version", tars.Version),
-				),
-			)
-			defer span.End()
+			index := strings.LastIndex(req.SServantName, ".")
+			attrs := []attribute.KeyValue{
+				attribute.String("tars.master.name", req.Status[masterName]),
+				attribute.String("tars.master.ip", ip),
+				attribute.String("tars.slave.name", req.SServantName[:index]),
+				attribute.String("tars.slave.ip", localIp),
+				attribute.String("tars.interface", req.SServantName),
+				attribute.String("tars.method", req.SFuncName),
+				attribute.String("tars.version", tars.Version),
+			}
 			cfg := tars.GetServerConfig()
 			if cfg.Enableset {
-				span.SetAttributes(attribute.String("tars.set_division", cfg.Setdivision))
+				attrs = append(attrs, attribute.String("tars.set_division", cfg.Setdivision))
 			}
+			ctx, span = tracer.Start(
+				ctx,
+				fmt.Sprintf("%s.%s", req.SServantName[index+1:], req.SFuncName),
+				trace.WithSpanKind(trace.SpanKindServer),
+				trace.WithAttributes(attrs...),
+			)
+			span.SetAttributes(
+				attribute.String("tars.client.port", port),
+				attribute.Int("tars.request.id", int(req.IRequestId)),
+			)
+			defer span.End()
+
+			var statusCode int32
+			defer func(t time.Time) {
+				elapsedTime := time.Since(t) / time.Millisecond
+				attrs = append(attrs, TarsRpcRetKey.Int64(int64(statusCode)))
+				o.rpcServerDuration.Record(ctx, int64(elapsedTime), attrs...)
+			}(time.Now())
+
 			err = next(ctx, d, f, req, resp, withContext)
 			if err != nil {
 				span.SetStatus(codes.Error, "server failed")
@@ -113,21 +150,32 @@ func (o *Opentelemetry) BuildHttpHandler() func(next http.Handler) http.Handler 
 			var span trace.Span
 			reqCtx := r.Context()
 			reqCtx = o.Propagators.Extract(reqCtx, propagation.HeaderCarrier(r.Header))
+			attrs := []attribute.KeyValue{
+				attribute.String("http.method", r.Method),
+				attribute.String("http.path", r.URL.Path),
+				attribute.String("http.scheme", r.URL.Scheme),
+				attribute.String("http.proto", r.Proto),
+				attribute.String("component", "web"),
+			}
 			reqCtx, span = tracer.Start(
 				reqCtx,
 				r.URL.Path,
 				trace.WithSpanKind(trace.SpanKindServer),
-				trace.WithAttributes(
-					attribute.String("http.method", r.Method),
-					attribute.String("http.url", r.URL.String()),
-					attribute.String("http.scheme", r.URL.Scheme),
-					attribute.String("http.proto", r.Proto),
-					attribute.String("peer.hostname", r.Host),
-					attribute.String("peer.address", r.RemoteAddr),
-					attribute.String("component", "web"),
-				),
+				trace.WithAttributes(attrs...),
+			)
+			span.SetAttributes(
+				attribute.String("http.url", r.URL.String()),
+				attribute.String("peer.hostname", r.Host),
+				attribute.String("peer.address", r.RemoteAddr),
 			)
 			defer span.End()
+
+			var statusCode int32
+			defer func(t time.Time) {
+				elapsedTime := time.Since(t) / time.Millisecond
+				attrs = append(attrs, TarsRpcRetKey.Int64(int64(statusCode)))
+				o.rpcServerDuration.Record(reqCtx, int64(elapsedTime), attrs...)
+			}(time.Now())
 
 			r = r.WithContext(reqCtx)
 			recorder := httptest.NewRecorder()
@@ -137,6 +185,7 @@ func (o *Opentelemetry) BuildHttpHandler() func(next http.Handler) http.Handler 
 				w.Header()[k] = v
 			}
 			w.WriteHeader(recorder.Code)
+			statusCode = int32(recorder.Code)
 			_, err := w.Write(recorder.Body.Bytes())
 			if err != nil {
 				span.SetStatus(codes.Error, "http server write failed")
@@ -158,20 +207,20 @@ func (o *Opentelemetry) BuildClientFilter() tars.ClientFilterMiddleware {
 				fmt.Sprintf("%s.%s", servants[2], msg.Req.SFuncName),
 				trace.WithSpanKind(trace.SpanKindClient),
 				trace.WithAttributes(
+					attribute.String("tars.master.ip", localIp),
 					attribute.String("tars.interface", msg.Req.SServantName),
 					attribute.String("tars.method", msg.Req.SFuncName),
-					attribute.Int("tars.request_id", int(msg.Req.IRequestId)),
-					attribute.String("tars.local_ip", localIp),
 					attribute.String("tars.protocol", "tars"),
-					attribute.String("tars.client.version", tars.Version),
+					attribute.String("tars.version", tars.Version),
+					attribute.Int("tars.request.id", int(msg.Req.IRequestId)),
 				),
 			)
 			ctx = o.inject(ctx, msg)
 			defer func() {
 				ip, _ := current.GetServerIPFromContext(ctx)
 				port, _ := current.GetServerPortFromContext(ctx)
-				span.SetAttributes(attribute.String("tars.server.ipv4", ip))
-				span.SetAttributes(attribute.String("tars.server.port", port))
+				span.SetAttributes(attribute.String("tars.slave.ip", ip))
+				span.SetAttributes(attribute.String("tars.slave.port", port))
 				span.End()
 			}()
 
@@ -193,6 +242,9 @@ func (o *Opentelemetry) inject(ctx context.Context, msg *tars.Message) context.C
 		msg.Req.Status = make(map[string]string)
 	}
 	o.Propagators.Inject(ctx, propagation.MapCarrier(msg.Req.Status))
+	// inject into the module Name
+	cfg := tars.GetClientConfig()
+	msg.Req.Status[masterName] = cfg.ModuleName
 	return ctx
 }
 
