@@ -1,14 +1,14 @@
 package main
 
 import (
+	cgin "github.com/TarsCloud/TarsGo/contrib/gin"
 	"github.com/TarsCloud/TarsGo/tars"
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
-	mux := &tars.TarsHttpMux{}
-	r := mux.GetGinEngine()
-	r.GET("/ping", func(c *gin.Context) {
+	g := cgin.New()
+	g.GET("/ping", func(c *gin.Context) {
 		c.JSON(200, gin.H{
 			"message": "pong",
 		})
@@ -16,6 +16,6 @@ func main() {
 
 	// Get server config
 	cfg := tars.GetServerConfig()
-	tars.AddHttpServant(mux, cfg.App+"."+cfg.Server+".HttpObj")
+	tars.AddHttpServant(g, cfg.App+"."+cfg.Server+".HttpObj")
 	tars.Run()
 }
