@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"log"
 
+	"OpentelemetryServer/otel"
 	"OpentelemetryServer/tars-protocol/StressTest"
-	"OpentelemetryServer/tracer"
 
 	"github.com/TarsCloud/TarsGo/contrib/middleware/opentelemetry"
 	"github.com/TarsCloud/TarsGo/tars"
@@ -15,17 +15,23 @@ import (
 func main() {
 	cfg := tars.GetServerConfig()
 	serviceNameKey := fmt.Sprintf("%s.%s", cfg.App, cfg.Server)
-	tp := tracer.NewTracerProvider(serviceNameKey, "")
+	tp := otel.NewTracerProvider(serviceNameKey, "")
+	mp := otel.NewMeterProvider(serviceNameKey, "")
 	defer func() {
+		if err := mp.Shutdown(context.Background()); err != nil {
+			log.Printf("Error shutting down metrics provider: %v", err)
+		}
+
 		if err := tp.Shutdown(context.Background()); err != nil {
-			log.Printf("Error shutting down tracer provider: %v", err)
+			log.Printf("Error shutting down traces provider: %v", err)
 		}
 	}()
+
 	filter := opentelemetry.New()
 	tars.UseServerFilterMiddleware(filter.BuildServerFilter())
 	tars.UseClientFilterMiddleware(filter.BuildClientFilter())
 	imp := new(OpentelemetryImp)                                               //New Imp
-	app := new(StressTest.ContextTest)                                         //New init the A Tars
+	app := new(StressTest.Opentelemetry)                                       //New init the A Tars
 	app.AddServantWithContext(imp, cfg.App+"."+cfg.Server+".OpenTelemetryObj") //Register Servant
 	tars.Run()
 }
