@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from AdminF.tars
 // Package adminf comment
-// This file was generated by tars2go 1.1.10
-// Generated from AdminF.tars
 package adminf
 
 import (
@@ -8,79 +8,67 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	m "github.com/TarsCloud/TarsGo/tars/model"
+
+	"github.com/TarsCloud/TarsGo/tars/model"
 	"github.com/TarsCloud/TarsGo/tars/protocol/codec"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/requestf"
 	"github.com/TarsCloud/TarsGo/tars/protocol/tup"
 	"github.com/TarsCloud/TarsGo/tars/util/current"
+	"github.com/TarsCloud/TarsGo/tars/util/endpoint"
 	"github.com/TarsCloud/TarsGo/tars/util/tools"
-	"unsafe"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var (
 	_ = fmt.Errorf
 	_ = codec.FromInt8
-	_ = unsafe.Pointer(nil)
 	_ = bytes.ErrTooLarge
 )
 
+type AdminFServant interface {
+	Shutdown() (err error)
+	Notify(command string) (ret string, err error)
+}
+
+type AdminFServantWithContext interface {
+	Shutdown(tarsCtx context.Context) (err error)
+	Notify(tarsCtx context.Context, command string) (ret string, err error)
+}
+
 // AdminF struct
 type AdminF struct {
-	servant m.Servant
+	servant model.Servant
+}
+
+// NewAdminF creates a new AdminF servant.
+func NewAdminF() *AdminF {
+	return new(AdminF)
+}
+
+// SetServant sets servant for the service.
+func (obj *AdminF) SetServant(servant model.Servant) {
+	obj.servant = servant
+}
+
+// TarsSetTimeout sets the timeout for the servant which is in ms.
+func (obj *AdminF) TarsSetTimeout(timeout int) {
+	obj.servant.TarsSetTimeout(timeout)
+}
+
+// TarsSetProtocol sets the protocol for the servant.
+func (obj *AdminF) TarsSetProtocol(p model.Protocol) {
+	obj.servant.TarsSetProtocol(p)
+}
+
+// Endpoints returns all active endpoint.Endpoint
+func (obj *AdminF) Endpoints() []*endpoint.Endpoint {
+	return obj.servant.Endpoints()
 }
 
 // Shutdown is the proxy function for the method defined in the tars file, with the context
-func (obj *AdminF) Shutdown(opts ...map[string]string) (err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "shutdown", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return nil
+func (obj *AdminF) Shutdown(opts ...map[string]string) error {
+	return obj.ShutdownWithContext(context.Background(), opts...)
 }
 
 // ShutdownWithContext is the proxy function for the method defined in the tars file, with the context
@@ -91,6 +79,7 @@ func (obj *AdminF) ShutdownWithContext(tarsCtx context.Context, opts ...map[stri
 		ty     byte
 	)
 	buf := codec.NewBuffer()
+
 	var statusMap map[string]string
 	var contextMap map[string]string
 	if len(opts) == 1 {
@@ -127,6 +116,7 @@ func (obj *AdminF) ShutdownWithContext(tarsCtx context.Context, opts ...map[stri
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -141,6 +131,7 @@ func (obj *AdminF) ShutdownOneWayWithContext(tarsCtx context.Context, opts ...ma
 		ty     byte
 	)
 	buf := codec.NewBuffer()
+
 	var statusMap map[string]string
 	var contextMap map[string]string
 	if len(opts) == 1 {
@@ -156,27 +147,6 @@ func (obj *AdminF) ShutdownOneWayWithContext(tarsCtx context.Context, opts ...ma
 		return err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -184,65 +154,8 @@ func (obj *AdminF) ShutdownOneWayWithContext(tarsCtx context.Context, opts ...ma
 }
 
 // Notify is the proxy function for the method defined in the tars file, with the context
-func (obj *AdminF) Notify(command string, opts ...map[string]string) (ret string, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(command, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "notify", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadString(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *AdminF) Notify(command string, opts ...map[string]string) (string, error) {
+	return obj.NotifyWithContext(context.Background(), command, opts...)
 }
 
 // NotifyWithContext is the proxy function for the method defined in the tars file, with the context
@@ -272,7 +185,6 @@ func (obj *AdminF) NotifyWithContext(tarsCtx context.Context, command string, op
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadString(&ret, 0, true)
 	if err != nil {
@@ -300,6 +212,7 @@ func (obj *AdminF) NotifyWithContext(tarsCtx context.Context, command string, op
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -334,55 +247,10 @@ func (obj *AdminF) NotifyOneWayWithContext(tarsCtx context.Context, command stri
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
 	return ret, nil
-}
-
-// SetServant sets servant for the service.
-func (obj *AdminF) SetServant(servant m.Servant) {
-	obj.servant = servant
-}
-
-// TarsSetTimeout sets the timeout for the servant which is in ms.
-func (obj *AdminF) TarsSetTimeout(timeout int) {
-	obj.servant.TarsSetTimeout(timeout)
-}
-
-// TarsSetProtocol sets the protocol for the servant.
-func (obj *AdminF) TarsSetProtocol(p m.Protocol) {
-	obj.servant.TarsSetProtocol(p)
-}
-
-type AdminFServant interface {
-	Shutdown() (err error)
-	Notify(command string) (ret string, err error)
-}
-type AdminFServantWithContext interface {
-	Shutdown(tarsCtx context.Context) (err error)
-	Notify(tarsCtx context.Context, command string) (ret string, err error)
 }
 
 // Dispatch is used to call the server side implement for the method defined in the tars file. withContext shows using context or not.
@@ -396,7 +264,6 @@ func (obj *AdminF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 	buf := codec.NewBuffer()
 	switch tarsReq.SFuncName {
 	case "shutdown":
-
 		if !withContext {
 			imp := val.(AdminFServant)
 			err = imp.Shutdown()
@@ -404,7 +271,6 @@ func (obj *AdminF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(AdminFServantWithContext)
 			err = imp.Shutdown(tarsCtx)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -436,14 +302,11 @@ func (obj *AdminF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		}
 	case "notify":
 		var command string
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&command, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -456,7 +319,6 @@ func (obj *AdminF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -471,7 +333,6 @@ func (obj *AdminF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -485,7 +346,6 @@ func (obj *AdminF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(AdminFServantWithContext)
 			funRet, err = imp.Notify(tarsCtx, command)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -497,7 +357,6 @@ func (obj *AdminF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
