@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from BaseF.tars
 // Package basef comment
-// This file was generated by tars2go 1.1.10
-// Generated from BaseF.tars
 package basef
 
 import (
@@ -13,7 +13,7 @@ import (
 var _ = fmt.Errorf
 var _ = codec.FromInt8
 
-//const as define in tars file
+// const as define in tars file
 const (
 	TARSVERSION             int16 = 0x01
 	TUPVERSION              int16 = 0x03
