@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from ConfigF.tars
 // Package configf comment
-// This file was generated by tars2go 1.1.10
-// Generated from ConfigF.tars
 package configf
 
 import (
@@ -8,147 +8,79 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	m "github.com/TarsCloud/TarsGo/tars/model"
+
+	"github.com/TarsCloud/TarsGo/tars/model"
 	"github.com/TarsCloud/TarsGo/tars/protocol/codec"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/requestf"
 	"github.com/TarsCloud/TarsGo/tars/protocol/tup"
 	"github.com/TarsCloud/TarsGo/tars/util/current"
+	"github.com/TarsCloud/TarsGo/tars/util/endpoint"
 	"github.com/TarsCloud/TarsGo/tars/util/tools"
-	"unsafe"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var (
 	_ = fmt.Errorf
 	_ = codec.FromInt8
-	_ = unsafe.Pointer(nil)
 	_ = bytes.ErrTooLarge
 )
 
+type ConfigServant interface {
+	ListConfig(app string, server string, vf *[]string) (ret int32, err error)
+	LoadConfig(app string, server string, filename string, config *string) (ret int32, err error)
+	LoadConfigByHost(appServerName string, filename string, host string, config *string) (ret int32, err error)
+	CheckConfig(appServerName string, filename string, host string, result *string) (ret int32, err error)
+	ListConfigByInfo(configInfo *ConfigInfo, vf *[]string) (ret int32, err error)
+	LoadConfigByInfo(configInfo *ConfigInfo, config *string) (ret int32, err error)
+	CheckConfigByInfo(configInfo *ConfigInfo, result *string) (ret int32, err error)
+	ListAllConfigByInfo(configInfo *GetConfigListInfo, vf *[]string) (ret int32, err error)
+}
+
+type ConfigServantWithContext interface {
+	ListConfig(tarsCtx context.Context, app string, server string, vf *[]string) (ret int32, err error)
+	LoadConfig(tarsCtx context.Context, app string, server string, filename string, config *string) (ret int32, err error)
+	LoadConfigByHost(tarsCtx context.Context, appServerName string, filename string, host string, config *string) (ret int32, err error)
+	CheckConfig(tarsCtx context.Context, appServerName string, filename string, host string, result *string) (ret int32, err error)
+	ListConfigByInfo(tarsCtx context.Context, configInfo *ConfigInfo, vf *[]string) (ret int32, err error)
+	LoadConfigByInfo(tarsCtx context.Context, configInfo *ConfigInfo, config *string) (ret int32, err error)
+	CheckConfigByInfo(tarsCtx context.Context, configInfo *ConfigInfo, result *string) (ret int32, err error)
+	ListAllConfigByInfo(tarsCtx context.Context, configInfo *GetConfigListInfo, vf *[]string) (ret int32, err error)
+}
+
 // Config struct
 type Config struct {
-	servant m.Servant
+	servant model.Servant
+}
+
+// NewConfig creates a new Config servant.
+func NewConfig() *Config {
+	return new(Config)
+}
+
+// SetServant sets servant for the service.
+func (obj *Config) SetServant(servant model.Servant) {
+	obj.servant = servant
+}
+
+// TarsSetTimeout sets the timeout for the servant which is in ms.
+func (obj *Config) TarsSetTimeout(timeout int) {
+	obj.servant.TarsSetTimeout(timeout)
+}
+
+// TarsSetProtocol sets the protocol for the servant.
+func (obj *Config) TarsSetProtocol(p model.Protocol) {
+	obj.servant.TarsSetProtocol(p)
+}
+
+// Endpoints returns all active endpoint.Endpoint
+func (obj *Config) Endpoints() []*endpoint.Endpoint {
+	return obj.servant.Endpoints()
 }
 
 // ListConfig is the proxy function for the method defined in the tars file, with the context
-func (obj *Config) ListConfig(app string, server string, vf *[]string, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(app, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(server, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteHead(codec.LIST, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*vf)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *vf {
-
-		err = buf.WriteString(v, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "ListConfig", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(3, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*vf = make([]string, length)
-		for i0, e0 := int32(0), length; i0 < e0; i0++ {
-
-			err = readBuf.ReadString(&(*vf)[i0], 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *Config) ListConfig(app string, server string, vf *[]string, opts ...map[string]string) (int32, error) {
+	return obj.ListConfigWithContext(context.Background(), app, server, vf, opts...)
 }
 
 // ListConfigWithContext is the proxy function for the method defined in the tars file, with the context
@@ -173,19 +105,15 @@ func (obj *Config) ListConfigWithContext(tarsCtx context.Context, app string, se
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*vf)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *vf {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -202,7 +130,6 @@ func (obj *Config) ListConfigWithContext(tarsCtx context.Context, app string, se
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -213,34 +140,28 @@ func (obj *Config) ListConfigWithContext(tarsCtx context.Context, app string, se
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*vf = make([]string, length)
-		for i1, e1 := int32(0), length; i1 < e1; i1++ {
-
-			err = readBuf.ReadString(&(*vf)[i1], 0, false)
+		for i0, e0 := int32(0), length; i0 < e0; i0++ {
+			err = readBuf.ReadString(&(*vf)[i0], 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -264,6 +185,7 @@ func (obj *Config) ListConfigWithContext(tarsCtx context.Context, app string, se
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -292,19 +214,15 @@ func (obj *Config) ListConfigOneWayWithContext(tarsCtx context.Context, app stri
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*vf)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *vf {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -322,27 +240,6 @@ func (obj *Config) ListConfigOneWayWithContext(tarsCtx context.Context, app stri
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -350,85 +247,8 @@ func (obj *Config) ListConfigOneWayWithContext(tarsCtx context.Context, app stri
 }
 
 // LoadConfig is the proxy function for the method defined in the tars file, with the context
-func (obj *Config) LoadConfig(app string, server string, filename string, config *string, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(app, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(server, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(filename, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(*config, 4)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "loadConfig", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	err = readBuf.ReadString(&(*config), 4, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *Config) LoadConfig(app string, server string, filename string, config *string, opts ...map[string]string) (int32, error) {
+	return obj.LoadConfigWithContext(context.Background(), app, server, filename, config, opts...)
 }
 
 // LoadConfigWithContext is the proxy function for the method defined in the tars file, with the context
@@ -473,7 +293,6 @@ func (obj *Config) LoadConfigWithContext(tarsCtx context.Context, app string, se
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -506,6 +325,7 @@ func (obj *Config) LoadConfigWithContext(tarsCtx context.Context, app string, se
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -555,27 +375,6 @@ func (obj *Config) LoadConfigOneWayWithContext(tarsCtx context.Context, app stri
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -583,85 +382,8 @@ func (obj *Config) LoadConfigOneWayWithContext(tarsCtx context.Context, app stri
 }
 
 // LoadConfigByHost is the proxy function for the method defined in the tars file, with the context
-func (obj *Config) LoadConfigByHost(appServerName string, filename string, host string, config *string, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(appServerName, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(filename, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(host, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(*config, 4)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "loadConfigByHost", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	err = readBuf.ReadString(&(*config), 4, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *Config) LoadConfigByHost(appServerName string, filename string, host string, config *string, opts ...map[string]string) (int32, error) {
+	return obj.LoadConfigByHostWithContext(context.Background(), appServerName, filename, host, config, opts...)
 }
 
 // LoadConfigByHostWithContext is the proxy function for the method defined in the tars file, with the context
@@ -706,7 +428,6 @@ func (obj *Config) LoadConfigByHostWithContext(tarsCtx context.Context, appServe
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -739,6 +460,7 @@ func (obj *Config) LoadConfigByHostWithContext(tarsCtx context.Context, appServe
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -788,27 +510,6 @@ func (obj *Config) LoadConfigByHostOneWayWithContext(tarsCtx context.Context, ap
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -816,85 +517,8 @@ func (obj *Config) LoadConfigByHostOneWayWithContext(tarsCtx context.Context, ap
 }
 
 // CheckConfig is the proxy function for the method defined in the tars file, with the context
-func (obj *Config) CheckConfig(appServerName string, filename string, host string, result *string, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(appServerName, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(filename, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(host, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(*result, 4)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "checkConfig", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	err = readBuf.ReadString(&(*result), 4, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *Config) CheckConfig(appServerName string, filename string, host string, result *string, opts ...map[string]string) (int32, error) {
+	return obj.CheckConfigWithContext(context.Background(), appServerName, filename, host, result, opts...)
 }
 
 // CheckConfigWithContext is the proxy function for the method defined in the tars file, with the context
@@ -939,7 +563,6 @@ func (obj *Config) CheckConfigWithContext(tarsCtx context.Context, appServerName
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -972,6 +595,7 @@ func (obj *Config) CheckConfigWithContext(tarsCtx context.Context, appServerName
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -1021,27 +645,6 @@ func (obj *Config) CheckConfigOneWayWithContext(tarsCtx context.Context, appServ
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -1049,118 +652,8 @@ func (obj *Config) CheckConfigOneWayWithContext(tarsCtx context.Context, appServ
 }
 
 // ListConfigByInfo is the proxy function for the method defined in the tars file, with the context
-func (obj *Config) ListConfigByInfo(configInfo *ConfigInfo, vf *[]string, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = configInfo.WriteBlock(buf, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteHead(codec.LIST, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*vf)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *vf {
-
-		err = buf.WriteString(v, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "ListConfigByInfo", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(2, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*vf = make([]string, length)
-		for i2, e2 := int32(0), length; i2 < e2; i2++ {
-
-			err = readBuf.ReadString(&(*vf)[i2], 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *Config) ListConfigByInfo(configInfo *ConfigInfo, vf *[]string, opts ...map[string]string) (int32, error) {
+	return obj.ListConfigByInfoWithContext(context.Background(), configInfo, vf, opts...)
 }
 
 // ListConfigByInfoWithContext is the proxy function for the method defined in the tars file, with the context
@@ -1180,19 +673,15 @@ func (obj *Config) ListConfigByInfoWithContext(tarsCtx context.Context, configIn
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*vf)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *vf {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -1209,7 +698,6 @@ func (obj *Config) ListConfigByInfoWithContext(tarsCtx context.Context, configIn
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -1220,34 +708,28 @@ func (obj *Config) ListConfigByInfoWithContext(tarsCtx context.Context, configIn
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*vf = make([]string, length)
-		for i3, e3 := int32(0), length; i3 < e3; i3++ {
-
-			err = readBuf.ReadString(&(*vf)[i3], 0, false)
+		for i1, e1 := int32(0), length; i1 < e1; i1++ {
+			err = readBuf.ReadString(&(*vf)[i1], 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -1271,6 +753,7 @@ func (obj *Config) ListConfigByInfoWithContext(tarsCtx context.Context, configIn
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -1294,19 +777,15 @@ func (obj *Config) ListConfigByInfoOneWayWithContext(tarsCtx context.Context, co
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*vf)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *vf {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -1324,27 +803,6 @@ func (obj *Config) ListConfigByInfoOneWayWithContext(tarsCtx context.Context, co
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -1352,75 +810,8 @@ func (obj *Config) ListConfigByInfoOneWayWithContext(tarsCtx context.Context, co
 }
 
 // LoadConfigByInfo is the proxy function for the method defined in the tars file, with the context
-func (obj *Config) LoadConfigByInfo(configInfo *ConfigInfo, config *string, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = configInfo.WriteBlock(buf, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(*config, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "loadConfigByInfo", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	err = readBuf.ReadString(&(*config), 2, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *Config) LoadConfigByInfo(configInfo *ConfigInfo, config *string, opts ...map[string]string) (int32, error) {
+	return obj.LoadConfigByInfoWithContext(context.Background(), configInfo, config, opts...)
 }
 
 // LoadConfigByInfoWithContext is the proxy function for the method defined in the tars file, with the context
@@ -1455,7 +846,6 @@ func (obj *Config) LoadConfigByInfoWithContext(tarsCtx context.Context, configIn
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -1488,6 +878,7 @@ func (obj *Config) LoadConfigByInfoWithContext(tarsCtx context.Context, configIn
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -1527,27 +918,6 @@ func (obj *Config) LoadConfigByInfoOneWayWithContext(tarsCtx context.Context, co
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -1555,75 +925,8 @@ func (obj *Config) LoadConfigByInfoOneWayWithContext(tarsCtx context.Context, co
 }
 
 // CheckConfigByInfo is the proxy function for the method defined in the tars file, with the context
-func (obj *Config) CheckConfigByInfo(configInfo *ConfigInfo, result *string, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = configInfo.WriteBlock(buf, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(*result, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "checkConfigByInfo", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	err = readBuf.ReadString(&(*result), 2, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *Config) CheckConfigByInfo(configInfo *ConfigInfo, result *string, opts ...map[string]string) (int32, error) {
+	return obj.CheckConfigByInfoWithContext(context.Background(), configInfo, result, opts...)
 }
 
 // CheckConfigByInfoWithContext is the proxy function for the method defined in the tars file, with the context
@@ -1658,7 +961,6 @@ func (obj *Config) CheckConfigByInfoWithContext(tarsCtx context.Context, configI
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -1691,6 +993,7 @@ func (obj *Config) CheckConfigByInfoWithContext(tarsCtx context.Context, configI
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -1730,27 +1033,6 @@ func (obj *Config) CheckConfigByInfoOneWayWithContext(tarsCtx context.Context, c
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -1758,118 +1040,8 @@ func (obj *Config) CheckConfigByInfoOneWayWithContext(tarsCtx context.Context, c
 }
 
 // ListAllConfigByInfo is the proxy function for the method defined in the tars file, with the context
-func (obj *Config) ListAllConfigByInfo(configInfo *GetConfigListInfo, vf *[]string, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = configInfo.WriteBlock(buf, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteHead(codec.LIST, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*vf)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *vf {
-
-		err = buf.WriteString(v, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "ListAllConfigByInfo", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(2, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*vf = make([]string, length)
-		for i4, e4 := int32(0), length; i4 < e4; i4++ {
-
-			err = readBuf.ReadString(&(*vf)[i4], 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *Config) ListAllConfigByInfo(configInfo *GetConfigListInfo, vf *[]string, opts ...map[string]string) (int32, error) {
+	return obj.ListAllConfigByInfoWithContext(context.Background(), configInfo, vf, opts...)
 }
 
 // ListAllConfigByInfoWithContext is the proxy function for the method defined in the tars file, with the context
@@ -1889,19 +1061,15 @@ func (obj *Config) ListAllConfigByInfoWithContext(tarsCtx context.Context, confi
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*vf)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *vf {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -1918,7 +1086,6 @@ func (obj *Config) ListAllConfigByInfoWithContext(tarsCtx context.Context, confi
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -1929,34 +1096,28 @@ func (obj *Config) ListAllConfigByInfoWithContext(tarsCtx context.Context, confi
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*vf = make([]string, length)
-		for i5, e5 := int32(0), length; i5 < e5; i5++ {
-
-			err = readBuf.ReadString(&(*vf)[i5], 0, false)
+		for i2, e2 := int32(0), length; i2 < e2; i2++ {
+			err = readBuf.ReadString(&(*vf)[i2], 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -1980,6 +1141,7 @@ func (obj *Config) ListAllConfigByInfoWithContext(tarsCtx context.Context, confi
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -2003,19 +1165,15 @@ func (obj *Config) ListAllConfigByInfoOneWayWithContext(tarsCtx context.Context,
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*vf)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *vf {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -2033,67 +1191,10 @@ func (obj *Config) ListAllConfigByInfoOneWayWithContext(tarsCtx context.Context,
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
 	return ret, nil
-}
-
-// SetServant sets servant for the service.
-func (obj *Config) SetServant(servant m.Servant) {
-	obj.servant = servant
-}
-
-// TarsSetTimeout sets the timeout for the servant which is in ms.
-func (obj *Config) TarsSetTimeout(timeout int) {
-	obj.servant.TarsSetTimeout(timeout)
-}
-
-// TarsSetProtocol sets the protocol for the servant.
-func (obj *Config) TarsSetProtocol(p m.Protocol) {
-	obj.servant.TarsSetProtocol(p)
-}
-
-type ConfigServant interface {
-	ListConfig(app string, server string, vf *[]string) (ret int32, err error)
-	LoadConfig(app string, server string, filename string, config *string) (ret int32, err error)
-	LoadConfigByHost(appServerName string, filename string, host string, config *string) (ret int32, err error)
-	CheckConfig(appServerName string, filename string, host string, result *string) (ret int32, err error)
-	ListConfigByInfo(configInfo *ConfigInfo, vf *[]string) (ret int32, err error)
-	LoadConfigByInfo(configInfo *ConfigInfo, config *string) (ret int32, err error)
-	CheckConfigByInfo(configInfo *ConfigInfo, result *string) (ret int32, err error)
-	ListAllConfigByInfo(configInfo *GetConfigListInfo, vf *[]string) (ret int32, err error)
-}
-type ConfigServantWithContext interface {
-	ListConfig(tarsCtx context.Context, app string, server string, vf *[]string) (ret int32, err error)
-	LoadConfig(tarsCtx context.Context, app string, server string, filename string, config *string) (ret int32, err error)
-	LoadConfigByHost(tarsCtx context.Context, appServerName string, filename string, host string, config *string) (ret int32, err error)
-	CheckConfig(tarsCtx context.Context, appServerName string, filename string, host string, result *string) (ret int32, err error)
-	ListConfigByInfo(tarsCtx context.Context, configInfo *ConfigInfo, vf *[]string) (ret int32, err error)
-	LoadConfigByInfo(tarsCtx context.Context, configInfo *ConfigInfo, config *string) (ret int32, err error)
-	CheckConfigByInfo(tarsCtx context.Context, configInfo *ConfigInfo, result *string) (ret int32, err error)
-	ListAllConfigByInfo(tarsCtx context.Context, configInfo *GetConfigListInfo, vf *[]string) (ret int32, err error)
 }
 
 // Dispatch is used to call the server side implement for the method defined in the tars file. withContext shows using context or not.
@@ -2111,19 +1212,15 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		var server string
 		var vf []string
 		vf = make([]string, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&app, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&server, 2, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2143,7 +1240,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2164,7 +1260,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2178,7 +1273,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(ConfigServantWithContext)
 			funRet, err = imp.ListConfig(tarsCtx, app, server, &vf)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2190,26 +1284,20 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteHead(codec.LIST, 3)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(vf)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range vf {
-
 				err = buf.WriteString(v, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2226,19 +1314,15 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(vf)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range vf {
-
 				err = buf.WriteString(v, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("vf", buf.ToBytes())
 
@@ -2268,24 +1352,19 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		var server string
 		var filename string
 		var config string
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&app, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&server, 2, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&filename, 3, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2312,7 +1391,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2339,7 +1417,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2353,7 +1430,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(ConfigServantWithContext)
 			funRet, err = imp.LoadConfig(tarsCtx, app, server, filename, &config)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2365,12 +1441,10 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteString(config, 4)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2387,7 +1461,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			rspTup.PutBuffer("config", buf.ToBytes())
 
 			buf.Reset()
@@ -2416,24 +1489,19 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		var filename string
 		var host string
 		var config string
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&appServerName, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&filename, 2, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&host, 3, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2460,7 +1528,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2487,7 +1554,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2501,7 +1567,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(ConfigServantWithContext)
 			funRet, err = imp.LoadConfigByHost(tarsCtx, appServerName, filename, host, &config)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2513,12 +1578,10 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteString(config, 4)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2535,7 +1598,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			rspTup.PutBuffer("config", buf.ToBytes())
 
 			buf.Reset()
@@ -2564,24 +1626,19 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		var filename string
 		var host string
 		var result string
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&appServerName, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&filename, 2, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&host, 3, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2608,7 +1665,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2635,7 +1691,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2649,7 +1704,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(ConfigServantWithContext)
 			funRet, err = imp.CheckConfig(tarsCtx, appServerName, filename, host, &result)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2661,12 +1715,10 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteString(result, 4)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2683,7 +1735,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			rspTup.PutBuffer("result", buf.ToBytes())
 
 			buf.Reset()
@@ -2711,14 +1762,11 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		var configInfo ConfigInfo
 		var vf []string
 		vf = make([]string, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = configInfo.ReadBlock(readBuf, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2731,7 +1779,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2747,7 +1794,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2761,7 +1807,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(ConfigServantWithContext)
 			funRet, err = imp.ListConfigByInfo(tarsCtx, &configInfo, &vf)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2773,26 +1818,20 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteHead(codec.LIST, 2)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(vf)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range vf {
-
 				err = buf.WriteString(v, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2809,19 +1848,15 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(vf)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range vf {
-
 				err = buf.WriteString(v, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("vf", buf.ToBytes())
 
@@ -2849,14 +1884,11 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 	case "loadConfigByInfo":
 		var configInfo ConfigInfo
 		var config string
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = configInfo.ReadBlock(readBuf, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2869,7 +1901,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2885,7 +1916,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2899,7 +1929,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(ConfigServantWithContext)
 			funRet, err = imp.LoadConfigByInfo(tarsCtx, &configInfo, &config)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2911,12 +1940,10 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteString(config, 2)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2933,7 +1960,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			rspTup.PutBuffer("config", buf.ToBytes())
 
 			buf.Reset()
@@ -2960,14 +1986,11 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 	case "checkConfigByInfo":
 		var configInfo ConfigInfo
 		var result string
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = configInfo.ReadBlock(readBuf, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2980,7 +2003,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2996,7 +2018,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -3010,7 +2031,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(ConfigServantWithContext)
 			funRet, err = imp.CheckConfigByInfo(tarsCtx, &configInfo, &result)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -3022,12 +2042,10 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteString(result, 2)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -3044,7 +2062,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			rspTup.PutBuffer("result", buf.ToBytes())
 
 			buf.Reset()
@@ -3072,14 +2089,11 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		var configInfo GetConfigListInfo
 		var vf []string
 		vf = make([]string, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = configInfo.ReadBlock(readBuf, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -3092,7 +2106,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -3108,7 +2121,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -3122,7 +2134,6 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(ConfigServantWithContext)
 			funRet, err = imp.ListAllConfigByInfo(tarsCtx, &configInfo, &vf)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -3134,26 +2145,20 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteHead(codec.LIST, 2)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(vf)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range vf {
-
 				err = buf.WriteString(v, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -3170,19 +2175,15 @@ func (obj *Config) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(vf)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range vf {
-
 				err = buf.WriteString(v, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("vf", buf.ToBytes())
 
