@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from ConfigF.tars
 // Package configf comment
-// This file was generated by tars2go 1.1.10
-// Generated from ConfigF.tars
 package configf
 
 import (
@@ -15,12 +15,12 @@ var _ = codec.FromInt8
 
 // ConfigInfo struct implement
 type ConfigInfo struct {
-	Appname     string `json:"appname"`
-	Servername  string `json:"servername"`
-	Filename    string `json:"filename"`
-	BAppOnly    bool   `json:"bAppOnly"`
-	Host        string `json:"host"`
-	Setdivision string `json:"setdivision"`
+	Appname     string `json:"appname" tars:"appname,tag:0,require:true"`
+	Servername  string `json:"servername" tars:"servername,tag:1,require:true"`
+	Filename    string `json:"filename" tars:"filename,tag:2,require:true"`
+	BAppOnly    bool   `json:"bAppOnly" tars:"bAppOnly,tag:3,require:true"`
+	Host        string `json:"host" tars:"host,tag:4,require:false"`
+	Setdivision string `json:"setdivision" tars:"setdivision,tag:5,require:false"`
 }
 
 func (st *ConfigInfo) ResetDefault() {
@@ -108,7 +108,6 @@ func (st *ConfigInfo) ReadBlock(readBuf *codec.Reader, tag byte, require bool) e
 
 // WriteTo encode struct to buffer
 func (st *ConfigInfo) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.Appname, 0)
 	if err != nil {
 		return err
@@ -129,14 +128,18 @@ func (st *ConfigInfo) WriteTo(buf *codec.Buffer) (err error) {
 		return err
 	}
 
-	err = buf.WriteString(st.Host, 4)
-	if err != nil {
-		return err
+	if st.Host != "" {
+		err = buf.WriteString(st.Host, 4)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.Setdivision, 5)
-	if err != nil {
-		return err
+	if st.Setdivision != "" {
+		err = buf.WriteString(st.Setdivision, 5)
+		if err != nil {
+			return err
+		}
 	}
 
 	return err
@@ -164,12 +167,12 @@ func (st *ConfigInfo) WriteBlock(buf *codec.Buffer, tag byte) error {
 
 // GetConfigListInfo struct implement
 type GetConfigListInfo struct {
-	Appname       string `json:"appname"`
-	Servername    string `json:"servername"`
-	BAppOnly      bool   `json:"bAppOnly"`
-	Host          string `json:"host"`
-	Setdivision   string `json:"setdivision"`
-	Containername string `json:"containername"`
+	Appname       string `json:"appname" tars:"appname,tag:0,require:true"`
+	Servername    string `json:"servername" tars:"servername,tag:1,require:false"`
+	BAppOnly      bool   `json:"bAppOnly" tars:"bAppOnly,tag:2,require:false"`
+	Host          string `json:"host" tars:"host,tag:3,require:false"`
+	Setdivision   string `json:"setdivision" tars:"setdivision,tag:4,require:false"`
+	Containername string `json:"containername" tars:"containername,tag:5,require:false"`
 }
 
 func (st *GetConfigListInfo) ResetDefault() {
@@ -260,35 +263,44 @@ func (st *GetConfigListInfo) ReadBlock(readBuf *codec.Reader, tag byte, require 
 
 // WriteTo encode struct to buffer
 func (st *GetConfigListInfo) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.Appname, 0)
 	if err != nil {
 		return err
 	}
 
-	err = buf.WriteString(st.Servername, 1)
-	if err != nil {
-		return err
+	if st.Servername != "" {
+		err = buf.WriteString(st.Servername, 1)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteBool(st.BAppOnly, 2)
-	if err != nil {
-		return err
+	if st.BAppOnly != false {
+		err = buf.WriteBool(st.BAppOnly, 2)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.Host, 3)
-	if err != nil {
-		return err
+	if st.Host != "" {
+		err = buf.WriteString(st.Host, 3)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.Setdivision, 4)
-	if err != nil {
-		return err
+	if st.Setdivision != "" {
+		err = buf.WriteString(st.Setdivision, 4)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.Containername, 5)
-	if err != nil {
-		return err
+	if st.Containername != "" {
+		err = buf.WriteString(st.Containername, 5)
+		if err != nil {
+			return err
+		}
 	}
 
 	return err
