@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from ./EndpointF.tars
 // Package endpointf comment
-// This file was generated by tars2go 1.1.10
-// Generated from EndpointF.tars
 package endpointf
 
 import (
@@ -15,19 +15,19 @@ var _ = codec.FromInt8
 
 // EndpointF struct implement
 type EndpointF struct {
-	Host        string `json:"host"`
-	Port        int32  `json:"port"`
-	Timeout     int32  `json:"timeout"`
-	Istcp       int32  `json:"istcp"`
-	Grid        int32  `json:"grid"`
-	Groupworkid int32  `json:"groupworkid"`
-	Grouprealid int32  `json:"grouprealid"`
-	SetId       string `json:"setId"`
-	Qos         int32  `json:"qos"`
-	BakFlag     int32  `json:"bakFlag"`
-	Weight      int32  `json:"weight"`
-	WeightType  int32  `json:"weightType"`
-	AuthType    int32  `json:"authType"`
+	Host        string `json:"host" tars:"host,tag:0,require:true"`
+	Port        int32  `json:"port" tars:"port,tag:1,require:true"`
+	Timeout     int32  `json:"timeout" tars:"timeout,tag:2,require:true"`
+	Istcp       int32  `json:"istcp" tars:"istcp,tag:3,require:true"`
+	Grid        int32  `json:"grid" tars:"grid,tag:4,require:true"`
+	Groupworkid int32  `json:"groupworkid" tars:"groupworkid,tag:5,require:false"`
+	Grouprealid int32  `json:"grouprealid" tars:"grouprealid,tag:6,require:false"`
+	SetId       string `json:"setId" tars:"setId,tag:7,require:false"`
+	Qos         int32  `json:"qos" tars:"qos,tag:8,require:false"`
+	BakFlag     int32  `json:"bakFlag" tars:"bakFlag,tag:9,require:false"`
+	Weight      int32  `json:"weight" tars:"weight,tag:11,require:false"`
+	WeightType  int32  `json:"weightType" tars:"weightType,tag:12,require:false"`
+	AuthType    int32  `json:"authType" tars:"authType,tag:13,require:false"`
 }
 
 func (st *EndpointF) ResetDefault() {
@@ -149,7 +149,6 @@ func (st *EndpointF) ReadBlock(readBuf *codec.Reader, tag byte, require bool) er
 
 // WriteTo encode struct to buffer
 func (st *EndpointF) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.Host, 0)
 	if err != nil {
 		return err
@@ -175,44 +174,60 @@ func (st *EndpointF) WriteTo(buf *codec.Buffer) (err error) {
 		return err
 	}
 
-	err = buf.WriteInt32(st.Groupworkid, 5)
-	if err != nil {
-		return err
+	if st.Groupworkid != 0 {
+		err = buf.WriteInt32(st.Groupworkid, 5)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteInt32(st.Grouprealid, 6)
-	if err != nil {
-		return err
+	if st.Grouprealid != 0 {
+		err = buf.WriteInt32(st.Grouprealid, 6)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SetId, 7)
-	if err != nil {
-		return err
+	if st.SetId != "" {
+		err = buf.WriteString(st.SetId, 7)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteInt32(st.Qos, 8)
-	if err != nil {
-		return err
+	if st.Qos != 0 {
+		err = buf.WriteInt32(st.Qos, 8)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteInt32(st.BakFlag, 9)
-	if err != nil {
-		return err
+	if st.BakFlag != 0 {
+		err = buf.WriteInt32(st.BakFlag, 9)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteInt32(st.Weight, 11)
-	if err != nil {
-		return err
+	if st.Weight != 0 {
+		err = buf.WriteInt32(st.Weight, 11)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteInt32(st.WeightType, 12)
-	if err != nil {
-		return err
+	if st.WeightType != 0 {
+		err = buf.WriteInt32(st.WeightType, 12)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteInt32(st.AuthType, 13)
-	if err != nil {
-		return err
+	if st.AuthType != 0 {
+		err = buf.WriteInt32(st.AuthType, 13)
+		if err != nil {
+			return err
+		}
 	}
 
 	return err
