@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from LogF.tars
 // Package logf comment
-// This file was generated by tars2go 1.1.10
-// Generated from LogF.tars
 package logf
 
 import (
@@ -8,117 +8,67 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	m "github.com/TarsCloud/TarsGo/tars/model"
+
+	"github.com/TarsCloud/TarsGo/tars/model"
 	"github.com/TarsCloud/TarsGo/tars/protocol/codec"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/requestf"
 	"github.com/TarsCloud/TarsGo/tars/protocol/tup"
 	"github.com/TarsCloud/TarsGo/tars/util/current"
+	"github.com/TarsCloud/TarsGo/tars/util/endpoint"
 	"github.com/TarsCloud/TarsGo/tars/util/tools"
-	"unsafe"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var (
 	_ = fmt.Errorf
 	_ = codec.FromInt8
-	_ = unsafe.Pointer(nil)
 	_ = bytes.ErrTooLarge
 )
 
+type LogServant interface {
+	Logger(app string, server string, file string, format string, buffer []string) (err error)
+	LoggerbyInfo(info *LogInfo, buffer []string) (err error)
+}
+
+type LogServantWithContext interface {
+	Logger(tarsCtx context.Context, app string, server string, file string, format string, buffer []string) (err error)
+	LoggerbyInfo(tarsCtx context.Context, info *LogInfo, buffer []string) (err error)
+}
+
 // Log struct
 type Log struct {
-	servant m.Servant
+	servant model.Servant
+}
+
+// NewLog creates a new Log servant.
+func NewLog() *Log {
+	return new(Log)
+}
+
+// SetServant sets servant for the service.
+func (obj *Log) SetServant(servant model.Servant) {
+	obj.servant = servant
+}
+
+// TarsSetTimeout sets the timeout for the servant which is in ms.
+func (obj *Log) TarsSetTimeout(timeout int) {
+	obj.servant.TarsSetTimeout(timeout)
+}
+
+// TarsSetProtocol sets the protocol for the servant.
+func (obj *Log) TarsSetProtocol(p model.Protocol) {
+	obj.servant.TarsSetProtocol(p)
+}
+
+// Endpoints returns all active endpoint.Endpoint
+func (obj *Log) Endpoints() []*endpoint.Endpoint {
+	return obj.servant.Endpoints()
 }
 
 // Logger is the proxy function for the method defined in the tars file, with the context
-func (obj *Log) Logger(app string, server string, file string, format string, buffer []string, opts ...map[string]string) (err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(app, 1)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteString(server, 2)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteString(file, 3)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteString(format, 4)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteHead(codec.LIST, 5)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteInt32(int32(len(buffer)), 0)
-	if err != nil {
-		return err
-	}
-
-	for _, v := range buffer {
-
-		err = buf.WriteString(v, 0)
-		if err != nil {
-			return err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "logger", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return nil
+func (obj *Log) Logger(app string, server string, file string, format string, buffer []string, opts ...map[string]string) error {
+	return obj.LoggerWithContext(context.Background(), app, server, file, format, buffer, opts...)
 }
 
 // LoggerWithContext is the proxy function for the method defined in the tars file, with the context
@@ -153,19 +103,15 @@ func (obj *Log) LoggerWithContext(tarsCtx context.Context, app string, server st
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(buffer)), 0)
 	if err != nil {
 		return err
 	}
-
 	for _, v := range buffer {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -204,6 +150,7 @@ func (obj *Log) LoggerWithContext(tarsCtx context.Context, app string, server st
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -242,19 +189,15 @@ func (obj *Log) LoggerOneWayWithContext(tarsCtx context.Context, app string, ser
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(buffer)), 0)
 	if err != nil {
 		return err
 	}
-
 	for _, v := range buffer {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -272,27 +215,6 @@ func (obj *Log) LoggerOneWayWithContext(tarsCtx context.Context, app string, ser
 		return err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -300,78 +222,8 @@ func (obj *Log) LoggerOneWayWithContext(tarsCtx context.Context, app string, ser
 }
 
 // LoggerbyInfo is the proxy function for the method defined in the tars file, with the context
-func (obj *Log) LoggerbyInfo(info *LogInfo, buffer []string, opts ...map[string]string) (err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = info.WriteBlock(buf, 1)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteHead(codec.LIST, 2)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteInt32(int32(len(buffer)), 0)
-	if err != nil {
-		return err
-	}
-
-	for _, v := range buffer {
-
-		err = buf.WriteString(v, 0)
-		if err != nil {
-			return err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "loggerbyInfo", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return nil
+func (obj *Log) LoggerbyInfo(info *LogInfo, buffer []string, opts ...map[string]string) error {
+	return obj.LoggerbyInfoWithContext(context.Background(), info, buffer, opts...)
 }
 
 // LoggerbyInfoWithContext is the proxy function for the method defined in the tars file, with the context
@@ -391,19 +243,15 @@ func (obj *Log) LoggerbyInfoWithContext(tarsCtx context.Context, info *LogInfo, 
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(buffer)), 0)
 	if err != nil {
 		return err
 	}
-
 	for _, v := range buffer {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -442,6 +290,7 @@ func (obj *Log) LoggerbyInfoWithContext(tarsCtx context.Context, info *LogInfo, 
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -465,19 +314,15 @@ func (obj *Log) LoggerbyInfoOneWayWithContext(tarsCtx context.Context, info *Log
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(buffer)), 0)
 	if err != nil {
 		return err
 	}
-
 	for _, v := range buffer {
-
 		err = buf.WriteString(v, 0)
 		if err != nil {
 			return err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -495,55 +340,10 @@ func (obj *Log) LoggerbyInfoOneWayWithContext(tarsCtx context.Context, info *Log
 		return err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
 	return nil
-}
-
-// SetServant sets servant for the service.
-func (obj *Log) SetServant(servant m.Servant) {
-	obj.servant = servant
-}
-
-// TarsSetTimeout sets the timeout for the servant which is in ms.
-func (obj *Log) TarsSetTimeout(timeout int) {
-	obj.servant.TarsSetTimeout(timeout)
-}
-
-// TarsSetProtocol sets the protocol for the servant.
-func (obj *Log) TarsSetProtocol(p m.Protocol) {
-	obj.servant.TarsSetProtocol(p)
-}
-
-type LogServant interface {
-	Logger(app string, server string, file string, format string, buffer []string) (err error)
-	LoggerbyInfo(info *LogInfo, buffer []string) (err error)
-}
-type LogServantWithContext interface {
-	Logger(tarsCtx context.Context, app string, server string, file string, format string, buffer []string) (err error)
-	LoggerbyInfo(tarsCtx context.Context, info *LogInfo, buffer []string) (err error)
 }
 
 // Dispatch is used to call the server side implement for the method defined in the tars file. withContext shows using context or not.
@@ -563,61 +363,49 @@ func (obj *Log) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *requ
 		var format string
 		var buffer []string
 		buffer = make([]string, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&app, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&server, 2, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&file, 3, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&format, 4, true)
 			if err != nil {
 				return err
 			}
-
 			_, ty, err = readBuf.SkipToNoCheck(5, true)
 			if err != nil {
 				return err
 			}
-
 			if ty == codec.LIST {
 				err = readBuf.ReadInt32(&length, 0, true)
 				if err != nil {
 					return err
 				}
-
 				buffer = make([]string, length)
 				for i0, e0 := int32(0), length; i0 < e0; i0++ {
-
-					err = readBuf.ReadString(&buffer[i0], 0, false)
+					err = readBuf.ReadString(&buffer[i0], 0, true)
 					if err != nil {
 						return err
 					}
-
 				}
 			} else if ty == codec.SimpleList {
 				err = fmt.Errorf("not support SimpleList type")
 				if err != nil {
 					return err
 				}
-
 			} else {
 				err = fmt.Errorf("require vector, but not")
 				if err != nil {
 					return err
 				}
-
 			}
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
@@ -659,34 +447,28 @@ func (obj *Log) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *requ
 			if err != nil {
 				return err
 			}
-
 			if ty == codec.LIST {
 				err = readBuf.ReadInt32(&length, 0, true)
 				if err != nil {
 					return err
 				}
-
 				buffer = make([]string, length)
 				for i1, e1 := int32(0), length; i1 < e1; i1++ {
-
-					err = readBuf.ReadString(&buffer[i1], 0, false)
+					err = readBuf.ReadString(&buffer[i1], 0, true)
 					if err != nil {
 						return err
 					}
-
 				}
 			} else if ty == codec.SimpleList {
 				err = fmt.Errorf("not support SimpleList type")
 				if err != nil {
 					return err
 				}
-
 			} else {
 				err = fmt.Errorf("require vector, but not")
 				if err != nil {
 					return err
 				}
-
 			}
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
@@ -726,7 +508,6 @@ func (obj *Log) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *requ
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -739,7 +520,6 @@ func (obj *Log) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *requ
 			imp := val.(LogServantWithContext)
 			err = imp.Logger(tarsCtx, app, server, file, format, buffer)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -773,46 +553,37 @@ func (obj *Log) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *requ
 		var info LogInfo
 		var buffer []string
 		buffer = make([]string, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = info.ReadBlock(readBuf, 1, true)
 			if err != nil {
 				return err
 			}
-
 			_, ty, err = readBuf.SkipToNoCheck(2, true)
 			if err != nil {
 				return err
 			}
-
 			if ty == codec.LIST {
 				err = readBuf.ReadInt32(&length, 0, true)
 				if err != nil {
 					return err
 				}
-
 				buffer = make([]string, length)
 				for i2, e2 := int32(0), length; i2 < e2; i2++ {
-
-					err = readBuf.ReadString(&buffer[i2], 0, false)
+					err = readBuf.ReadString(&buffer[i2], 0, true)
 					if err != nil {
 						return err
 					}
-
 				}
 			} else if ty == codec.SimpleList {
 				err = fmt.Errorf("not support SimpleList type")
 				if err != nil {
 					return err
 				}
-
 			} else {
 				err = fmt.Errorf("require vector, but not")
 				if err != nil {
 					return err
 				}
-
 			}
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
@@ -833,34 +604,28 @@ func (obj *Log) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *requ
 			if err != nil {
 				return err
 			}
-
 			if ty == codec.LIST {
 				err = readBuf.ReadInt32(&length, 0, true)
 				if err != nil {
 					return err
 				}
-
 				buffer = make([]string, length)
 				for i3, e3 := int32(0), length; i3 < e3; i3++ {
-
-					err = readBuf.ReadString(&buffer[i3], 0, false)
+					err = readBuf.ReadString(&buffer[i3], 0, true)
 					if err != nil {
 						return err
 					}
-
 				}
 			} else if ty == codec.SimpleList {
 				err = fmt.Errorf("not support SimpleList type")
 				if err != nil {
 					return err
 				}
-
 			} else {
 				err = fmt.Errorf("require vector, but not")
 				if err != nil {
 					return err
 				}
-
 			}
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
@@ -883,7 +648,6 @@ func (obj *Log) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *requ
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -896,7 +660,6 @@ func (obj *Log) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *requ
 			imp := val.(LogServantWithContext)
 			err = imp.LoggerbyInfo(tarsCtx, &info, buffer)
 		}
-
 		if err != nil {
 			return err
 		}
