@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from LogF.tars
 // Package logf comment
-// This file was generated by tars2go 1.1.10
-// Generated from LogF.tars
 package logf
 
 import (
@@ -15,17 +15,17 @@ var _ = codec.FromInt8
 
 // LogInfo struct implement
 type LogInfo struct {
-	Appname           string `json:"appname"`
-	Servername        string `json:"servername"`
-	SFilename         string `json:"sFilename"`
-	SFormat           string `json:"sFormat"`
-	Setdivision       string `json:"setdivision"`
-	BHasSufix         bool   `json:"bHasSufix"`
-	BHasAppNamePrefix bool   `json:"bHasAppNamePrefix"`
-	BHasSquareBracket bool   `json:"bHasSquareBracket"`
-	SConcatStr        string `json:"sConcatStr"`
-	SSepar            string `json:"sSepar"`
-	SLogType          string `json:"sLogType"`
+	Appname           string `json:"appname" tars:"appname,tag:0,require:true"`
+	Servername        string `json:"servername" tars:"servername,tag:1,require:true"`
+	SFilename         string `json:"sFilename" tars:"sFilename,tag:2,require:true"`
+	SFormat           string `json:"sFormat" tars:"sFormat,tag:3,require:true"`
+	Setdivision       string `json:"setdivision" tars:"setdivision,tag:4,require:false"`
+	BHasSufix         bool   `json:"bHasSufix" tars:"bHasSufix,tag:5,require:false"`
+	BHasAppNamePrefix bool   `json:"bHasAppNamePrefix" tars:"bHasAppNamePrefix,tag:6,require:false"`
+	BHasSquareBracket bool   `json:"bHasSquareBracket" tars:"bHasSquareBracket,tag:7,require:false"`
+	SConcatStr        string `json:"sConcatStr" tars:"sConcatStr,tag:8,require:false"`
+	SSepar            string `json:"sSepar" tars:"sSepar,tag:9,require:false"`
+	SLogType          string `json:"sLogType" tars:"sLogType,tag:10,require:false"`
 }
 
 func (st *LogInfo) ResetDefault() {
@@ -143,7 +143,6 @@ func (st *LogInfo) ReadBlock(readBuf *codec.Reader, tag byte, require bool) erro
 
 // WriteTo encode struct to buffer
 func (st *LogInfo) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.Appname, 0)
 	if err != nil {
 		return err
@@ -164,39 +163,53 @@ func (st *LogInfo) WriteTo(buf *codec.Buffer) (err error) {
 		return err
 	}
 
-	err = buf.WriteString(st.Setdivision, 4)
-	if err != nil {
-		return err
+	if st.Setdivision != "" {
+		err = buf.WriteString(st.Setdivision, 4)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteBool(st.BHasSufix, 5)
-	if err != nil {
-		return err
+	if st.BHasSufix != true {
+		err = buf.WriteBool(st.BHasSufix, 5)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteBool(st.BHasAppNamePrefix, 6)
-	if err != nil {
-		return err
+	if st.BHasAppNamePrefix != true {
+		err = buf.WriteBool(st.BHasAppNamePrefix, 6)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteBool(st.BHasSquareBracket, 7)
-	if err != nil {
-		return err
+	if st.BHasSquareBracket != false {
+		err = buf.WriteBool(st.BHasSquareBracket, 7)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SConcatStr, 8)
-	if err != nil {
-		return err
+	if st.SConcatStr != "_" {
+		err = buf.WriteString(st.SConcatStr, 8)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SSepar, 9)
-	if err != nil {
-		return err
+	if st.SSepar != "|" {
+		err = buf.WriteString(st.SSepar, 9)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SLogType, 10)
-	if err != nil {
-		return err
+	if st.SLogType != "" {
+		err = buf.WriteString(st.SLogType, 10)
+		if err != nil {
+			return err
+		}
 	}
 
 	return err
