@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from NodeF.tars
 // Package nodef comment
-// This file was generated by tars2go 1.1.10
-// Generated from NodeF.tars
 package nodef
 
 import (
@@ -15,10 +15,10 @@ var _ = codec.FromInt8
 
 // ServerInfo struct implement
 type ServerInfo struct {
-	Application string `json:"application"`
-	ServerName  string `json:"serverName"`
-	Pid         int32  `json:"pid"`
-	Adapter     string `json:"adapter"`
+	Application string `json:"application" tars:"application,tag:0,require:true"`
+	ServerName  string `json:"serverName" tars:"serverName,tag:1,require:true"`
+	Pid         int32  `json:"pid" tars:"pid,tag:2,require:true"`
+	Adapter     string `json:"adapter" tars:"adapter,tag:3,require:false"`
 }
 
 func (st *ServerInfo) ResetDefault() {
@@ -95,7 +95,6 @@ func (st *ServerInfo) ReadBlock(readBuf *codec.Reader, tag byte, require bool) e
 
 // WriteTo encode struct to buffer
 func (st *ServerInfo) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.Application, 0)
 	if err != nil {
 		return err
@@ -111,9 +110,11 @@ func (st *ServerInfo) WriteTo(buf *codec.Buffer) (err error) {
 		return err
 	}
 
-	err = buf.WriteString(st.Adapter, 3)
-	if err != nil {
-		return err
+	if st.Adapter != "" {
+		err = buf.WriteString(st.Adapter, 3)
+		if err != nil {
+			return err
+		}
 	}
 
 	return err
