@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from NodeF.tars
 // Package nodef comment
-// This file was generated by tars2go 1.1.10
-// Generated from NodeF.tars
 package nodef
 
 import (
@@ -8,89 +8,71 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	m "github.com/TarsCloud/TarsGo/tars/model"
+
+	"github.com/TarsCloud/TarsGo/tars/model"
 	"github.com/TarsCloud/TarsGo/tars/protocol/codec"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/requestf"
 	"github.com/TarsCloud/TarsGo/tars/protocol/tup"
 	"github.com/TarsCloud/TarsGo/tars/util/current"
+	"github.com/TarsCloud/TarsGo/tars/util/endpoint"
 	"github.com/TarsCloud/TarsGo/tars/util/tools"
-	"unsafe"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var (
 	_ = fmt.Errorf
 	_ = codec.FromInt8
-	_ = unsafe.Pointer(nil)
 	_ = bytes.ErrTooLarge
 )
 
+type ServerFServant interface {
+	KeepAlive(serverInfo *ServerInfo) (ret int32, err error)
+	KeepActiving(serverInfo *ServerInfo) (ret int32, err error)
+	ReportVersion(app string, serverName string, version string) (ret int32, err error)
+	GetLatestKeepAliveTime() (ret uint32, err error)
+}
+
+type ServerFServantWithContext interface {
+	KeepAlive(tarsCtx context.Context, serverInfo *ServerInfo) (ret int32, err error)
+	KeepActiving(tarsCtx context.Context, serverInfo *ServerInfo) (ret int32, err error)
+	ReportVersion(tarsCtx context.Context, app string, serverName string, version string) (ret int32, err error)
+	GetLatestKeepAliveTime(tarsCtx context.Context) (ret uint32, err error)
+}
+
 // ServerF struct
 type ServerF struct {
-	servant m.Servant
+	servant model.Servant
+}
+
+// NewServerF creates a new ServerF servant.
+func NewServerF() *ServerF {
+	return new(ServerF)
+}
+
+// SetServant sets servant for the service.
+func (obj *ServerF) SetServant(servant model.Servant) {
+	obj.servant = servant
+}
+
+// TarsSetTimeout sets the timeout for the servant which is in ms.
+func (obj *ServerF) TarsSetTimeout(timeout int) {
+	obj.servant.TarsSetTimeout(timeout)
+}
+
+// TarsSetProtocol sets the protocol for the servant.
+func (obj *ServerF) TarsSetProtocol(p model.Protocol) {
+	obj.servant.TarsSetProtocol(p)
+}
+
+// Endpoints returns all active endpoint.Endpoint
+func (obj *ServerF) Endpoints() []*endpoint.Endpoint {
+	return obj.servant.Endpoints()
 }
 
 // KeepAlive is the proxy function for the method defined in the tars file, with the context
-func (obj *ServerF) KeepAlive(serverInfo *ServerInfo, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = serverInfo.WriteBlock(buf, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "keepAlive", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *ServerF) KeepAlive(serverInfo *ServerInfo, opts ...map[string]string) (int32, error) {
+	return obj.KeepAliveWithContext(context.Background(), serverInfo, opts...)
 }
 
 // KeepAliveWithContext is the proxy function for the method defined in the tars file, with the context
@@ -120,7 +102,6 @@ func (obj *ServerF) KeepAliveWithContext(tarsCtx context.Context, serverInfo *Se
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -148,6 +129,7 @@ func (obj *ServerF) KeepAliveWithContext(tarsCtx context.Context, serverInfo *Se
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -182,52 +164,26 @@ func (obj *ServerF) KeepAliveOneWayWithContext(tarsCtx context.Context, serverIn
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
 	return ret, nil
 }
 
-// ReportVersion is the proxy function for the method defined in the tars file, with the context
-func (obj *ServerF) ReportVersion(app string, serverName string, version string, opts ...map[string]string) (ret int32, err error) {
+// KeepActiving is the proxy function for the method defined in the tars file, with the context
+func (obj *ServerF) KeepActiving(serverInfo *ServerInfo, opts ...map[string]string) (int32, error) {
+	return obj.KeepActivingWithContext(context.Background(), serverInfo, opts...)
+}
+
+// KeepActivingWithContext is the proxy function for the method defined in the tars file, with the context
+func (obj *ServerF) KeepActivingWithContext(tarsCtx context.Context, serverInfo *ServerInfo, opts ...map[string]string) (ret int32, err error) {
 	var (
 		length int32
 		have   bool
 		ty     byte
 	)
 	buf := codec.NewBuffer()
-	err = buf.WriteString(app, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(serverName, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(version, 3)
+	err = serverInfo.WriteBlock(buf, 1)
 	if err != nil {
 		return ret, err
 	}
@@ -240,14 +196,12 @@ func (obj *ServerF) ReportVersion(app string, serverName string, version string,
 		contextMap = opts[0]
 		statusMap = opts[1]
 	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
 
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "reportVersion", buf.ToBytes(), statusMap, contextMap, tarsResp)
+	tarsResp := new(requestf.ResponsePacket)
+	err = obj.servant.TarsInvoke(tarsCtx, 0, "keepActiving", buf.ToBytes(), statusMap, contextMap, tarsResp)
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -275,10 +229,50 @@ func (obj *ServerF) ReportVersion(app string, serverName string, version string,
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
 	return ret, nil
+}
+
+// KeepActivingOneWayWithContext is the proxy function for the method defined in the tars file, with the context
+func (obj *ServerF) KeepActivingOneWayWithContext(tarsCtx context.Context, serverInfo *ServerInfo, opts ...map[string]string) (ret int32, err error) {
+	var (
+		length int32
+		have   bool
+		ty     byte
+	)
+	buf := codec.NewBuffer()
+	err = serverInfo.WriteBlock(buf, 1)
+	if err != nil {
+		return ret, err
+	}
+
+	var statusMap map[string]string
+	var contextMap map[string]string
+	if len(opts) == 1 {
+		contextMap = opts[0]
+	} else if len(opts) == 2 {
+		contextMap = opts[0]
+		statusMap = opts[1]
+	}
+
+	tarsResp := new(requestf.ResponsePacket)
+	err = obj.servant.TarsInvoke(tarsCtx, 1, "keepActiving", buf.ToBytes(), statusMap, contextMap, tarsResp)
+	if err != nil {
+		return ret, err
+	}
+
+	_ = length
+	_ = have
+	_ = ty
+	return ret, nil
+}
+
+// ReportVersion is the proxy function for the method defined in the tars file, with the context
+func (obj *ServerF) ReportVersion(app string, serverName string, version string, opts ...map[string]string) (int32, error) {
+	return obj.ReportVersionWithContext(context.Background(), app, serverName, version, opts...)
 }
 
 // ReportVersionWithContext is the proxy function for the method defined in the tars file, with the context
@@ -318,7 +312,6 @@ func (obj *ServerF) ReportVersionWithContext(tarsCtx context.Context, app string
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -346,6 +339,7 @@ func (obj *ServerF) ReportVersionWithContext(tarsCtx context.Context, app string
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -390,6 +384,46 @@ func (obj *ServerF) ReportVersionOneWayWithContext(tarsCtx context.Context, app 
 		return ret, err
 	}
 
+	_ = length
+	_ = have
+	_ = ty
+	return ret, nil
+}
+
+// GetLatestKeepAliveTime is the proxy function for the method defined in the tars file, with the context
+func (obj *ServerF) GetLatestKeepAliveTime(opts ...map[string]string) (uint32, error) {
+	return obj.GetLatestKeepAliveTimeWithContext(context.Background(), opts...)
+}
+
+// GetLatestKeepAliveTimeWithContext is the proxy function for the method defined in the tars file, with the context
+func (obj *ServerF) GetLatestKeepAliveTimeWithContext(tarsCtx context.Context, opts ...map[string]string) (ret uint32, err error) {
+	var (
+		length int32
+		have   bool
+		ty     byte
+	)
+	buf := codec.NewBuffer()
+
+	var statusMap map[string]string
+	var contextMap map[string]string
+	if len(opts) == 1 {
+		contextMap = opts[0]
+	} else if len(opts) == 2 {
+		contextMap = opts[0]
+		statusMap = opts[1]
+	}
+
+	tarsResp := new(requestf.ResponsePacket)
+	err = obj.servant.TarsInvoke(tarsCtx, 0, "getLatestKeepAliveTime", buf.ToBytes(), statusMap, contextMap, tarsResp)
+	if err != nil {
+		return ret, err
+	}
+	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
+	err = readBuf.ReadUint32(&ret, 0, true)
+	if err != nil {
+		return ret, err
+	}
+
 	if len(opts) == 1 {
 		for k := range contextMap {
 			delete(contextMap, k)
@@ -411,34 +445,41 @@ func (obj *ServerF) ReportVersionOneWayWithContext(tarsCtx context.Context, app 
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
 	return ret, nil
 }
 
-// SetServant sets servant for the service.
-func (obj *ServerF) SetServant(servant m.Servant) {
-	obj.servant = servant
-}
+// GetLatestKeepAliveTimeOneWayWithContext is the proxy function for the method defined in the tars file, with the context
+func (obj *ServerF) GetLatestKeepAliveTimeOneWayWithContext(tarsCtx context.Context, opts ...map[string]string) (ret uint32, err error) {
+	var (
+		length int32
+		have   bool
+		ty     byte
+	)
+	buf := codec.NewBuffer()
 
-// TarsSetTimeout sets the timeout for the servant which is in ms.
-func (obj *ServerF) TarsSetTimeout(timeout int) {
-	obj.servant.TarsSetTimeout(timeout)
-}
+	var statusMap map[string]string
+	var contextMap map[string]string
+	if len(opts) == 1 {
+		contextMap = opts[0]
+	} else if len(opts) == 2 {
+		contextMap = opts[0]
+		statusMap = opts[1]
+	}
 
-// TarsSetProtocol sets the protocol for the servant.
-func (obj *ServerF) TarsSetProtocol(p m.Protocol) {
-	obj.servant.TarsSetProtocol(p)
-}
+	tarsResp := new(requestf.ResponsePacket)
+	err = obj.servant.TarsInvoke(tarsCtx, 1, "getLatestKeepAliveTime", buf.ToBytes(), statusMap, contextMap, tarsResp)
+	if err != nil {
+		return ret, err
+	}
 
-type ServerFServant interface {
-	KeepAlive(serverInfo *ServerInfo) (ret int32, err error)
-	ReportVersion(app string, serverName string, version string) (ret int32, err error)
-}
-type ServerFServantWithContext interface {
-	KeepAlive(tarsCtx context.Context, serverInfo *ServerInfo) (ret int32, err error)
-	ReportVersion(tarsCtx context.Context, app string, serverName string, version string) (ret int32, err error)
+	_ = length
+	_ = have
+	_ = ty
+	return ret, nil
 }
 
 // Dispatch is used to call the server side implement for the method defined in the tars file. withContext shows using context or not.
@@ -453,14 +494,11 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 	switch tarsReq.SFuncName {
 	case "keepAlive":
 		var serverInfo ServerInfo
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = serverInfo.ReadBlock(readBuf, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -473,7 +511,6 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -489,7 +526,6 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -503,7 +539,6 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 			imp := val.(ServerFServantWithContext)
 			funRet, err = imp.KeepAlive(tarsCtx, &serverInfo)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -515,7 +550,95 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 			if err != nil {
 				return err
 			}
+		} else if tarsReq.IVersion == basef.TUPVERSION {
+			rspTup := tup.NewUniAttribute()
 
+			err = buf.WriteInt32(funRet, 0)
+			if err != nil {
+				return err
+			}
+
+			rspTup.PutBuffer("", buf.ToBytes())
+			rspTup.PutBuffer("tars_ret", buf.ToBytes())
+
+			buf.Reset()
+			err = rspTup.Encode(buf)
+			if err != nil {
+				return err
+			}
+		} else if tarsReq.IVersion == basef.JSONVERSION {
+			rspJson := map[string]interface{}{}
+			rspJson["tars_ret"] = funRet
+
+			var rspByte []byte
+			if rspByte, err = json.Marshal(rspJson); err != nil {
+				return err
+			}
+
+			buf.Reset()
+			err = buf.WriteSliceUint8(rspByte)
+			if err != nil {
+				return err
+			}
+		}
+	case "keepActiving":
+		var serverInfo ServerInfo
+		if tarsReq.IVersion == basef.TARSVERSION {
+			err = serverInfo.ReadBlock(readBuf, 1, true)
+			if err != nil {
+				return err
+			}
+		} else if tarsReq.IVersion == basef.TUPVERSION {
+			reqTup := tup.NewUniAttribute()
+			reqTup.Decode(readBuf)
+
+			var tupBuffer []byte
+
+			reqTup.GetBuffer("serverInfo", &tupBuffer)
+			readBuf.Reset(tupBuffer)
+			err = serverInfo.ReadBlock(readBuf, 0, true)
+			if err != nil {
+				return err
+			}
+		} else if tarsReq.IVersion == basef.JSONVERSION {
+			var jsonData map[string]interface{}
+			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
+			decoder.UseNumber()
+			err = decoder.Decode(&jsonData)
+			if err != nil {
+				return fmt.Errorf("decode reqpacket failed, error: %+v", err)
+			}
+			{
+				jsonStr, _ := json.Marshal(jsonData["serverInfo"])
+				serverInfo.ResetDefault()
+				if err = json.Unmarshal(jsonStr, &serverInfo); err != nil {
+					return err
+				}
+			}
+		} else {
+			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
+			return err
+		}
+
+		var funRet int32
+		if !withContext {
+			imp := val.(ServerFServant)
+			funRet, err = imp.KeepActiving(&serverInfo)
+		} else {
+			imp := val.(ServerFServantWithContext)
+			funRet, err = imp.KeepActiving(tarsCtx, &serverInfo)
+		}
+		if err != nil {
+			return err
+		}
+
+		if tarsReq.IVersion == basef.TARSVERSION {
+			buf.Reset()
+
+			err = buf.WriteInt32(funRet, 0)
+			if err != nil {
+				return err
+			}
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -551,24 +674,19 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 		var app string
 		var serverName string
 		var version string
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&app, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&serverName, 2, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&version, 3, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -595,7 +713,6 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -622,7 +739,6 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -636,7 +752,6 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 			imp := val.(ServerFServantWithContext)
 			funRet, err = imp.ReportVersion(tarsCtx, app, serverName, version)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -648,11 +763,61 @@ func (obj *ServerF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
 			err = buf.WriteInt32(funRet, 0)
+			if err != nil {
+				return err
+			}
+
+			rspTup.PutBuffer("", buf.ToBytes())
+			rspTup.PutBuffer("tars_ret", buf.ToBytes())
+
+			buf.Reset()
+			err = rspTup.Encode(buf)
+			if err != nil {
+				return err
+			}
+		} else if tarsReq.IVersion == basef.JSONVERSION {
+			rspJson := map[string]interface{}{}
+			rspJson["tars_ret"] = funRet
+
+			var rspByte []byte
+			if rspByte, err = json.Marshal(rspJson); err != nil {
+				return err
+			}
+
+			buf.Reset()
+			err = buf.WriteSliceUint8(rspByte)
+			if err != nil {
+				return err
+			}
+		}
+	case "getLatestKeepAliveTime":
+		var funRet uint32
+		if !withContext {
+			imp := val.(ServerFServant)
+			funRet, err = imp.GetLatestKeepAliveTime()
+		} else {
+			imp := val.(ServerFServantWithContext)
+			funRet, err = imp.GetLatestKeepAliveTime(tarsCtx)
+		}
+		if err != nil {
+			return err
+		}
+
+		if tarsReq.IVersion == basef.TARSVERSION {
+			buf.Reset()
+
+			err = buf.WriteUint32(funRet, 0)
+			if err != nil {
+				return err
+			}
+		} else if tarsReq.IVersion == basef.TUPVERSION {
+			rspTup := tup.NewUniAttribute()
+
+			err = buf.WriteUint32(funRet, 0)
 			if err != nil {
 				return err
 			}
