@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from PropertyF.tars
 // Package propertyf comment
-// This file was generated by tars2go 1.1.10
-// Generated from PropertyF.tars
 package propertyf
 
 import (
@@ -15,14 +15,14 @@ var _ = codec.FromInt8
 
 // StatPropMsgHead struct implement
 type StatPropMsgHead struct {
-	ModuleName   string `json:"moduleName"`
-	Ip           string `json:"ip"`
-	PropertyName string `json:"propertyName"`
-	SetName      string `json:"setName"`
-	SetArea      string `json:"setArea"`
-	SetID        string `json:"setID"`
-	SContainer   string `json:"sContainer"`
-	IPropertyVer int32  `json:"iPropertyVer"`
+	ModuleName   string `json:"moduleName" tars:"moduleName,tag:0,require:true"`
+	Ip           string `json:"ip" tars:"ip,tag:1,require:true"`
+	PropertyName string `json:"propertyName" tars:"propertyName,tag:2,require:true"`
+	SetName      string `json:"setName" tars:"setName,tag:3,require:false"`
+	SetArea      string `json:"setArea" tars:"setArea,tag:4,require:false"`
+	SetID        string `json:"setID" tars:"setID,tag:5,require:false"`
+	SContainer   string `json:"sContainer" tars:"sContainer,tag:6,require:false"`
+	IPropertyVer int32  `json:"iPropertyVer" tars:"iPropertyVer,tag:7,require:false"`
 }
 
 func (st *StatPropMsgHead) ResetDefault() {
@@ -120,7 +120,6 @@ func (st *StatPropMsgHead) ReadBlock(readBuf *codec.Reader, tag byte, require bo
 
 // WriteTo encode struct to buffer
 func (st *StatPropMsgHead) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.ModuleName, 0)
 	if err != nil {
 		return err
@@ -136,29 +135,39 @@ func (st *StatPropMsgHead) WriteTo(buf *codec.Buffer) (err error) {
 		return err
 	}
 
-	err = buf.WriteString(st.SetName, 3)
-	if err != nil {
-		return err
+	if st.SetName != "" {
+		err = buf.WriteString(st.SetName, 3)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SetArea, 4)
-	if err != nil {
-		return err
+	if st.SetArea != "" {
+		err = buf.WriteString(st.SetArea, 4)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SetID, 5)
-	if err != nil {
-		return err
+	if st.SetID != "" {
+		err = buf.WriteString(st.SetID, 5)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SContainer, 6)
-	if err != nil {
-		return err
+	if st.SContainer != "" {
+		err = buf.WriteString(st.SContainer, 6)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteInt32(st.IPropertyVer, 7)
-	if err != nil {
-		return err
+	if st.IPropertyVer != 1 {
+		err = buf.WriteInt32(st.IPropertyVer, 7)
+		if err != nil {
+			return err
+		}
 	}
 
 	return err
@@ -186,8 +195,8 @@ func (st *StatPropMsgHead) WriteBlock(buf *codec.Buffer, tag byte) error {
 
 // StatPropInfo struct implement
 type StatPropInfo struct {
-	Policy string `json:"policy"`
-	Value  string `json:"value"`
+	Policy string `json:"policy" tars:"policy,tag:0,require:true"`
+	Value  string `json:"value" tars:"value,tag:1,require:true"`
 }
 
 func (st *StatPropInfo) ResetDefault() {
@@ -254,7 +263,6 @@ func (st *StatPropInfo) ReadBlock(readBuf *codec.Reader, tag byte, require bool)
 
 // WriteTo encode struct to buffer
 func (st *StatPropInfo) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.Policy, 0)
 	if err != nil {
 		return err
@@ -290,7 +298,7 @@ func (st *StatPropInfo) WriteBlock(buf *codec.Buffer, tag byte) error {
 
 // StatPropMsgBody struct implement
 type StatPropMsgBody struct {
-	VInfo []StatPropInfo `json:"vInfo"`
+	VInfo []StatPropInfo `json:"vInfo" tars:"vInfo,tag:0,require:true"`
 }
 
 func (st *StatPropMsgBody) ResetDefault() {
@@ -310,34 +318,28 @@ func (st *StatPropMsgBody) ReadFrom(readBuf *codec.Reader) error {
 	if err != nil {
 		return err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return err
 		}
-
 		st.VInfo = make([]StatPropInfo, length)
 		for i0, e0 := int32(0), length; i0 < e0; i0++ {
-
-			err = st.VInfo[i0].ReadBlock(readBuf, 0, false)
+			err = st.VInfo[i0].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return err
 		}
-
 	}
 
 	_ = err
@@ -381,24 +383,19 @@ func (st *StatPropMsgBody) ReadBlock(readBuf *codec.Reader, tag byte, require bo
 
 // WriteTo encode struct to buffer
 func (st *StatPropMsgBody) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteHead(codec.LIST, 0)
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(st.VInfo)), 0)
 	if err != nil {
 		return err
 	}
-
 	for _, v := range st.VInfo {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return err
 		}
-
 	}
 
 	return err
