@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from PropertyF.tars
 // Package propertyf comment
-// This file was generated by tars2go 1.1.10
-// Generated from PropertyF.tars
 package propertyf
 
 import (
@@ -8,108 +8,65 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	m "github.com/TarsCloud/TarsGo/tars/model"
+
+	"github.com/TarsCloud/TarsGo/tars/model"
 	"github.com/TarsCloud/TarsGo/tars/protocol/codec"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/requestf"
 	"github.com/TarsCloud/TarsGo/tars/protocol/tup"
 	"github.com/TarsCloud/TarsGo/tars/util/current"
+	"github.com/TarsCloud/TarsGo/tars/util/endpoint"
 	"github.com/TarsCloud/TarsGo/tars/util/tools"
-	"unsafe"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var (
 	_ = fmt.Errorf
 	_ = codec.FromInt8
-	_ = unsafe.Pointer(nil)
 	_ = bytes.ErrTooLarge
 )
 
+type PropertyFServant interface {
+	ReportPropMsg(statmsg map[StatPropMsgHead]StatPropMsgBody) (ret int32, err error)
+}
+
+type PropertyFServantWithContext interface {
+	ReportPropMsg(tarsCtx context.Context, statmsg map[StatPropMsgHead]StatPropMsgBody) (ret int32, err error)
+}
+
 // PropertyF struct
 type PropertyF struct {
-	servant m.Servant
+	servant model.Servant
+}
+
+// NewPropertyF creates a new PropertyF servant.
+func NewPropertyF() *PropertyF {
+	return new(PropertyF)
+}
+
+// SetServant sets servant for the service.
+func (obj *PropertyF) SetServant(servant model.Servant) {
+	obj.servant = servant
+}
+
+// TarsSetTimeout sets the timeout for the servant which is in ms.
+func (obj *PropertyF) TarsSetTimeout(timeout int) {
+	obj.servant.TarsSetTimeout(timeout)
+}
+
+// TarsSetProtocol sets the protocol for the servant.
+func (obj *PropertyF) TarsSetProtocol(p model.Protocol) {
+	obj.servant.TarsSetProtocol(p)
+}
+
+// Endpoints returns all active endpoint.Endpoint
+func (obj *PropertyF) Endpoints() []*endpoint.Endpoint {
+	return obj.servant.Endpoints()
 }
 
 // ReportPropMsg is the proxy function for the method defined in the tars file, with the context
-func (obj *PropertyF) ReportPropMsg(statmsg map[StatPropMsgHead]StatPropMsgBody, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteHead(codec.MAP, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(statmsg)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for k1, v1 := range statmsg {
-
-		err = k1.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-		err = v1.WriteBlock(buf, 1)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "reportPropMsg", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *PropertyF) ReportPropMsg(statmsg map[StatPropMsgHead]StatPropMsgBody, opts ...map[string]string) (int32, error) {
+	return obj.ReportPropMsgWithContext(context.Background(), statmsg, opts...)
 }
 
 // ReportPropMsgWithContext is the proxy function for the method defined in the tars file, with the context
@@ -124,24 +81,19 @@ func (obj *PropertyF) ReportPropMsgWithContext(tarsCtx context.Context, statmsg 
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(statmsg)), 0)
 	if err != nil {
 		return ret, err
 	}
-
-	for k2, v2 := range statmsg {
-
-		err = k2.WriteBlock(buf, 0)
+	for k1, v1 := range statmsg {
+		err = k1.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
-		err = v2.WriteBlock(buf, 1)
+		err = v1.WriteBlock(buf, 1)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -158,7 +110,6 @@ func (obj *PropertyF) ReportPropMsgWithContext(tarsCtx context.Context, statmsg 
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -186,6 +137,7 @@ func (obj *PropertyF) ReportPropMsgWithContext(tarsCtx context.Context, statmsg 
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -204,24 +156,19 @@ func (obj *PropertyF) ReportPropMsgOneWayWithContext(tarsCtx context.Context, st
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(statmsg)), 0)
 	if err != nil {
 		return ret, err
 	}
-
-	for k3, v3 := range statmsg {
-
-		err = k3.WriteBlock(buf, 0)
+	for k2, v2 := range statmsg {
+		err = k2.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
-		err = v3.WriteBlock(buf, 1)
+		err = v2.WriteBlock(buf, 1)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -239,53 +186,10 @@ func (obj *PropertyF) ReportPropMsgOneWayWithContext(tarsCtx context.Context, st
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
 	return ret, nil
-}
-
-// SetServant sets servant for the service.
-func (obj *PropertyF) SetServant(servant m.Servant) {
-	obj.servant = servant
-}
-
-// TarsSetTimeout sets the timeout for the servant which is in ms.
-func (obj *PropertyF) TarsSetTimeout(timeout int) {
-	obj.servant.TarsSetTimeout(timeout)
-}
-
-// TarsSetProtocol sets the protocol for the servant.
-func (obj *PropertyF) TarsSetProtocol(p m.Protocol) {
-	obj.servant.TarsSetProtocol(p)
-}
-
-type PropertyFServant interface {
-	ReportPropMsg(statmsg map[StatPropMsgHead]StatPropMsgBody) (ret int32, err error)
-}
-type PropertyFServantWithContext interface {
-	ReportPropMsg(tarsCtx context.Context, statmsg map[StatPropMsgHead]StatPropMsgBody) (ret int32, err error)
 }
 
 // Dispatch is used to call the server side implement for the method defined in the tars file. withContext shows using context or not.
@@ -301,35 +205,28 @@ func (obj *PropertyF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq
 	case "reportPropMsg":
 		var statmsg map[StatPropMsgHead]StatPropMsgBody
 		statmsg = make(map[StatPropMsgHead]StatPropMsgBody)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			_, err = readBuf.SkipTo(codec.MAP, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadInt32(&length, 0, true)
 			if err != nil {
 				return err
 			}
-
 			statmsg = make(map[StatPropMsgHead]StatPropMsgBody)
-			for i4, e4 := int32(0), length; i4 < e4; i4++ {
-				var k4 StatPropMsgHead
-				var v4 StatPropMsgBody
-
-				err = k4.ReadBlock(readBuf, 0, false)
+			for i3, e3 := int32(0), length; i3 < e3; i3++ {
+				var k3 StatPropMsgHead
+				var v3 StatPropMsgBody
+				err = k3.ReadBlock(readBuf, 0, true)
 				if err != nil {
 					return err
 				}
-
-				err = v4.ReadBlock(readBuf, 1, false)
+				err = v3.ReadBlock(readBuf, 1, true)
 				if err != nil {
 					return err
 				}
-
-				statmsg[k4] = v4
+				statmsg[k3] = v3
 			}
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
@@ -343,28 +240,23 @@ func (obj *PropertyF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadInt32(&length, 0, true)
 			if err != nil {
 				return err
 			}
-
 			statmsg = make(map[StatPropMsgHead]StatPropMsgBody)
-			for i5, e5 := int32(0), length; i5 < e5; i5++ {
-				var k5 StatPropMsgHead
-				var v5 StatPropMsgBody
-
-				err = k5.ReadBlock(readBuf, 0, false)
+			for i4, e4 := int32(0), length; i4 < e4; i4++ {
+				var k4 StatPropMsgHead
+				var v4 StatPropMsgBody
+				err = k4.ReadBlock(readBuf, 0, true)
 				if err != nil {
 					return err
 				}
-
-				err = v5.ReadBlock(readBuf, 1, false)
+				err = v4.ReadBlock(readBuf, 1, true)
 				if err != nil {
 					return err
 				}
-
-				statmsg[k5] = v5
+				statmsg[k4] = v4
 			}
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
@@ -380,7 +272,6 @@ func (obj *PropertyF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -394,7 +285,6 @@ func (obj *PropertyF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq
 			imp := val.(PropertyFServantWithContext)
 			funRet, err = imp.ReportPropMsg(tarsCtx, statmsg)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -406,7 +296,6 @@ func (obj *PropertyF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
