@@ -1,126 +1,84 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from QueryF.tars
 // Package queryf comment
-// This file was generated by tars2go 1.1.10
-// Generated from QueryF.tars
 package queryf
 
 import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"github.com/TarsCloud/TarsGo/tars/protocol/res/endpointf"
 	"fmt"
-	m "github.com/TarsCloud/TarsGo/tars/model"
+
+	"github.com/TarsCloud/TarsGo/tars/model"
 	"github.com/TarsCloud/TarsGo/tars/protocol/codec"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/requestf"
 	"github.com/TarsCloud/TarsGo/tars/protocol/tup"
 	"github.com/TarsCloud/TarsGo/tars/util/current"
+	"github.com/TarsCloud/TarsGo/tars/util/endpoint"
 	"github.com/TarsCloud/TarsGo/tars/util/tools"
-	"unsafe"
+
+	"github.com/TarsCloud/TarsGo/tars/protocol/res/endpointf"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var (
 	_ = fmt.Errorf
 	_ = codec.FromInt8
-	_ = unsafe.Pointer(nil)
 	_ = bytes.ErrTooLarge
 )
 
+type QueryFServant interface {
+	FindObjectById(id string) (ret []endpointf.EndpointF, err error)
+	FindObjectById4Any(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+	FindObjectById4All(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+	FindObjectByIdInSameGroup(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+	FindObjectByIdInSameStation(id string, sStation string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+	FindObjectByIdInSameSet(id string, setId string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+}
+
+type QueryFServantWithContext interface {
+	FindObjectById(tarsCtx context.Context, id string) (ret []endpointf.EndpointF, err error)
+	FindObjectById4Any(tarsCtx context.Context, id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+	FindObjectById4All(tarsCtx context.Context, id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+	FindObjectByIdInSameGroup(tarsCtx context.Context, id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+	FindObjectByIdInSameStation(tarsCtx context.Context, id string, sStation string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+	FindObjectByIdInSameSet(tarsCtx context.Context, id string, setId string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
+}
+
 // QueryF struct
 type QueryF struct {
-	servant m.Servant
+	servant model.Servant
+}
+
+// NewQueryF creates a new QueryF servant.
+func NewQueryF() *QueryF {
+	return new(QueryF)
+}
+
+// SetServant sets servant for the service.
+func (obj *QueryF) SetServant(servant model.Servant) {
+	obj.servant = servant
+}
+
+// TarsSetTimeout sets the timeout for the servant which is in ms.
+func (obj *QueryF) TarsSetTimeout(timeout int) {
+	obj.servant.TarsSetTimeout(timeout)
+}
+
+// TarsSetProtocol sets the protocol for the servant.
+func (obj *QueryF) TarsSetProtocol(p model.Protocol) {
+	obj.servant.TarsSetProtocol(p)
+}
+
+// Endpoints returns all active endpoint.Endpoint
+func (obj *QueryF) Endpoints() []*endpoint.Endpoint {
+	return obj.servant.Endpoints()
 }
 
 // FindObjectById is the proxy function for the method defined in the tars file, with the context
-func (obj *QueryF) FindObjectById(id string, opts ...map[string]string) (ret []endpointf.EndpointF, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(id, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "findObjectById", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	_, ty, err = readBuf.SkipToNoCheck(0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		ret = make([]endpointf.EndpointF, length)
-		for i0, e0 := int32(0), length; i0 < e0; i0++ {
-
-			err = ret[i0].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *QueryF) FindObjectById(id string, opts ...map[string]string) ([]endpointf.EndpointF, error) {
+	return obj.FindObjectByIdWithContext(context.Background(), id, opts...)
 }
 
 // FindObjectByIdWithContext is the proxy function for the method defined in the tars file, with the context
@@ -150,40 +108,33 @@ func (obj *QueryF) FindObjectByIdWithContext(tarsCtx context.Context, id string,
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	_, ty, err = readBuf.SkipToNoCheck(0, true)
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		ret = make([]endpointf.EndpointF, length)
-		for i1, e1 := int32(0), length; i1 < e1; i1++ {
-
-			err = ret[i1].ReadBlock(readBuf, 0, false)
+		for i0, e0 := int32(0), length; i0 < e0; i0++ {
+			err = ret[i0].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -207,6 +158,7 @@ func (obj *QueryF) FindObjectByIdWithContext(tarsCtx context.Context, id string,
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -241,27 +193,6 @@ func (obj *QueryF) FindObjectByIdOneWayWithContext(tarsCtx context.Context, id s
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -269,171 +200,8 @@ func (obj *QueryF) FindObjectByIdOneWayWithContext(tarsCtx context.Context, id s
 }
 
 // FindObjectById4Any is the proxy function for the method defined in the tars file, with the context
-func (obj *QueryF) FindObjectById4Any(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(id, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteHead(codec.LIST, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*activeEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *activeEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	err = buf.WriteHead(codec.LIST, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *inactiveEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "findObjectById4Any", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(2, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*activeEp = make([]endpointf.EndpointF, length)
-		for i2, e2 := int32(0), length; i2 < e2; i2++ {
-
-			err = (*activeEp)[i2].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(3, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i3, e3 := int32(0), length; i3 < e3; i3++ {
-
-			err = (*inactiveEp)[i3].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *QueryF) FindObjectById4Any(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (int32, error) {
+	return obj.FindObjectById4AnyWithContext(context.Background(), id, activeEp, inactiveEp, opts...)
 }
 
 // FindObjectById4AnyWithContext is the proxy function for the method defined in the tars file, with the context
@@ -453,38 +221,30 @@ func (obj *QueryF) FindObjectById4AnyWithContext(tarsCtx context.Context, id str
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 3)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -501,7 +261,6 @@ func (obj *QueryF) FindObjectById4AnyWithContext(tarsCtx context.Context, id str
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -512,68 +271,55 @@ func (obj *QueryF) FindObjectById4AnyWithContext(tarsCtx context.Context, id str
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*activeEp = make([]endpointf.EndpointF, length)
-		for i4, e4 := int32(0), length; i4 < e4; i4++ {
-
-			err = (*activeEp)[i4].ReadBlock(readBuf, 0, false)
+		for i1, e1 := int32(0), length; i1 < e1; i1++ {
+			err = (*activeEp)[i1].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
-
 	_, ty, err = readBuf.SkipToNoCheck(3, true)
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i5, e5 := int32(0), length; i5 < e5; i5++ {
-
-			err = (*inactiveEp)[i5].ReadBlock(readBuf, 0, false)
+		for i2, e2 := int32(0), length; i2 < e2; i2++ {
+			err = (*inactiveEp)[i2].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -597,6 +343,7 @@ func (obj *QueryF) FindObjectById4AnyWithContext(tarsCtx context.Context, id str
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -620,38 +367,30 @@ func (obj *QueryF) FindObjectById4AnyOneWayWithContext(tarsCtx context.Context, 
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 3)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -669,27 +408,6 @@ func (obj *QueryF) FindObjectById4AnyOneWayWithContext(tarsCtx context.Context, 
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -697,171 +415,8 @@ func (obj *QueryF) FindObjectById4AnyOneWayWithContext(tarsCtx context.Context, 
 }
 
 // FindObjectById4All is the proxy function for the method defined in the tars file, with the context
-func (obj *QueryF) FindObjectById4All(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(id, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteHead(codec.LIST, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*activeEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *activeEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	err = buf.WriteHead(codec.LIST, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *inactiveEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "findObjectById4All", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(2, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*activeEp = make([]endpointf.EndpointF, length)
-		for i6, e6 := int32(0), length; i6 < e6; i6++ {
-
-			err = (*activeEp)[i6].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(3, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i7, e7 := int32(0), length; i7 < e7; i7++ {
-
-			err = (*inactiveEp)[i7].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *QueryF) FindObjectById4All(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (int32, error) {
+	return obj.FindObjectById4AllWithContext(context.Background(), id, activeEp, inactiveEp, opts...)
 }
 
 // FindObjectById4AllWithContext is the proxy function for the method defined in the tars file, with the context
@@ -881,38 +436,30 @@ func (obj *QueryF) FindObjectById4AllWithContext(tarsCtx context.Context, id str
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 3)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -929,7 +476,6 @@ func (obj *QueryF) FindObjectById4AllWithContext(tarsCtx context.Context, id str
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -940,68 +486,55 @@ func (obj *QueryF) FindObjectById4AllWithContext(tarsCtx context.Context, id str
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*activeEp = make([]endpointf.EndpointF, length)
-		for i8, e8 := int32(0), length; i8 < e8; i8++ {
-
-			err = (*activeEp)[i8].ReadBlock(readBuf, 0, false)
+		for i3, e3 := int32(0), length; i3 < e3; i3++ {
+			err = (*activeEp)[i3].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
-
 	_, ty, err = readBuf.SkipToNoCheck(3, true)
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i9, e9 := int32(0), length; i9 < e9; i9++ {
-
-			err = (*inactiveEp)[i9].ReadBlock(readBuf, 0, false)
+		for i4, e4 := int32(0), length; i4 < e4; i4++ {
+			err = (*inactiveEp)[i4].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -1025,6 +558,7 @@ func (obj *QueryF) FindObjectById4AllWithContext(tarsCtx context.Context, id str
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -1048,38 +582,30 @@ func (obj *QueryF) FindObjectById4AllOneWayWithContext(tarsCtx context.Context, 
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 3)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -1097,27 +623,6 @@ func (obj *QueryF) FindObjectById4AllOneWayWithContext(tarsCtx context.Context, 
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -1125,171 +630,8 @@ func (obj *QueryF) FindObjectById4AllOneWayWithContext(tarsCtx context.Context, 
 }
 
 // FindObjectByIdInSameGroup is the proxy function for the method defined in the tars file, with the context
-func (obj *QueryF) FindObjectByIdInSameGroup(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(id, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteHead(codec.LIST, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*activeEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *activeEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	err = buf.WriteHead(codec.LIST, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *inactiveEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "findObjectByIdInSameGroup", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(2, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*activeEp = make([]endpointf.EndpointF, length)
-		for i10, e10 := int32(0), length; i10 < e10; i10++ {
-
-			err = (*activeEp)[i10].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(3, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i11, e11 := int32(0), length; i11 < e11; i11++ {
-
-			err = (*inactiveEp)[i11].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *QueryF) FindObjectByIdInSameGroup(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (int32, error) {
+	return obj.FindObjectByIdInSameGroupWithContext(context.Background(), id, activeEp, inactiveEp, opts...)
 }
 
 // FindObjectByIdInSameGroupWithContext is the proxy function for the method defined in the tars file, with the context
@@ -1309,38 +651,30 @@ func (obj *QueryF) FindObjectByIdInSameGroupWithContext(tarsCtx context.Context,
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 3)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -1357,7 +691,6 @@ func (obj *QueryF) FindObjectByIdInSameGroupWithContext(tarsCtx context.Context,
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -1368,68 +701,55 @@ func (obj *QueryF) FindObjectByIdInSameGroupWithContext(tarsCtx context.Context,
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*activeEp = make([]endpointf.EndpointF, length)
-		for i12, e12 := int32(0), length; i12 < e12; i12++ {
-
-			err = (*activeEp)[i12].ReadBlock(readBuf, 0, false)
+		for i5, e5 := int32(0), length; i5 < e5; i5++ {
+			err = (*activeEp)[i5].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
-
 	_, ty, err = readBuf.SkipToNoCheck(3, true)
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i13, e13 := int32(0), length; i13 < e13; i13++ {
-
-			err = (*inactiveEp)[i13].ReadBlock(readBuf, 0, false)
+		for i6, e6 := int32(0), length; i6 < e6; i6++ {
+			err = (*inactiveEp)[i6].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -1453,6 +773,7 @@ func (obj *QueryF) FindObjectByIdInSameGroupWithContext(tarsCtx context.Context,
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -1476,38 +797,30 @@ func (obj *QueryF) FindObjectByIdInSameGroupOneWayWithContext(tarsCtx context.Co
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 3)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -1525,27 +838,6 @@ func (obj *QueryF) FindObjectByIdInSameGroupOneWayWithContext(tarsCtx context.Co
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -1553,176 +845,8 @@ func (obj *QueryF) FindObjectByIdInSameGroupOneWayWithContext(tarsCtx context.Co
 }
 
 // FindObjectByIdInSameStation is the proxy function for the method defined in the tars file, with the context
-func (obj *QueryF) FindObjectByIdInSameStation(id string, sStation string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(id, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(sStation, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteHead(codec.LIST, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*activeEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *activeEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	err = buf.WriteHead(codec.LIST, 4)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *inactiveEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "findObjectByIdInSameStation", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(3, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*activeEp = make([]endpointf.EndpointF, length)
-		for i14, e14 := int32(0), length; i14 < e14; i14++ {
-
-			err = (*activeEp)[i14].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(4, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i15, e15 := int32(0), length; i15 < e15; i15++ {
-
-			err = (*inactiveEp)[i15].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *QueryF) FindObjectByIdInSameStation(id string, sStation string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (int32, error) {
+	return obj.FindObjectByIdInSameStationWithContext(context.Background(), id, sStation, activeEp, inactiveEp, opts...)
 }
 
 // FindObjectByIdInSameStationWithContext is the proxy function for the method defined in the tars file, with the context
@@ -1747,38 +871,30 @@ func (obj *QueryF) FindObjectByIdInSameStationWithContext(tarsCtx context.Contex
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 4)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -1795,7 +911,6 @@ func (obj *QueryF) FindObjectByIdInSameStationWithContext(tarsCtx context.Contex
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -1806,68 +921,55 @@ func (obj *QueryF) FindObjectByIdInSameStationWithContext(tarsCtx context.Contex
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*activeEp = make([]endpointf.EndpointF, length)
-		for i16, e16 := int32(0), length; i16 < e16; i16++ {
-
-			err = (*activeEp)[i16].ReadBlock(readBuf, 0, false)
+		for i7, e7 := int32(0), length; i7 < e7; i7++ {
+			err = (*activeEp)[i7].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
-
 	_, ty, err = readBuf.SkipToNoCheck(4, true)
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i17, e17 := int32(0), length; i17 < e17; i17++ {
-
-			err = (*inactiveEp)[i17].ReadBlock(readBuf, 0, false)
+		for i8, e8 := int32(0), length; i8 < e8; i8++ {
+			err = (*inactiveEp)[i8].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -1891,6 +993,7 @@ func (obj *QueryF) FindObjectByIdInSameStationWithContext(tarsCtx context.Contex
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -1919,38 +1022,30 @@ func (obj *QueryF) FindObjectByIdInSameStationOneWayWithContext(tarsCtx context.
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 4)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -1968,27 +1063,6 @@ func (obj *QueryF) FindObjectByIdInSameStationOneWayWithContext(tarsCtx context.
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -1996,176 +1070,8 @@ func (obj *QueryF) FindObjectByIdInSameStationOneWayWithContext(tarsCtx context.
 }
 
 // FindObjectByIdInSameSet is the proxy function for the method defined in the tars file, with the context
-func (obj *QueryF) FindObjectByIdInSameSet(id string, setId string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteString(id, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteString(setId, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteHead(codec.LIST, 3)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*activeEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *activeEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	err = buf.WriteHead(codec.LIST, 4)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range *inactiveEp {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "findObjectByIdInSameSet", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(3, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*activeEp = make([]endpointf.EndpointF, length)
-		for i18, e18 := int32(0), length; i18 < e18; i18++ {
-
-			err = (*activeEp)[i18].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	_, ty, err = readBuf.SkipToNoCheck(4, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if ty == codec.LIST {
-		err = readBuf.ReadInt32(&length, 0, true)
-		if err != nil {
-			return ret, err
-		}
-
-		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i19, e19 := int32(0), length; i19 < e19; i19++ {
-
-			err = (*inactiveEp)[i19].ReadBlock(readBuf, 0, false)
-			if err != nil {
-				return ret, err
-			}
-
-		}
-	} else if ty == codec.SimpleList {
-		err = fmt.Errorf("not support SimpleList type")
-		if err != nil {
-			return ret, err
-		}
-
-	} else {
-		err = fmt.Errorf("require vector, but not")
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *QueryF) FindObjectByIdInSameSet(id string, setId string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF, opts ...map[string]string) (int32, error) {
+	return obj.FindObjectByIdInSameSetWithContext(context.Background(), id, setId, activeEp, inactiveEp, opts...)
 }
 
 // FindObjectByIdInSameSetWithContext is the proxy function for the method defined in the tars file, with the context
@@ -2190,38 +1096,30 @@ func (obj *QueryF) FindObjectByIdInSameSetWithContext(tarsCtx context.Context, i
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 4)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -2238,7 +1136,6 @@ func (obj *QueryF) FindObjectByIdInSameSetWithContext(tarsCtx context.Context, i
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -2249,68 +1146,55 @@ func (obj *QueryF) FindObjectByIdInSameSetWithContext(tarsCtx context.Context, i
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*activeEp = make([]endpointf.EndpointF, length)
-		for i20, e20 := int32(0), length; i20 < e20; i20++ {
-
-			err = (*activeEp)[i20].ReadBlock(readBuf, 0, false)
+		for i9, e9 := int32(0), length; i9 < e9; i9++ {
+			err = (*activeEp)[i9].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
-
 	_, ty, err = readBuf.SkipToNoCheck(4, true)
 	if err != nil {
 		return ret, err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return ret, err
 		}
-
 		*inactiveEp = make([]endpointf.EndpointF, length)
-		for i21, e21 := int32(0), length; i21 < e21; i21++ {
-
-			err = (*inactiveEp)[i21].ReadBlock(readBuf, 0, false)
+		for i10, e10 := int32(0), length; i10 < e10; i10++ {
+			err = (*inactiveEp)[i10].ReadBlock(readBuf, 0, true)
 			if err != nil {
 				return ret, err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
 		err = fmt.Errorf("not support SimpleList type")
 		if err != nil {
 			return ret, err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	if len(opts) == 1 {
@@ -2334,6 +1218,7 @@ func (obj *QueryF) FindObjectByIdInSameSetWithContext(tarsCtx context.Context, i
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -2362,38 +1247,30 @@ func (obj *QueryF) FindObjectByIdInSameSetOneWayWithContext(tarsCtx context.Cont
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*activeEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *activeEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.LIST, 4)
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(*inactiveEp)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range *inactiveEp {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -2411,63 +1288,10 @@ func (obj *QueryF) FindObjectByIdInSameSetOneWayWithContext(tarsCtx context.Cont
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
 	return ret, nil
-}
-
-// SetServant sets servant for the service.
-func (obj *QueryF) SetServant(servant m.Servant) {
-	obj.servant = servant
-}
-
-// TarsSetTimeout sets the timeout for the servant which is in ms.
-func (obj *QueryF) TarsSetTimeout(timeout int) {
-	obj.servant.TarsSetTimeout(timeout)
-}
-
-// TarsSetProtocol sets the protocol for the servant.
-func (obj *QueryF) TarsSetProtocol(p m.Protocol) {
-	obj.servant.TarsSetProtocol(p)
-}
-
-type QueryFServant interface {
-	FindObjectById(id string) (ret []endpointf.EndpointF, err error)
-	FindObjectById4Any(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-	FindObjectById4All(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-	FindObjectByIdInSameGroup(id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-	FindObjectByIdInSameStation(id string, sStation string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-	FindObjectByIdInSameSet(id string, setId string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-}
-type QueryFServantWithContext interface {
-	FindObjectById(tarsCtx context.Context, id string) (ret []endpointf.EndpointF, err error)
-	FindObjectById4Any(tarsCtx context.Context, id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-	FindObjectById4All(tarsCtx context.Context, id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-	FindObjectByIdInSameGroup(tarsCtx context.Context, id string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-	FindObjectByIdInSameStation(tarsCtx context.Context, id string, sStation string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
-	FindObjectByIdInSameSet(tarsCtx context.Context, id string, setId string, activeEp *[]endpointf.EndpointF, inactiveEp *[]endpointf.EndpointF) (ret int32, err error)
 }
 
 // Dispatch is used to call the server side implement for the method defined in the tars file. withContext shows using context or not.
@@ -2482,14 +1306,11 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 	switch tarsReq.SFuncName {
 	case "findObjectById":
 		var id string
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&id, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2502,7 +1323,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2517,7 +1337,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2531,7 +1350,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(QueryFServantWithContext)
 			funRet, err = imp.FindObjectById(tarsCtx, id)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2543,21 +1361,16 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(funRet)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range funRet {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2565,19 +1378,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(funRet)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range funRet {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 
 			rspTup.PutBuffer("", buf.ToBytes())
@@ -2609,14 +1418,11 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		activeEp = make([]endpointf.EndpointF, 0)
 		var inactiveEp []endpointf.EndpointF
 		inactiveEp = make([]endpointf.EndpointF, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&id, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2629,7 +1435,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2644,7 +1449,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2658,7 +1462,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(QueryFServantWithContext)
 			funRet, err = imp.FindObjectById4Any(tarsCtx, id, &activeEp, &inactiveEp)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2670,45 +1473,34 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteHead(codec.LIST, 2)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 			err = buf.WriteHead(codec.LIST, 3)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2725,19 +1517,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("activeEp", buf.ToBytes())
 
@@ -2746,19 +1534,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("inactiveEp", buf.ToBytes())
 
@@ -2790,14 +1574,11 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		activeEp = make([]endpointf.EndpointF, 0)
 		var inactiveEp []endpointf.EndpointF
 		inactiveEp = make([]endpointf.EndpointF, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&id, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2810,7 +1591,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -2825,7 +1605,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -2839,7 +1618,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(QueryFServantWithContext)
 			funRet, err = imp.FindObjectById4All(tarsCtx, id, &activeEp, &inactiveEp)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -2851,45 +1629,34 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteHead(codec.LIST, 2)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 			err = buf.WriteHead(codec.LIST, 3)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -2906,19 +1673,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("activeEp", buf.ToBytes())
 
@@ -2927,19 +1690,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("inactiveEp", buf.ToBytes())
 
@@ -2971,14 +1730,11 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		activeEp = make([]endpointf.EndpointF, 0)
 		var inactiveEp []endpointf.EndpointF
 		inactiveEp = make([]endpointf.EndpointF, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&id, 1, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -2991,7 +1747,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -3006,7 +1761,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -3020,7 +1774,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(QueryFServantWithContext)
 			funRet, err = imp.FindObjectByIdInSameGroup(tarsCtx, id, &activeEp, &inactiveEp)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -3032,45 +1785,34 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteHead(codec.LIST, 2)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 			err = buf.WriteHead(codec.LIST, 3)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -3087,19 +1829,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("activeEp", buf.ToBytes())
 
@@ -3108,19 +1846,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("inactiveEp", buf.ToBytes())
 
@@ -3153,19 +1887,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		activeEp = make([]endpointf.EndpointF, 0)
 		var inactiveEp []endpointf.EndpointF
 		inactiveEp = make([]endpointf.EndpointF, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&id, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&sStation, 2, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -3185,7 +1915,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -3206,7 +1935,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -3220,7 +1948,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(QueryFServantWithContext)
 			funRet, err = imp.FindObjectByIdInSameStation(tarsCtx, id, sStation, &activeEp, &inactiveEp)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -3232,45 +1959,34 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteHead(codec.LIST, 3)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 			err = buf.WriteHead(codec.LIST, 4)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -3287,19 +2003,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("activeEp", buf.ToBytes())
 
@@ -3308,19 +2020,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("inactiveEp", buf.ToBytes())
 
@@ -3353,19 +2061,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 		activeEp = make([]endpointf.EndpointF, 0)
 		var inactiveEp []endpointf.EndpointF
 		inactiveEp = make([]endpointf.EndpointF, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			err = readBuf.ReadString(&id, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadString(&setId, 2, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -3385,7 +2089,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -3406,7 +2109,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -3420,7 +2122,6 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			imp := val.(QueryFServantWithContext)
 			funRet, err = imp.FindObjectByIdInSameSet(tarsCtx, id, setId, &activeEp, &inactiveEp)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -3432,45 +2133,34 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteHead(codec.LIST, 3)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 			err = buf.WriteHead(codec.LIST, 4)
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -3487,19 +2177,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(activeEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range activeEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("activeEp", buf.ToBytes())
 
@@ -3508,19 +2194,15 @@ func (obj *QueryF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *r
 			if err != nil {
 				return err
 			}
-
 			err = buf.WriteInt32(int32(len(inactiveEp)), 0)
 			if err != nil {
 				return err
 			}
-
 			for _, v := range inactiveEp {
-
 				err = v.WriteBlock(buf, 0)
 				if err != nil {
 					return err
 				}
-
 			}
 			rspTup.PutBuffer("inactiveEp", buf.ToBytes())
 
