@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from RequestF.tars
 // Package requestf comment
-// This file was generated by tars2go 1.1.10
-// Generated from RequestF.tars
 package requestf
 
 import (
@@ -15,16 +15,16 @@ var _ = codec.FromInt8
 
 // RequestPacket struct implement
 type RequestPacket struct {
-	IVersion     int16             `json:"iVersion"`
-	CPacketType  int8              `json:"cPacketType"`
-	IMessageType int32             `json:"iMessageType"`
-	IRequestId   int32             `json:"iRequestId"`
-	SServantName string            `json:"sServantName"`
-	SFuncName    string            `json:"sFuncName"`
-	SBuffer      []int8            `json:"sBuffer"`
-	ITimeout     int32             `json:"iTimeout"`
-	Context      map[string]string `json:"context"`
-	Status       map[string]string `json:"status"`
+	IVersion     int16             `json:"iVersion" tars:"iVersion,tag:1,require:true"`
+	CPacketType  int8              `json:"cPacketType" tars:"cPacketType,tag:2,require:true"`
+	IMessageType int32             `json:"iMessageType" tars:"iMessageType,tag:3,require:true"`
+	IRequestId   int32             `json:"iRequestId" tars:"iRequestId,tag:4,require:true"`
+	SServantName string            `json:"sServantName" tars:"sServantName,tag:5,require:true"`
+	SFuncName    string            `json:"sFuncName" tars:"sFuncName,tag:6,require:true"`
+	SBuffer      []int8            `json:"sBuffer" tars:"sBuffer,tag:7,require:true"`
+	ITimeout     int32             `json:"iTimeout" tars:"iTimeout,tag:8,require:true"`
+	Context      map[string]string `json:"context" tars:"context,tag:9,require:true"`
+	Status       map[string]string `json:"status" tars:"status,tag:10,require:true"`
 }
 
 func (st *RequestPacket) ResetDefault() {
@@ -79,45 +79,36 @@ func (st *RequestPacket) ReadFrom(readBuf *codec.Reader) error {
 	if err != nil {
 		return err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return err
 		}
-
 		st.SBuffer = make([]int8, length)
 		for i0, e0 := int32(0), length; i0 < e0; i0++ {
-
-			err = readBuf.ReadInt8(&st.SBuffer[i0], 0, false)
+			err = readBuf.ReadInt8(&st.SBuffer[i0], 0, true)
 			if err != nil {
 				return err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
-
 		_, err = readBuf.SkipTo(codec.BYTE, 0, true)
 		if err != nil {
 			return err
 		}
-
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return err
 		}
-
 		err = readBuf.ReadSliceInt8(&st.SBuffer, length, true)
 		if err != nil {
 			return err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return err
 		}
-
 	}
 
 	err = readBuf.ReadInt32(&st.ITimeout, 8, true)
@@ -129,27 +120,22 @@ func (st *RequestPacket) ReadFrom(readBuf *codec.Reader) error {
 	if err != nil {
 		return err
 	}
-
 	err = readBuf.ReadInt32(&length, 0, true)
 	if err != nil {
 		return err
 	}
-
 	st.Context = make(map[string]string)
 	for i1, e1 := int32(0), length; i1 < e1; i1++ {
 		var k1 string
 		var v1 string
-
-		err = readBuf.ReadString(&k1, 0, false)
+		err = readBuf.ReadString(&k1, 0, true)
 		if err != nil {
 			return err
 		}
-
-		err = readBuf.ReadString(&v1, 1, false)
+		err = readBuf.ReadString(&v1, 1, true)
 		if err != nil {
 			return err
 		}
-
 		st.Context[k1] = v1
 	}
 
@@ -157,27 +143,22 @@ func (st *RequestPacket) ReadFrom(readBuf *codec.Reader) error {
 	if err != nil {
 		return err
 	}
-
 	err = readBuf.ReadInt32(&length, 0, true)
 	if err != nil {
 		return err
 	}
-
 	st.Status = make(map[string]string)
 	for i2, e2 := int32(0), length; i2 < e2; i2++ {
 		var k2 string
 		var v2 string
-
-		err = readBuf.ReadString(&k2, 0, false)
+		err = readBuf.ReadString(&k2, 0, true)
 		if err != nil {
 			return err
 		}
-
-		err = readBuf.ReadString(&v2, 1, false)
+		err = readBuf.ReadString(&v2, 1, true)
 		if err != nil {
 			return err
 		}
-
 		st.Status[k2] = v2
 	}
 
@@ -222,7 +203,6 @@ func (st *RequestPacket) ReadBlock(readBuf *codec.Reader, tag byte, require bool
 
 // WriteTo encode struct to buffer
 func (st *RequestPacket) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteInt16(st.IVersion, 1)
 	if err != nil {
 		return err
@@ -257,17 +237,14 @@ func (st *RequestPacket) WriteTo(buf *codec.Buffer) (err error) {
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteHead(codec.BYTE, 0)
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(st.SBuffer)), 0)
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteSliceInt8(st.SBuffer)
 	if err != nil {
 		return err
@@ -282,48 +259,38 @@ func (st *RequestPacket) WriteTo(buf *codec.Buffer) (err error) {
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(st.Context)), 0)
 	if err != nil {
 		return err
 	}
-
 	for k3, v3 := range st.Context {
-
 		err = buf.WriteString(k3, 0)
 		if err != nil {
 			return err
 		}
-
 		err = buf.WriteString(v3, 1)
 		if err != nil {
 			return err
 		}
-
 	}
 
 	err = buf.WriteHead(codec.MAP, 10)
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(st.Status)), 0)
 	if err != nil {
 		return err
 	}
-
 	for k4, v4 := range st.Status {
-
 		err = buf.WriteString(k4, 0)
 		if err != nil {
 			return err
 		}
-
 		err = buf.WriteString(v4, 1)
 		if err != nil {
 			return err
 		}
-
 	}
 
 	return err
@@ -351,15 +318,15 @@ func (st *RequestPacket) WriteBlock(buf *codec.Buffer, tag byte) error {
 
 // ResponsePacket struct implement
 type ResponsePacket struct {
-	IVersion     int16             `json:"iVersion"`
-	CPacketType  int8              `json:"cPacketType"`
-	IRequestId   int32             `json:"iRequestId"`
-	IMessageType int32             `json:"iMessageType"`
-	IRet         int32             `json:"iRet"`
-	SBuffer      []int8            `json:"sBuffer"`
-	Status       map[string]string `json:"status"`
-	SResultDesc  string            `json:"sResultDesc"`
-	Context      map[string]string `json:"context"`
+	IVersion     int16             `json:"iVersion" tars:"iVersion,tag:1,require:true"`
+	CPacketType  int8              `json:"cPacketType" tars:"cPacketType,tag:2,require:true"`
+	IRequestId   int32             `json:"iRequestId" tars:"iRequestId,tag:3,require:true"`
+	IMessageType int32             `json:"iMessageType" tars:"iMessageType,tag:4,require:true"`
+	IRet         int32             `json:"iRet" tars:"iRet,tag:5,require:true"`
+	SBuffer      []int8            `json:"sBuffer" tars:"sBuffer,tag:6,require:true"`
+	Status       map[string]string `json:"status" tars:"status,tag:7,require:true"`
+	SResultDesc  string            `json:"sResultDesc" tars:"sResultDesc,tag:8,require:false"`
+	Context      map[string]string `json:"context" tars:"context,tag:9,require:false"`
 }
 
 func (st *ResponsePacket) ResetDefault() {
@@ -407,72 +374,58 @@ func (st *ResponsePacket) ReadFrom(readBuf *codec.Reader) error {
 	if err != nil {
 		return err
 	}
-
 	if ty == codec.LIST {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return err
 		}
-
 		st.SBuffer = make([]int8, length)
 		for i0, e0 := int32(0), length; i0 < e0; i0++ {
-
-			err = readBuf.ReadInt8(&st.SBuffer[i0], 0, false)
+			err = readBuf.ReadInt8(&st.SBuffer[i0], 0, true)
 			if err != nil {
 				return err
 			}
-
 		}
 	} else if ty == codec.SimpleList {
-
 		_, err = readBuf.SkipTo(codec.BYTE, 0, true)
 		if err != nil {
 			return err
 		}
-
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return err
 		}
-
 		err = readBuf.ReadSliceInt8(&st.SBuffer, length, true)
 		if err != nil {
 			return err
 		}
-
 	} else {
 		err = fmt.Errorf("require vector, but not")
 		if err != nil {
 			return err
 		}
-
 	}
 
 	_, err = readBuf.SkipTo(codec.MAP, 7, true)
 	if err != nil {
 		return err
 	}
-
 	err = readBuf.ReadInt32(&length, 0, true)
 	if err != nil {
 		return err
 	}
-
 	st.Status = make(map[string]string)
 	for i1, e1 := int32(0), length; i1 < e1; i1++ {
 		var k1 string
 		var v1 string
-
-		err = readBuf.ReadString(&k1, 0, false)
+		err = readBuf.ReadString(&k1, 0, true)
 		if err != nil {
 			return err
 		}
-
-		err = readBuf.ReadString(&v1, 1, false)
+		err = readBuf.ReadString(&v1, 1, true)
 		if err != nil {
 			return err
 		}
-
 		st.Status[k1] = v1
 	}
 
@@ -485,28 +438,23 @@ func (st *ResponsePacket) ReadFrom(readBuf *codec.Reader) error {
 	if err != nil {
 		return err
 	}
-
 	if have {
 		err = readBuf.ReadInt32(&length, 0, true)
 		if err != nil {
 			return err
 		}
-
 		st.Context = make(map[string]string)
 		for i2, e2 := int32(0), length; i2 < e2; i2++ {
 			var k2 string
 			var v2 string
-
-			err = readBuf.ReadString(&k2, 0, false)
+			err = readBuf.ReadString(&k2, 0, true)
 			if err != nil {
 				return err
 			}
-
-			err = readBuf.ReadString(&v2, 1, false)
+			err = readBuf.ReadString(&v2, 1, true)
 			if err != nil {
 				return err
 			}
-
 			st.Context[k2] = v2
 		}
 	}
@@ -552,7 +500,6 @@ func (st *ResponsePacket) ReadBlock(readBuf *codec.Reader, tag byte, require boo
 
 // WriteTo encode struct to buffer
 func (st *ResponsePacket) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteInt16(st.IVersion, 1)
 	if err != nil {
 		return err
@@ -582,17 +529,14 @@ func (st *ResponsePacket) WriteTo(buf *codec.Buffer) (err error) {
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteHead(codec.BYTE, 0)
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(st.SBuffer)), 0)
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteSliceInt8(st.SBuffer)
 	if err != nil {
 		return err
@@ -602,53 +546,47 @@ func (st *ResponsePacket) WriteTo(buf *codec.Buffer) (err error) {
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(st.Status)), 0)
 	if err != nil {
 		return err
 	}
-
 	for k3, v3 := range st.Status {
-
 		err = buf.WriteString(k3, 0)
 		if err != nil {
 			return err
 		}
-
 		err = buf.WriteString(v3, 1)
 		if err != nil {
 			return err
 		}
-
 	}
 
-	err = buf.WriteString(st.SResultDesc, 8)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteHead(codec.MAP, 9)
-	if err != nil {
-		return err
-	}
-
-	err = buf.WriteInt32(int32(len(st.Context)), 0)
-	if err != nil {
-		return err
-	}
-
-	for k4, v4 := range st.Context {
-
-		err = buf.WriteString(k4, 0)
+	if st.SResultDesc != "" {
+		err = buf.WriteString(st.SResultDesc, 8)
 		if err != nil {
 			return err
 		}
+	}
 
-		err = buf.WriteString(v4, 1)
+	if len(st.Context) > 0 {
+		err = buf.WriteHead(codec.MAP, 9)
 		if err != nil {
 			return err
 		}
-
+		err = buf.WriteInt32(int32(len(st.Context)), 0)
+		if err != nil {
+			return err
+		}
+		for k4, v4 := range st.Context {
+			err = buf.WriteString(k4, 0)
+			if err != nil {
+				return err
+			}
+			err = buf.WriteString(v4, 1)
+			if err != nil {
+				return err
+			}
+		}
 	}
 
 	return err
