@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from StatF.tars
 // Package statf comment
-// This file was generated by tars2go 1.1.10
-// Generated from StatF.tars
 package statf
 
 import (
@@ -15,17 +15,17 @@ var _ = codec.FromInt8
 
 // StatMicMsgHead struct implement
 type StatMicMsgHead struct {
-	MasterName    string `json:"masterName"`
-	SlaveName     string `json:"slaveName"`
-	InterfaceName string `json:"interfaceName"`
-	MasterIp      string `json:"masterIp"`
-	SlaveIp       string `json:"slaveIp"`
-	SlavePort     int32  `json:"slavePort"`
-	ReturnValue   int32  `json:"returnValue"`
-	SlaveSetName  string `json:"slaveSetName"`
-	SlaveSetArea  string `json:"slaveSetArea"`
-	SlaveSetID    string `json:"slaveSetID"`
-	TarsVersion   string `json:"tarsVersion"`
+	MasterName    string `json:"masterName" tars:"masterName,tag:0,require:true"`
+	SlaveName     string `json:"slaveName" tars:"slaveName,tag:1,require:true"`
+	InterfaceName string `json:"interfaceName" tars:"interfaceName,tag:2,require:true"`
+	MasterIp      string `json:"masterIp" tars:"masterIp,tag:3,require:true"`
+	SlaveIp       string `json:"slaveIp" tars:"slaveIp,tag:4,require:true"`
+	SlavePort     int32  `json:"slavePort" tars:"slavePort,tag:5,require:true"`
+	ReturnValue   int32  `json:"returnValue" tars:"returnValue,tag:6,require:true"`
+	SlaveSetName  string `json:"slaveSetName" tars:"slaveSetName,tag:7,require:false"`
+	SlaveSetArea  string `json:"slaveSetArea" tars:"slaveSetArea,tag:8,require:false"`
+	SlaveSetID    string `json:"slaveSetID" tars:"slaveSetID,tag:9,require:false"`
+	TarsVersion   string `json:"tarsVersion" tars:"tarsVersion,tag:10,require:false"`
 }
 
 func (st *StatMicMsgHead) ResetDefault() {
@@ -137,7 +137,6 @@ func (st *StatMicMsgHead) ReadBlock(readBuf *codec.Reader, tag byte, require boo
 
 // WriteTo encode struct to buffer
 func (st *StatMicMsgHead) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.MasterName, 0)
 	if err != nil {
 		return err
@@ -173,24 +172,32 @@ func (st *StatMicMsgHead) WriteTo(buf *codec.Buffer) (err error) {
 		return err
 	}
 
-	err = buf.WriteString(st.SlaveSetName, 7)
-	if err != nil {
-		return err
+	if st.SlaveSetName != "" {
+		err = buf.WriteString(st.SlaveSetName, 7)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SlaveSetArea, 8)
-	if err != nil {
-		return err
+	if st.SlaveSetArea != "" {
+		err = buf.WriteString(st.SlaveSetArea, 8)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.SlaveSetID, 9)
-	if err != nil {
-		return err
+	if st.SlaveSetID != "" {
+		err = buf.WriteString(st.SlaveSetID, 9)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = buf.WriteString(st.TarsVersion, 10)
-	if err != nil {
-		return err
+	if st.TarsVersion != "" {
+		err = buf.WriteString(st.TarsVersion, 10)
+		if err != nil {
+			return err
+		}
 	}
 
 	return err
@@ -218,13 +225,13 @@ func (st *StatMicMsgHead) WriteBlock(buf *codec.Buffer, tag byte) error {
 
 // StatMicMsgBody struct implement
 type StatMicMsgBody struct {
-	Count         int32           `json:"count"`
-	TimeoutCount  int32           `json:"timeoutCount"`
-	ExecCount     int32           `json:"execCount"`
-	IntervalCount map[int32]int32 `json:"intervalCount"`
-	TotalRspTime  int64           `json:"totalRspTime"`
-	MaxRspTime    int32           `json:"maxRspTime"`
-	MinRspTime    int32           `json:"minRspTime"`
+	Count         int32           `json:"count" tars:"count,tag:0,require:true"`
+	TimeoutCount  int32           `json:"timeoutCount" tars:"timeoutCount,tag:1,require:true"`
+	ExecCount     int32           `json:"execCount" tars:"execCount,tag:2,require:true"`
+	IntervalCount map[int32]int32 `json:"intervalCount" tars:"intervalCount,tag:3,require:true"`
+	TotalRspTime  int64           `json:"totalRspTime" tars:"totalRspTime,tag:4,require:true"`
+	MaxRspTime    int32           `json:"maxRspTime" tars:"maxRspTime,tag:5,require:true"`
+	MinRspTime    int32           `json:"minRspTime" tars:"minRspTime,tag:6,require:true"`
 }
 
 func (st *StatMicMsgBody) ResetDefault() {
@@ -259,27 +266,22 @@ func (st *StatMicMsgBody) ReadFrom(readBuf *codec.Reader) error {
 	if err != nil {
 		return err
 	}
-
 	err = readBuf.ReadInt32(&length, 0, true)
 	if err != nil {
 		return err
 	}
-
 	st.IntervalCount = make(map[int32]int32)
 	for i0, e0 := int32(0), length; i0 < e0; i0++ {
 		var k0 int32
 		var v0 int32
-
-		err = readBuf.ReadInt32(&k0, 0, false)
+		err = readBuf.ReadInt32(&k0, 0, true)
 		if err != nil {
 			return err
 		}
-
-		err = readBuf.ReadInt32(&v0, 1, false)
+		err = readBuf.ReadInt32(&v0, 1, true)
 		if err != nil {
 			return err
 		}
-
 		st.IntervalCount[k0] = v0
 	}
 
@@ -339,7 +341,6 @@ func (st *StatMicMsgBody) ReadBlock(readBuf *codec.Reader, tag byte, require boo
 
 // WriteTo encode struct to buffer
 func (st *StatMicMsgBody) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteInt32(st.Count, 0)
 	if err != nil {
 		return err
@@ -359,24 +360,19 @@ func (st *StatMicMsgBody) WriteTo(buf *codec.Buffer) (err error) {
 	if err != nil {
 		return err
 	}
-
 	err = buf.WriteInt32(int32(len(st.IntervalCount)), 0)
 	if err != nil {
 		return err
 	}
-
 	for k1, v1 := range st.IntervalCount {
-
 		err = buf.WriteInt32(k1, 0)
 		if err != nil {
 			return err
 		}
-
 		err = buf.WriteInt32(v1, 1)
 		if err != nil {
 			return err
 		}
-
 	}
 
 	err = buf.WriteInt64(st.TotalRspTime, 4)
@@ -419,15 +415,15 @@ func (st *StatMicMsgBody) WriteBlock(buf *codec.Buffer, tag byte) error {
 
 // StatSampleMsg struct implement
 type StatSampleMsg struct {
-	Unid          string `json:"unid"`
-	MasterName    string `json:"masterName"`
-	SlaveName     string `json:"slaveName"`
-	InterfaceName string `json:"interfaceName"`
-	MasterIp      string `json:"masterIp"`
-	SlaveIp       string `json:"slaveIp"`
-	Depth         int32  `json:"depth"`
-	Width         int32  `json:"width"`
-	ParentWidth   int32  `json:"parentWidth"`
+	Unid          string `json:"unid" tars:"unid,tag:0,require:true"`
+	MasterName    string `json:"masterName" tars:"masterName,tag:1,require:true"`
+	SlaveName     string `json:"slaveName" tars:"slaveName,tag:2,require:true"`
+	InterfaceName string `json:"interfaceName" tars:"interfaceName,tag:3,require:true"`
+	MasterIp      string `json:"masterIp" tars:"masterIp,tag:4,require:true"`
+	SlaveIp       string `json:"slaveIp" tars:"slaveIp,tag:5,require:true"`
+	Depth         int32  `json:"depth" tars:"depth,tag:6,require:true"`
+	Width         int32  `json:"width" tars:"width,tag:7,require:true"`
+	ParentWidth   int32  `json:"parentWidth" tars:"parentWidth,tag:8,require:true"`
 }
 
 func (st *StatSampleMsg) ResetDefault() {
@@ -529,7 +525,6 @@ func (st *StatSampleMsg) ReadBlock(readBuf *codec.Reader, tag byte, require bool
 
 // WriteTo encode struct to buffer
 func (st *StatSampleMsg) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteString(st.Unid, 0)
 	if err != nil {
 		return err
@@ -600,7 +595,7 @@ func (st *StatSampleMsg) WriteBlock(buf *codec.Buffer, tag byte) error {
 
 // ProxyInfo struct implement
 type ProxyInfo struct {
-	BFromClient bool `json:"bFromClient"`
+	BFromClient bool `json:"bFromClient" tars:"bFromClient,tag:0,require:true"`
 }
 
 func (st *ProxyInfo) ResetDefault() {
@@ -662,7 +657,6 @@ func (st *ProxyInfo) ReadBlock(readBuf *codec.Reader, tag byte, require bool) er
 
 // WriteTo encode struct to buffer
 func (st *ProxyInfo) WriteTo(buf *codec.Buffer) (err error) {
-
 	err = buf.WriteBool(st.BFromClient, 0)
 	if err != nil {
 		return err
