@@ -1,6 +1,6 @@
+// Code generated by tars2go 1.2.3, DO NOT EDIT.
+// This file was generated from StatF.tars
 // Package statf comment
-// This file was generated by tars2go 1.1.10
-// Generated from StatF.tars
 package statf
 
 import (
@@ -8,113 +8,67 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	m "github.com/TarsCloud/TarsGo/tars/model"
+
+	"github.com/TarsCloud/TarsGo/tars/model"
 	"github.com/TarsCloud/TarsGo/tars/protocol/codec"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/requestf"
 	"github.com/TarsCloud/TarsGo/tars/protocol/tup"
 	"github.com/TarsCloud/TarsGo/tars/util/current"
+	"github.com/TarsCloud/TarsGo/tars/util/endpoint"
 	"github.com/TarsCloud/TarsGo/tars/util/tools"
-	"unsafe"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var (
 	_ = fmt.Errorf
 	_ = codec.FromInt8
-	_ = unsafe.Pointer(nil)
 	_ = bytes.ErrTooLarge
 )
 
+type StatFServant interface {
+	ReportMicMsg(msg map[StatMicMsgHead]StatMicMsgBody, bFromClient bool) (ret int32, err error)
+	ReportSampleMsg(msg []StatSampleMsg) (ret int32, err error)
+}
+
+type StatFServantWithContext interface {
+	ReportMicMsg(tarsCtx context.Context, msg map[StatMicMsgHead]StatMicMsgBody, bFromClient bool) (ret int32, err error)
+	ReportSampleMsg(tarsCtx context.Context, msg []StatSampleMsg) (ret int32, err error)
+}
+
 // StatF struct
 type StatF struct {
-	servant m.Servant
+	servant model.Servant
+}
+
+// NewStatF creates a new StatF servant.
+func NewStatF() *StatF {
+	return new(StatF)
+}
+
+// SetServant sets servant for the service.
+func (obj *StatF) SetServant(servant model.Servant) {
+	obj.servant = servant
+}
+
+// TarsSetTimeout sets the timeout for the servant which is in ms.
+func (obj *StatF) TarsSetTimeout(timeout int) {
+	obj.servant.TarsSetTimeout(timeout)
+}
+
+// TarsSetProtocol sets the protocol for the servant.
+func (obj *StatF) TarsSetProtocol(p model.Protocol) {
+	obj.servant.TarsSetProtocol(p)
+}
+
+// Endpoints returns all active endpoint.Endpoint
+func (obj *StatF) Endpoints() []*endpoint.Endpoint {
+	return obj.servant.Endpoints()
 }
 
 // ReportMicMsg is the proxy function for the method defined in the tars file, with the context
-func (obj *StatF) ReportMicMsg(msg map[StatMicMsgHead]StatMicMsgBody, bFromClient bool, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteHead(codec.MAP, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(msg)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for k0, v0 := range msg {
-
-		err = k0.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-		err = v0.WriteBlock(buf, 1)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	err = buf.WriteBool(bFromClient, 2)
-	if err != nil {
-		return ret, err
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "reportMicMsg", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *StatF) ReportMicMsg(msg map[StatMicMsgHead]StatMicMsgBody, bFromClient bool, opts ...map[string]string) (int32, error) {
+	return obj.ReportMicMsgWithContext(context.Background(), msg, bFromClient, opts...)
 }
 
 // ReportMicMsgWithContext is the proxy function for the method defined in the tars file, with the context
@@ -129,24 +83,19 @@ func (obj *StatF) ReportMicMsgWithContext(tarsCtx context.Context, msg map[StatM
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(msg)), 0)
 	if err != nil {
 		return ret, err
 	}
-
-	for k1, v1 := range msg {
-
-		err = k1.WriteBlock(buf, 0)
+	for k0, v0 := range msg {
+		err = k0.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
-		err = v1.WriteBlock(buf, 1)
+		err = v0.WriteBlock(buf, 1)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteBool(bFromClient, 2)
@@ -168,7 +117,6 @@ func (obj *StatF) ReportMicMsgWithContext(tarsCtx context.Context, msg map[StatM
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -196,6 +144,7 @@ func (obj *StatF) ReportMicMsgWithContext(tarsCtx context.Context, msg map[StatM
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -214,24 +163,19 @@ func (obj *StatF) ReportMicMsgOneWayWithContext(tarsCtx context.Context, msg map
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(msg)), 0)
 	if err != nil {
 		return ret, err
 	}
-
-	for k2, v2 := range msg {
-
-		err = k2.WriteBlock(buf, 0)
+	for k1, v1 := range msg {
+		err = k1.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
-		err = v2.WriteBlock(buf, 1)
+		err = v1.WriteBlock(buf, 1)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	err = buf.WriteBool(bFromClient, 2)
@@ -254,27 +198,6 @@ func (obj *StatF) ReportMicMsgOneWayWithContext(tarsCtx context.Context, msg map
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
@@ -282,79 +205,8 @@ func (obj *StatF) ReportMicMsgOneWayWithContext(tarsCtx context.Context, msg map
 }
 
 // ReportSampleMsg is the proxy function for the method defined in the tars file, with the context
-func (obj *StatF) ReportSampleMsg(msg []StatSampleMsg, opts ...map[string]string) (ret int32, err error) {
-	var (
-		length int32
-		have   bool
-		ty     byte
-	)
-	buf := codec.NewBuffer()
-	err = buf.WriteHead(codec.LIST, 1)
-	if err != nil {
-		return ret, err
-	}
-
-	err = buf.WriteInt32(int32(len(msg)), 0)
-	if err != nil {
-		return ret, err
-	}
-
-	for _, v := range msg {
-
-		err = v.WriteBlock(buf, 0)
-		if err != nil {
-			return ret, err
-		}
-
-	}
-
-	var statusMap map[string]string
-	var contextMap map[string]string
-	if len(opts) == 1 {
-		contextMap = opts[0]
-	} else if len(opts) == 2 {
-		contextMap = opts[0]
-		statusMap = opts[1]
-	}
-	tarsResp := new(requestf.ResponsePacket)
-	tarsCtx := context.Background()
-
-	err = obj.servant.TarsInvoke(tarsCtx, 0, "reportSampleMsg", buf.ToBytes(), statusMap, contextMap, tarsResp)
-	if err != nil {
-		return ret, err
-	}
-
-	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
-	err = readBuf.ReadInt32(&ret, 0, true)
-	if err != nil {
-		return ret, err
-	}
-
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
-	_ = length
-	_ = have
-	_ = ty
-	return ret, nil
+func (obj *StatF) ReportSampleMsg(msg []StatSampleMsg, opts ...map[string]string) (int32, error) {
+	return obj.ReportSampleMsgWithContext(context.Background(), msg, opts...)
 }
 
 // ReportSampleMsgWithContext is the proxy function for the method defined in the tars file, with the context
@@ -369,19 +221,15 @@ func (obj *StatF) ReportSampleMsgWithContext(tarsCtx context.Context, msg []Stat
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(msg)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range msg {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -398,7 +246,6 @@ func (obj *StatF) ReportSampleMsgWithContext(tarsCtx context.Context, msg []Stat
 	if err != nil {
 		return ret, err
 	}
-
 	readBuf := codec.NewReader(tools.Int8ToByte(tarsResp.SBuffer))
 	err = readBuf.ReadInt32(&ret, 0, true)
 	if err != nil {
@@ -426,6 +273,7 @@ func (obj *StatF) ReportSampleMsgWithContext(tarsCtx context.Context, msg []Stat
 			statusMap[k] = v
 		}
 	}
+
 	_ = length
 	_ = have
 	_ = ty
@@ -444,19 +292,15 @@ func (obj *StatF) ReportSampleMsgOneWayWithContext(tarsCtx context.Context, msg 
 	if err != nil {
 		return ret, err
 	}
-
 	err = buf.WriteInt32(int32(len(msg)), 0)
 	if err != nil {
 		return ret, err
 	}
-
 	for _, v := range msg {
-
 		err = v.WriteBlock(buf, 0)
 		if err != nil {
 			return ret, err
 		}
-
 	}
 
 	var statusMap map[string]string
@@ -474,55 +318,10 @@ func (obj *StatF) ReportSampleMsgOneWayWithContext(tarsCtx context.Context, msg 
 		return ret, err
 	}
 
-	if len(opts) == 1 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-	} else if len(opts) == 2 {
-		for k := range contextMap {
-			delete(contextMap, k)
-		}
-		for k, v := range tarsResp.Context {
-			contextMap[k] = v
-		}
-		for k := range statusMap {
-			delete(statusMap, k)
-		}
-		for k, v := range tarsResp.Status {
-			statusMap[k] = v
-		}
-	}
 	_ = length
 	_ = have
 	_ = ty
 	return ret, nil
-}
-
-// SetServant sets servant for the service.
-func (obj *StatF) SetServant(servant m.Servant) {
-	obj.servant = servant
-}
-
-// TarsSetTimeout sets the timeout for the servant which is in ms.
-func (obj *StatF) TarsSetTimeout(timeout int) {
-	obj.servant.TarsSetTimeout(timeout)
-}
-
-// TarsSetProtocol sets the protocol for the servant.
-func (obj *StatF) TarsSetProtocol(p m.Protocol) {
-	obj.servant.TarsSetProtocol(p)
-}
-
-type StatFServant interface {
-	ReportMicMsg(msg map[StatMicMsgHead]StatMicMsgBody, bFromClient bool) (ret int32, err error)
-	ReportSampleMsg(msg []StatSampleMsg) (ret int32, err error)
-}
-type StatFServantWithContext interface {
-	ReportMicMsg(tarsCtx context.Context, msg map[StatMicMsgHead]StatMicMsgBody, bFromClient bool) (ret int32, err error)
-	ReportSampleMsg(tarsCtx context.Context, msg []StatSampleMsg) (ret int32, err error)
 }
 
 // Dispatch is used to call the server side implement for the method defined in the tars file. withContext shows using context or not.
@@ -539,42 +338,33 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 		var msg map[StatMicMsgHead]StatMicMsgBody
 		msg = make(map[StatMicMsgHead]StatMicMsgBody)
 		var bFromClient bool
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			_, err = readBuf.SkipTo(codec.MAP, 1, true)
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadInt32(&length, 0, true)
 			if err != nil {
 				return err
 			}
-
 			msg = make(map[StatMicMsgHead]StatMicMsgBody)
-			for i3, e3 := int32(0), length; i3 < e3; i3++ {
-				var k3 StatMicMsgHead
-				var v3 StatMicMsgBody
-
-				err = k3.ReadBlock(readBuf, 0, false)
+			for i2, e2 := int32(0), length; i2 < e2; i2++ {
+				var k2 StatMicMsgHead
+				var v2 StatMicMsgBody
+				err = k2.ReadBlock(readBuf, 0, true)
 				if err != nil {
 					return err
 				}
-
-				err = v3.ReadBlock(readBuf, 1, false)
+				err = v2.ReadBlock(readBuf, 1, true)
 				if err != nil {
 					return err
 				}
-
-				msg[k3] = v3
+				msg[k2] = v2
 			}
-
 			err = readBuf.ReadBool(&bFromClient, 2, true)
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
 			reqTup.Decode(readBuf)
@@ -587,28 +377,23 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 			if err != nil {
 				return err
 			}
-
 			err = readBuf.ReadInt32(&length, 0, true)
 			if err != nil {
 				return err
 			}
-
 			msg = make(map[StatMicMsgHead]StatMicMsgBody)
-			for i4, e4 := int32(0), length; i4 < e4; i4++ {
-				var k4 StatMicMsgHead
-				var v4 StatMicMsgBody
-
-				err = k4.ReadBlock(readBuf, 0, false)
+			for i3, e3 := int32(0), length; i3 < e3; i3++ {
+				var k3 StatMicMsgHead
+				var v3 StatMicMsgBody
+				err = k3.ReadBlock(readBuf, 0, true)
 				if err != nil {
 					return err
 				}
-
-				err = v4.ReadBlock(readBuf, 1, false)
+				err = v3.ReadBlock(readBuf, 1, true)
 				if err != nil {
 					return err
 				}
-
-				msg[k4] = v4
+				msg[k3] = v3
 			}
 
 			reqTup.GetBuffer("bFromClient", &tupBuffer)
@@ -617,7 +402,6 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
 			decoder := json.NewDecoder(bytes.NewReader(readBuf.ToBytes()))
@@ -638,7 +422,6 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -652,7 +435,6 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 			imp := val.(StatFServantWithContext)
 			funRet, err = imp.ReportMicMsg(tarsCtx, msg, bFromClient)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -664,7 +446,6 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
@@ -699,41 +480,33 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 	case "reportSampleMsg":
 		var msg []StatSampleMsg
 		msg = make([]StatSampleMsg, 0)
-
 		if tarsReq.IVersion == basef.TARSVERSION {
-
 			_, ty, err = readBuf.SkipToNoCheck(1, true)
 			if err != nil {
 				return err
 			}
-
 			if ty == codec.LIST {
 				err = readBuf.ReadInt32(&length, 0, true)
 				if err != nil {
 					return err
 				}
-
 				msg = make([]StatSampleMsg, length)
-				for i5, e5 := int32(0), length; i5 < e5; i5++ {
-
-					err = msg[i5].ReadBlock(readBuf, 0, false)
+				for i4, e4 := int32(0), length; i4 < e4; i4++ {
+					err = msg[i4].ReadBlock(readBuf, 0, true)
 					if err != nil {
 						return err
 					}
-
 				}
 			} else if ty == codec.SimpleList {
 				err = fmt.Errorf("not support SimpleList type")
 				if err != nil {
 					return err
 				}
-
 			} else {
 				err = fmt.Errorf("require vector, but not")
 				if err != nil {
 					return err
 				}
-
 			}
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			reqTup := tup.NewUniAttribute()
@@ -747,34 +520,28 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 			if err != nil {
 				return err
 			}
-
 			if ty == codec.LIST {
 				err = readBuf.ReadInt32(&length, 0, true)
 				if err != nil {
 					return err
 				}
-
 				msg = make([]StatSampleMsg, length)
-				for i6, e6 := int32(0), length; i6 < e6; i6++ {
-
-					err = msg[i6].ReadBlock(readBuf, 0, false)
+				for i5, e5 := int32(0), length; i5 < e5; i5++ {
+					err = msg[i5].ReadBlock(readBuf, 0, true)
 					if err != nil {
 						return err
 					}
-
 				}
 			} else if ty == codec.SimpleList {
 				err = fmt.Errorf("not support SimpleList type")
 				if err != nil {
 					return err
 				}
-
 			} else {
 				err = fmt.Errorf("require vector, but not")
 				if err != nil {
 					return err
 				}
-
 			}
 		} else if tarsReq.IVersion == basef.JSONVERSION {
 			var jsonData map[string]interface{}
@@ -790,7 +557,6 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 					return err
 				}
 			}
-
 		} else {
 			err = fmt.Errorf("decode reqpacket fail, error version: %d", tarsReq.IVersion)
 			return err
@@ -804,7 +570,6 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 			imp := val.(StatFServantWithContext)
 			funRet, err = imp.ReportSampleMsg(tarsCtx, msg)
 		}
-
 		if err != nil {
 			return err
 		}
@@ -816,7 +581,6 @@ func (obj *StatF) Dispatch(tarsCtx context.Context, val interface{}, tarsReq *re
 			if err != nil {
 				return err
 			}
-
 		} else if tarsReq.IVersion == basef.TUPVERSION {
 			rspTup := tup.NewUniAttribute()
 
