@@ -3,11 +3,12 @@ package protocol
 import (
 	"bytes"
 	"encoding/binary"
+	"testing"
+
 	"github.com/TarsCloud/TarsGo/tars/protocol/codec"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/requestf"
 	"github.com/stretchr/testify/assert"
-	"testing"
 )
 
 func TestSetMaxPackageLength(t *testing.T) {
@@ -92,6 +93,7 @@ func TestTarsProtocol_ParsePackage(t *testing.T) {
 		})
 	}
 }
+
 func TestTarsProtocol_RequestPack(t *testing.T) {
 	req := &requestf.RequestPacket{
 		IVersion:     basef.TARSVERSION,
@@ -150,5 +152,4 @@ func TestTarsProtocol_ResponseUnpack(t *testing.T) {
 	}
 
 	assert.Equal(t, got, resp, "Failed to test ResponseUnpack")
-
 }
