@@ -8,8 +8,8 @@
 // hack/cmake/CMakeTestGoCompiler.cmake
 // hack/cmake/golang.cmake
 // hack/cmake/tars-tools.cmake
-// hack/scripts/makefile.tars
-// hack/scripts/makefile.tars.gomod
+// hack/scripts/makefile.tars.gomod.mk
+// hack/scripts/makefile.tars.mk
 package bindata
 
 import (
@@ -186,7 +186,7 @@ func cmakeGolangCmake() (*asset, error) {
 	return a, nil
 }
 
-var _cmakeTarsToolsCmake = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xec\x59\x4f\x8f\xe3\x4a\x11\x3f\xc7\x9f\xa2\xf1\xcc\xc1\x79\x22\x89\xb4\xdc\x46\x1a\x81\x27\xe9\xc9\x98\xf5\xd8\xc1\x71\x76\x76\xa5\x48\x96\xc7\xee\x24\xd6\x3a\xb6\x65\x77\xe6\xed\xbe\x51\x6e\x48\x80\x40\x7a\x20\x21\x0e\x5c\x10\x12\x02\x6e\xdc\x10\xe2\x89\x6f\xf3\x76\x97\x8f\x81\xaa\xdb\x7f\x63\x27\x93\xc9\xcb\xec\x2e\x82\x5c\x66\x6c\x57\x55\xff\xea\x4f\x57\x55\x57\x0b\x02\x79\x43\x9c\x15\x25\x56\x14\x87\x0e\x49\x12\xa9\xaf\x5f\x5f\xcb\xda\x00\xcd\x43\x44\x82\x3b\x34\xd4\x47\xb2\x79\x85\xf4\x89\x39\x9a\x98\xd6\x0b\xd9\x50\xe4\x0b\x15\xa7\xaf\xdb\x42\x42\x63\x2f\x98\x4b\x06\x1e\xe2\x97\xc8\xc0\x23\x55\xee\x63\x24\x4e\x83\x53\x11\x89\x62\xc6\x2c\x9e\xde\xf3\xff\xd6\x62\x5b\x48\x08\x95\xfa\xd7\xf2\x73\x6c\x5d\xeb\x83\x89\x8a\x2d\x46\x72\x7a\x5f\x7b\xb7\x06\x3e\xfe\xb6\x3f\x31\x0c\xac\x99\xd6\x58\x9f\x18\x7d\x6c\x0d\x14\x63\xdd\x73\x96\xf6\x6b\x22\xb6\x05\xc1\x0b\x1c\x7f\xe5\x12\xe9\x41\xe2\xde\x3c\xf4\xed\x60\xde\x65\x0f\x6d\xa1\x84\x44\x36\xfa\x57\xca\x0b\x6c\xa5\x5a\x0e\x14\x03\xf7\x4d\xdd\x78\x95\xc3\xba\x50\x34\xd9\x78\xc5\x65\xf9\xde\x6d\x59\x0b\x55\xb9\x30\xe0\xdb\x21\xbc\xc6\x44\x33\x95\xeb\x3d\xd7\xbd\xf5\x02\xd0\x76\x26\xdd\x28\xda\x0f\x9e\xb5\x85\x16\xc8\x31\x65\x63\xfc\xac\x3f\x1a\x95\x6c\x0c\x84\x3d\x6a\xc7\xc9\xb3\x79\xd8\x25\x6f\xc0\x46\x39\xa9\xb5\xe1\x8f\x06\x5a\xe2\x27\x44\xda\x4f\xfa\xbe\x92\x41\x6a\xe0\x7a\x33\x29\x35\x3a\xa3\xbf\xc1\x17\xd6\x95\x3e\x36\x21\x4e\xfa\x72\xff\x0a\xa3\xb1\x69\x28\xda\x10\x89\x09\xa1\xe8\x4b\x72\x8b\x16\x61\x42\xc5\xb6\xa0\x5c\xa2\x0d\x8e\xb1\x69\xe0\x9f\x4c\x64\x15\x89\x15\x04\x85\xc4\x05\xa5\xd1\x59\xaf\xf7\x25\xb9\xed\x02\x84\xae\x13\x2e\xc5\xb6\x80\xb5\x01\xc8\x2a\x83\x30\xf5\xe7\x58\xdb\x8e\x80\x86\xaf\x49\x00\x31\x76\x82\x72\x16\x03\xab\x58\x1e\xe3\x22\x38\xcb\x3e\x8a\x57\x41\x27\x26\x3e\xb1\x13\xd2\xcd\x02\x34\xe7\x9c\x8c\x54\x5d\x1e\x6c\x67\x5c\x45\x7e\x68\xbb\x75\x3e\x53\x36\xb6\x33\x51\x3b\xce\x39\x84\x13\x74\xa9\xa8\x58\xba\x31\x14\x13\xa3\xd3\xfb\x32\xde\x35\x12\xf1\x4b\xdc\x9f\x98\xd8\x1a\x19\x7a\x1f\x8f\xc7\xf9\x4e\xcf\x37\x0e\x7f\x5e\xa3\x0e\x46\xc4\x59\x84\x28\xd5\x04\xd9\xbe\xdf\x9e\x82\x1d\xea\xd2\xb9\x4e\x87\x08\xe7\xda\xee\x92\x6d\xca\xc6\x9a\xb9\x58\x98\xad\x02\x87\x7a\x61\x20\xcd\x49\x60\x81\x47\x91\x29\x1b\x43\x6c\xb6\x05\xa1\x35\xf3\x7c\x22\x0d\x55\xfd\xc2\x32\x70\x7f\x62\x8c\x31\x62\xcc\x8a\x36\x9a\x98\xe8\x0b\xe6\x7f\x20\x03\x6b\xea\x13\xd3\x62\x1f\xaf\x2c\x55\x19\x33\x6e\x6f\x96\x06\x17\xa3\x87\x37\xad\x59\x18\x13\xdb\x59\xf0\xd7\x80\x2a\xc3\xc3\x48\xd6\x6d\xa1\xd5\x6a\xcd\x09\xb5\x60\xe1\xc0\x5e\x12\xcb\x09\x97\x51\x18\x90\x20\x75\x97\x26\x5f\xe7\x2c\xc0\xbe\x46\xf0\xc6\xba\xc1\x0f\x72\xa6\x99\xb0\xc4\xc9\x93\x2c\xf0\xb1\xa4\x31\x31\x38\xfe\x21\xd6\x32\x3a\xbe\xd9\xd2\x07\x58\x68\xdd\x5d\x70\x0e\xdb\x75\x2d\x67\x95\xd0\x70\x09\xeb\x2c\xed\xc0\x95\xe0\x75\xab\xc5\xb3\x0d\x78\xa6\x24\x6f\xcd\xbf\xdd\xe8\xc6\x73\x45\x1b\x56\x52\x51\xb1\x0e\xa7\x29\x5c\x9b\xe5\x87\x35\xea\x84\x2b\xea\x7a\x31\x02\x73\x77\xa2\x38\xa4\xa1\x13\xfa\x10\xb5\xe9\x36\xd3\x55\xeb\x52\x95\x87\x6b\x31\x7f\xc7\x14\x14\x0b\x89\x58\x33\xb3\x6f\xbb\x44\xd6\x24\xa2\x06\x79\x03\x3c\xc2\xda\x60\x5c\x41\x58\x26\x03\x2a\x6e\x24\xdf\x4b\xa8\x24\x8f\x80\x1c\x6d\x84\xc7\xa6\x81\x18\x03\x09\xdc\x2c\x3e\xd8\x73\xc9\xc8\xd4\x8e\xe7\x84\x4a\x6c\x9d\x21\x36\xd7\x48\x56\x55\x54\x20\xd9\x90\xce\xc5\x25\x84\x5a\xae\x17\x13\x87\x86\xf1\x5b\x28\xc1\x11\x89\xa9\x47\x12\x69\x64\xe8\x23\x6c\x98\x0a\x1e\x23\x79\x30\x50\x4c\x45\xd7\x64\xd5\xe2\x3b\x49\xc5\xb2\xc6\xf4\x18\x83\xc1\x36\xe5\xc2\x7e\x69\xe5\xe9\x16\xf0\x66\x9b\x07\x12\xc4\x87\xdf\xfe\xe1\xfd\xcf\x7f\xfd\xee\x1f\x7f\x06\xab\xbe\xff\xdd\xcf\xbe\xfd\xe6\xef\x1f\x7e\xff\xd3\x77\x5f\xff\xed\xfd\x2f\xfe\xf2\xef\x3f\xfe\xea\xc3\x5f\x7f\xf9\xee\x9f\x5f\x0b\x4b\xdb\x89\x43\xb6\xd9\x12\x12\xdf\x91\x18\xc9\xa3\x51\x69\xcb\x41\x2c\xf2\x8d\x9f\xc2\x38\xbd\x97\x8d\x21\xb7\x50\x5a\x8d\x73\xad\x40\x99\xd3\xfb\x91\xa1\xff\x18\xf7\x2b\x35\x19\x88\xd3\x7c\x55\xd9\xbd\x63\xa3\x9f\x4a\x45\xe2\x17\x5d\x56\x3c\x04\x66\xe6\x79\x68\xf1\x5e\xc5\xbe\xf5\x49\x61\xe6\xf6\xde\x29\xc0\x9b\x55\xb6\x3a\xec\x45\x9e\x4d\x24\x16\x62\x65\x89\x6c\x41\x97\x44\x24\x70\x49\xe0\x70\x1d\x32\xb7\xd6\x88\x73\x63\xb7\x4e\x20\x11\x03\x01\x3a\x39\xd6\x0f\x84\xae\x12\x8a\x9c\x98\xd8\x94\x20\xba\x8c\x50\x1e\x30\xdf\x47\xf6\x5d\xe8\xb9\xc8\xf7\x82\xd5\x1b\xc4\xaa\x00\x72\xc2\x60\xe6\x7b\x0e\xfd\x9e\xd0\x1a\x63\x53\x32\x26\x1a\x44\x47\x96\x7e\x79\x3e\xdb\x55\x4c\x0a\xcd\xf2\xb2\xd2\xaa\xa4\xe6\x26\x89\x87\xe4\xff\xe5\x6b\xd8\xdc\x9d\xa8\xb1\xd7\xa1\xcb\xa8\x57\x98\x98\xd5\x07\x8e\x22\xdd\xaa\xc7\x83\x11\x2f\x51\x27\x9e\x7d\x3a\x10\x60\xe3\x22\x05\x7c\x62\x63\x38\x61\xf4\x16\x41\xe3\x56\x44\xfb\x83\x80\x7a\x1c\x91\xd0\x52\x2e\xf3\xce\xf4\xf8\xf0\xea\xc8\xa0\x5d\xdd\x1b\x5d\x0b\xab\x63\x2c\x7d\x24\x64\x8f\x40\x05\x1d\xa9\xd4\x6e\x41\x1b\x93\x56\x95\xb4\x49\x3c\xbd\x2f\x27\xd8\xf5\x93\x20\x2f\x5c\x9e\xad\xf6\x08\xe8\x15\x77\x3f\x91\x55\x0f\x81\x95\x3b\xfa\x73\xc2\x94\xb9\xb9\xdc\x3e\x6c\x73\xf4\x13\xec\x6d\x67\x01\xa9\x76\x1b\x5e\x56\xaf\x9c\xaf\x66\x77\xa8\xd8\x5b\x74\xfe\x15\xaa\xe9\xf1\x24\x06\xb5\xbc\x99\xe5\x7a\xb3\x19\x89\x49\x40\x1f\xb6\x69\x8a\xac\x4e\x55\xa1\xc8\x52\x52\x43\x0b\x5c\x34\xbf\x35\x11\x5d\xea\x2d\x49\x42\xed\x65\x24\x6c\xe9\x84\x6b\x3c\x42\xa5\x1d\xde\xd4\x70\xb4\xcd\x4c\xc2\x66\x8f\x0a\xc0\x85\x72\x23\xdc\xec\x93\x4d\xa5\x36\x5b\x4e\x28\xe0\xa5\x96\x73\x97\x8a\xf5\x40\x2b\x9d\xb9\xf6\x77\xe2\x56\x15\x33\x17\xf0\x96\x28\x3d\xec\x1d\xb1\x2b\xca\x9a\x9b\x74\xf7\x6c\xf6\x37\x59\xc3\xd9\x7c\xc6\x6e\x6a\x72\x6a\xe5\xab\xdc\xf0\x34\xac\x72\x48\xb2\x4d\x8d\x9c\x0d\x2a\xd6\x3d\x3b\xf2\x7a\x1c\x92\x65\x07\xae\x15\xad\x6e\x7d\x2f\x59\xa0\xce\x65\xb2\x4a\xc8\xf9\x8f\xaa\x61\xdf\xb9\xb4\xa3\xc8\xf7\x1c\x1b\xba\xf9\xf3\xd3\x7b\x99\x9d\x92\x2e\x97\xa1\xbb\xf2\x89\x05\x07\xca\xf3\xa2\x06\x75\x2e\x21\xe2\x49\x40\xcf\x5d\x72\x47\x7c\x38\x58\x74\xec\x15\x0d\x53\x0b\x64\xa9\xa9\xbe\xa7\x1f\xab\x6a\xe9\x14\x4a\x17\x5e\xec\x46\x76\x4c\xdf\xb2\xe9\x8f\xb3\x8a\xfd\xb4\x4a\xef\xa5\xf6\x0f\xa9\xe7\xbc\x26\xf4\x3c\x3f\xe2\x3d\xc7\xda\xfa\xbf\xc8\x16\x99\x97\xa7\x41\xe7\x78\xbf\xc6\xfe\xe5\xff\xa1\xb9\x87\xaa\x10\x7e\xff\x2b\xa1\xb7\xd5\xad\x4f\x10\x8b\x79\xdb\x0a\xd9\x9d\x8f\x0d\xde\xfd\xe6\x5f\xdf\x7e\xf3\xa7\xdd\xa5\x89\x2b\xfc\xa8\xca\x5a\x2b\x91\x50\xdf\xca\x25\x97\x1f\x79\xf3\x22\xd4\x64\xb3\x4a\x55\x4d\xcb\x10\xf7\x59\xb7\xea\x56\x3b\x70\x51\x1a\x0d\xdd\x6e\x57\xdc\x52\x1f\x0f\x98\x77\xee\x42\x97\x57\x49\xc4\xeb\x64\x36\x71\x3d\x62\xa1\x3c\x41\x59\xa9\x4c\x27\xc1\x7b\xd7\xca\x14\x4c\x43\xb1\x04\xa9\x9b\x93\xd3\xd6\x09\xda\x63\x74\x7a\x82\x0e\x1f\x9e\x3e\xc4\xbb\x7d\x7c\x0a\x9c\x8f\x1d\xa0\x02\x4f\xe9\xb2\x05\x1e\xeb\xc9\xb7\xc9\xa4\xdf\x6d\x00\xe0\x9c\x4d\xa7\xd4\x8e\x13\x36\xf0\x9c\x4e\xb3\xb9\xe7\x74\xca\x83\x16\xfe\x56\x27\x01\x05\xb0\x4a\xfe\x38\x02\x32\x7e\x3e\x8c\xb2\xf9\x48\x79\x16\xfa\x39\xc1\x4c\x8f\x65\x47\xc0\x47\xfc\x84\x78\x33\x00\xa8\xe2\x8f\xe8\xf2\x53\xac\xbd\xb8\xbf\xd2\xaf\xf1\xba\x97\x43\xee\x65\x88\x7b\x1c\x70\x6d\x00\xf4\x69\xdc\xfe\x99\x41\x6d\x74\xfd\xc1\x18\x53\xf7\x4f\x34\xe5\xe5\x47\xf4\x7e\x6f\x11\x2e\x49\x09\xe9\x67\xe5\xef\x4f\x0e\xae\xd1\xc3\x8f\x40\x95\xcd\xe6\xe1\xa1\x34\x7c\xd9\x5a\xa0\x18\xe1\xae\x36\x26\x2d\x8a\x5c\xfa\x9e\x9d\x0c\x90\x36\x1c\xf7\xe1\xf5\x96\x4e\xa6\xd1\x72\x05\x07\x6b\x66\x1c\xdb\xf7\x77\x91\x8b\x6d\x01\xa5\x3f\xc6\xd9\xd0\xcd\x1c\x72\x37\xbc\x13\x62\x6e\x79\x66\x77\x24\xb1\x8b\x7f\x7e\x9b\xd4\x16\x84\xba\x5d\xf3\xa6\x70\xdf\x9e\xb0\xc1\x5e\xe5\xbe\x8c\xdd\x81\xd7\x97\xc9\x9c\xb6\xbf\xcf\xb6\xf8\xa6\x62\xb5\x46\x8d\x78\x87\xfa\x5d\xd5\x31\x65\x76\x41\xb6\x24\x49\x62\xcf\x89\x24\x1e\xb1\x89\x2f\x7e\x62\xbb\x58\x80\x83\x2b\xae\xe7\xce\x50\xfe\xcb\x90\x97\xee\xee\xea\x9c\x85\x5a\x0d\x9c\x25\x9d\xcb\x9c\xf5\x4b\xc1\xb3\x82\xb3\xe1\xc6\xb0\x61\xd5\x89\xa2\x0e\x2c\xf3\xd5\x08\x37\xad\x9a\x7f\xac\xae\xaa\xca\xe6\xa5\x6e\x5c\x97\x38\x4a\xab\xa6\x1f\x2b\x1c\xd9\x1d\x72\x23\x47\x7e\xc1\xbc\xc9\x91\x9f\x39\xcf\x1a\x38\x8a\x03\x69\x8d\x8d\x9d\x3d\x6b\x4b\x55\x0e\xa6\x65\x9e\xa7\x8a\x0c\xe1\x3f\x01\x00\x00\xff\xff\x1c\xe0\x38\x88\x1a\x26\x00\x00")
+var _cmakeTarsToolsCmake = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xec\x5a\x4f\x8f\xe3\x4a\x11\x3f\xc7\x9f\xa2\xf1\xcc\xc1\x79\x22\x89\xb4\xdc\x46\x1a\x81\x27\xe9\xc9\x98\xf5\xd8\xc1\x71\x76\x76\xa5\x48\x96\xc7\xee\x24\xd6\x3a\xb6\x65\x77\xe6\xed\xbe\x51\x6e\x48\x80\x40\x7a\x20\x21\x0e\x5c\x10\x12\x02\x6e\xdc\x10\xe2\x89\x6f\xf3\x76\x97\x8f\x81\xaa\xdb\x7f\x63\x27\x93\xc9\xcb\xec\x2e\x82\x5c\x76\x6d\x57\x55\xff\xea\x4f\x57\x55\x57\x8f\x20\x90\x37\xc4\x59\x51\x62\x45\x71\xe8\x90\x24\x91\xfa\xfa\xf5\xb5\xac\x0d\xd0\x3c\x44\x24\xb8\x43\x43\x7d\x24\x9b\x57\x48\x9f\x98\xa3\x89\x69\xbd\x90\x0d\x45\xbe\x50\x71\xfa\xba\x2d\x24\x34\xf6\x82\xb9\x64\xe0\x21\x7e\x89\x0c\x3c\x52\xe5\x3e\x46\xe2\x34\x38\x15\x91\x28\x66\xcc\xe2\xe9\x3d\xff\xdf\x5a\x6c\x0b\x09\xa1\x52\xff\x5a\x7e\x8e\xad\x6b\x7d\x30\x51\xb1\xc5\x48\x4e\xef\x6b\xef\xd6\xc0\xc7\xdf\xf6\x27\x86\x81\x35\xd3\x1a\xeb\x13\xa3\x8f\xad\x81\x62\xac\x7b\xce\xd2\x7e\x4d\xc4\xb6\x20\x78\x81\xe3\xaf\x5c\x22\x3d\x48\xdc\x9b\x87\xbe\x1d\xcc\xbb\xec\xa1\x2d\x94\x90\xc8\x46\xff\x4a\x79\x81\xad\x54\xcb\x81\x62\xe0\xbe\xa9\x1b\xaf\x72\x58\x17\x8a\x26\x1b\xaf\xb8\x2c\xdf\xbb\x2d\x6b\xa1\x2a\x17\x06\x7c\x3b\x84\xd7\x98\x68\xa6\x72\xbd\xe7\xba\xb7\x5e\x00\xda\xce\xa4\x1b\x45\xfb\xc1\xb3\xb6\xd0\x02\x39\xa6\x6c\x8c\x9f\xf5\x47\xa3\x92\x8d\x81\xb0\x47\xed\x38\x79\x36\x0f\xbb\xe4\x0d\xd8\x28\x27\xb5\x36\xfc\xd1\x40\x4b\xfc\x84\x48\xfb\x49\xdf\x57\x32\x48\x0d\x5c\x6f\x26\xa5\x46\x67\xf4\x37\xf8\xc2\xba\xd2\xc7\x26\xc4\x49\x5f\xee\x5f\x61\x34\x36\x0d\x45\x1b\x22\x31\x21\x14\x7d\x49\x6e\xd1\x22\x4c\xa8\xd8\x16\x94\x4b\xb4\xc1\x31\x36\x0d\xfc\x93\x89\xac\x22\xb1\x82\xa0\x90\xb8\xa0\x34\x3a\xeb\xf5\xbe\x24\xb7\x5d\x80\xd0\x75\xc2\xa5\xd8\x16\xb0\x36\x00\x59\x65\x10\xa6\xfe\x1c\x6b\xdb\x11\xd0\xf0\x35\x09\x20\xc6\x4e\x50\xce\x62\x60\x15\xcb\x63\x5c\x04\x67\xd9\x47\xf1\x2a\xe8\xc4\xc4\x27\x76\x42\xba\x59\x80\xe6\x9c\x93\x91\xaa\xcb\x83\xed\x8c\xab\xc8\x0f\x6d\xb7\xce\x67\xca\xc6\x76\x26\x6a\xc7\x39\x87\x70\x82\x2e\x15\x15\x4b\x37\x86\x62\x62\x74\x7a\x5f\xc6\xbb\x46\x22\x7e\x89\xfb\x13\x13\x5b\x23\x43\xef\xe3\xf1\x38\xdf\xe9\xf9\xc6\xe1\xcf\x6b\xd4\xc1\x88\x38\x8b\x10\xa5\x9a\x20\xdb\xf7\xdb\x53\xb0\x43\x5d\x3a\xd7\xe9\x10\xe1\x5c\xdb\x5d\xb2\x4d\xd9\x58\x33\x17\x0b\xb3\x55\xe0\x50\x2f\x0c\xa4\x39\x09\x2c\xf0\x28\x32\x65\x63\x88\xcd\xb6\x20\xb4\x66\x9e\x4f\xa4\xa1\xaa\x5f\x58\x06\xee\x4f\x8c\x31\x46\x8c\x59\xd1\x46\x13\x13\x7d\xc1\xfc\x0f\x64\x60\x4d\x7d\x62\x5a\xec\xe3\x95\xa5\x2a\x63\xc6\xed\xcd\xd2\xe0\x62\xf4\xf0\xa6\x35\x0b\x63\x62\x3b\x0b\xfe\x1a\x50\x65\x78\x18\xc9\xba\x2d\xb4\x5a\xad\x39\xa1\x16\x2c\x1c\xd8\x4b\x62\x39\xe1\x32\x0a\x03\x12\xa4\xee\xd2\xe4\xeb\x9c\x05\xd8\xd7\x08\xde\x58\x37\xf8\x41\xce\x34\x13\x96\x38\x79\x92\x05\x3e\x96\x34\x26\x06\xc7\x3f\xc4\x5a\x46\xc7\x37\x5b\xfa\x00\x0b\xad\xbb\x0b\xce\x61\xbb\xae\xe5\xac\x12\x1a\x2e\x61\x9d\xa5\x1d\xb8\x12\xbc\x6e\xb5\x78\xb6\x01\xcf\x94\xe4\xad\xf9\xb7\x1b\xdd\x78\xae\x68\xc3\x4a\x2a\x2a\xd6\xe1\x34\x85\x6b\xb3\xfc\xb0\x46\x9d\x70\x45\x5d\x2f\x46\x60\xee\x4e\x14\x87\x34\x74\x42\x1f\xa2\x36\xdd\x66\xba\x6a\x5d\xaa\xf2\x70\x2d\xe6\xef\x98\x82\x62\x21\x11\x6b\x66\xf6\x6d\x97\xc8\x9a\x44\xd4\x20\x6f\x80\x47\x58\x1b\x8c\x2b\x08\xcb\x64\x40\xc5\x8d\xe4\x7b\x09\x95\xe4\x11\x90\xa3\x8d\xf0\xd8\x34\x10\x63\x20\x81\x9b\xc5\x07\x7b\x2e\x19\x99\xda\xf1\x9c\x50\x89\xad\x33\xc4\xe6\x1a\xc9\xaa\x8a\x0a\x24\x1b\xd2\xb9\xb8\x84\x50\xcb\xf5\x62\xe2\xd0\x30\x7e\x0b\x25\x38\x22\x31\xf5\x48\x22\x8d\x0c\x7d\x84\x0d\x53\xc1\x63\x24\x0f\x06\x8a\xa9\xe8\x9a\xac\x5a\x7c\x27\xa9\x58\xd6\x98\x1e\x63\x30\xd8\xa6\x5c\xd8\x2f\xad\x3c\xdd\x02\xde\x6c\xf3\x40\x82\xf8\xf0\xdb\x3f\xbc\xff\xf9\xaf\xdf\xfd\xe3\xcf\x60\xd5\xf7\xbf\xfb\xd9\xb7\xdf\xfc\xfd\xc3\xef\x7f\xfa\xee\xeb\xbf\xbd\xff\xc5\x5f\xfe\xfd\xc7\x5f\x7d\xf8\xeb\x2f\xdf\xfd\xf3\x6b\x61\x69\x3b\x71\xc8\x36\x5b\x42\xe2\x3b\x12\x23\x79\x34\x2a\x6d\x39\x88\x45\xbe\xf1\x53\x18\xa7\xf7\xb2\x31\xe4\x16\x4a\xab\x71\xae\x15\x28\x73\x7a\x3f\x32\xf4\x1f\xe3\x7e\xa5\x26\x03\x71\x9a\xaf\x2a\xbb\x77\x6c\xf4\x53\xa9\x48\xfc\xa2\xcb\x8a\x87\xc0\xcc\x3c\x0f\x2d\xde\xab\xd8\xb7\x3e\x29\xcc\xdc\xde\x3b\x05\x78\xb3\xca\x56\x87\xbd\xc8\xb3\x89\xc4\x42\xac\x2c\x91\x2d\xe8\x92\x88\x04\x2e\x09\x1c\xae\x43\xe6\xd6\x1a\x71\x6e\xec\xd6\x09\x24\x62\x20\x40\x27\xc7\xfa\x81\xd0\x55\x42\x91\x13\x13\x9b\x12\x44\x97\x11\xca\x03\xe6\xfb\xc8\xbe\x0b\x3d\x17\xf9\x5e\xb0\x7a\x83\x58\x15\x40\x4e\x18\xcc\x7c\xcf\xa1\xdf\x13\x5a\x63\x6c\x4a\xc6\x44\x83\xe8\xc8\xd2\x2f\xcf\x67\xbb\x8a\x49\xa1\x59\x5e\x56\x5a\x95\xd4\xdc\x24\xf1\x90\xfc\xbf\x7c\x0d\x9b\xbb\x13\x35\xf6\x3a\x74\x19\xf5\x0a\x13\xb3\xfa\xc0\x51\xa4\x5b\xf5\x78\x30\xe2\x25\xea\xc4\xb3\x4f\x07\x02\x6c\x5c\xa4\x80\x4f\x6c\x0c\x27\x8c\xde\x22\x68\xdc\x8a\x68\x7f\x10\x50\x8f\x23\x12\x5a\xca\x65\xde\x99\x1e\x1f\x5e\x1d\x19\xb4\xab\x7b\xa3\x6b\x61\x75\x8c\xa5\x8f\x84\xec\x11\xa8\xa0\x23\x95\xda\x2d\x68\x63\xd2\xaa\x92\x36\x89\xa7\xf7\xe5\x04\xbb\x7e\x12\xe4\x85\xcb\xb3\xd5\x1e\x01\x1d\xdc\xad\x8c\x2b\x7d\x42\x2a\x83\xb5\x39\x4f\x62\xe6\xca\x3e\x79\x3c\xe2\x3c\x06\x9e\x28\x08\x0e\xc2\x94\x45\x40\xb9\xb3\xd8\x16\x03\x4f\xb0\xed\x9d\x05\x64\xe1\x6d\x78\x59\x29\x73\xbe\x9a\xdd\xa1\x62\xdb\xd1\xf9\x57\xa8\xa6\xc7\xd3\xb8\xdb\x9b\x59\xae\x37\x9b\x91\x98\x04\xf4\x61\x9b\xa6\xc8\xea\x54\x15\x8a\x2c\x5b\x35\x74\xc7\x45\x5f\x5c\x13\xd1\xa5\xde\x92\x24\xd4\x5e\x46\xc2\x96\x26\xb9\xc6\x23\x54\x3a\xe5\x4d\x0d\x47\xdb\xcc\x24\x6c\xb6\xaf\x00\x5c\x28\xf7\xc8\xcd\x3e\xd9\x54\x6a\xb3\x1b\x85\xda\x5e\xea\x46\x77\xa9\x58\x0f\xb4\xd2\x71\x6c\x7f\x27\x6e\x55\x31\x73\x01\xef\x96\xd2\x73\xe0\x11\x1b\xa6\xac\xef\x49\x77\xcf\x66\xeb\x93\xf5\xa2\xcd\xc7\xef\xa6\xfe\xa7\x56\xd9\xca\xbd\x50\xc3\x2a\x87\xe4\xe1\xd4\xc8\xd9\x0c\x63\xdd\xb3\x23\xaf\xc7\x21\x59\x76\xe0\x5a\xd1\xea\xd6\xf7\x92\x05\xea\x5c\x26\xab\x84\x9c\xff\xa8\x1a\xf6\x9d\x4b\x3b\x8a\x7c\xcf\xb1\xa1\xd1\x3f\x3f\xbd\x97\xd9\x01\xea\x72\x19\xba\x2b\x9f\x58\x70\xd6\x3c\x2f\xca\x53\xe7\x12\x22\x9e\x04\xf4\xdc\x25\x77\xc4\x87\x33\x47\xc7\x5e\xd1\x30\xb5\x40\x96\x9a\xea\x7b\xfa\xb1\xaa\x96\x0e\xa8\x74\xe1\xc5\x6e\x64\xc7\xf4\x2d\x1b\x0c\x39\xab\xd8\x4f\x0b\xf8\x5e\x6a\xff\x90\x7a\xce\x6b\x42\xcf\xf3\xd3\xdf\x73\xac\xad\xff\x8b\x6c\x91\x79\x79\x1a\x74\x8e\xf7\x6b\x6c\x6d\xfe\x1f\x9a\x7b\xa8\x0a\xe1\xf7\xbf\x12\x7a\x5b\xdd\xfa\x04\xb1\x98\x77\xb4\x90\xdd\xf9\x44\xe1\xdd\x6f\xfe\xf5\xed\x37\x7f\xda\x5d\x9a\xb8\xc2\x8f\xaa\xac\xb5\x12\x09\xf5\xad\x5c\x72\xf9\x69\x38\x2f\x42\x4d\x36\xab\x54\xd5\xb4\x0c\x71\x9f\x75\xab\x6e\xb5\x03\x17\xa5\xd1\xd0\xed\x76\xc5\x2d\xf5\xf1\x80\x51\xe8\x2e\x74\x79\x95\x44\xbc\x4e\x66\xc3\xd8\x23\x16\xca\x13\x94\x95\xca\x74\x48\xbc\x77\xad\x4c\xc1\x34\x14\x4b\x90\xba\x39\x54\x6d\x9d\xa0\x3d\xa6\xaa\x27\xe8\xf0\xb9\xea\x43\xbc\xdb\x27\xab\xc0\xf9\xd8\xd9\x2a\xf0\x94\xee\x61\xe0\xb1\x9e\x7c\x9b\x4c\xfa\xdd\x66\x03\xce\xd9\x74\x4a\xed\x38\x61\xb3\xd0\xe9\x34\x1b\x89\x4e\xa7\x3c\x68\xe1\xdf\xea\x90\xa0\x00\x56\xc9\x1f\x47\x40\xc6\x8f\x8e\x51\x36\x3a\x29\x8f\x49\x3f\x27\x98\xe9\xb1\xec\x08\xf8\x88\x9f\x10\x6f\x06\x00\x55\xfc\x11\x5d\x7e\x8a\xb5\x17\xf7\x57\xfa\x35\x5e\xf7\x72\xc8\xbd\x0c\x71\x8f\x03\xae\xcd\x86\x3e\x8d\xdb\x3f\x33\xa8\x8d\xae\x3f\x18\x63\xea\xfe\x89\xa6\xbc\xfc\x88\xde\xef\x2d\xc2\x25\x29\x21\xfd\xac\xfc\xfd\xc9\xc1\x35\x7a\xf8\x11\xa8\xb2\xb1\x3d\x3c\x94\x86\x2f\x5b\x0b\x14\x23\xdc\xd5\xc6\xa4\x45\x91\x4b\xdf\xb3\x93\x01\xd2\x86\xe3\x3e\xbc\xde\xd2\xc9\x34\x5a\xae\xe0\x60\xcd\x8c\x63\xfb\xfe\x2e\x72\xb1\x2d\xa0\xf4\xc7\x38\x1b\xba\x99\x43\xae\x8d\x77\x42\xcc\x2d\xcf\xec\x8e\x24\xf6\x37\x01\xfc\xa2\xa9\x2d\x08\x75\xbb\xe6\x4d\xe1\xbe\x3d\x61\x83\xbd\xca\x7d\x19\xbb\x1e\xaf\x2f\x93\x39\x6d\x7f\x9f\x6d\xf1\x4d\xc5\x6a\x8d\x1a\xf1\x0e\xf5\xbb\xaa\x63\xca\xec\xee\x6c\x49\x92\xc4\x9e\x13\x49\x3c\x62\x13\x5f\xfc\xc4\x76\xb1\x00\x07\x57\xdc\xdc\x9d\xa1\xfc\x97\x21\x2f\x5d\xeb\xd5\x39\x0b\xb5\x1a\x38\x4b\x3a\x97\x39\xeb\xf7\x85\x67\x05\x67\xc3\x65\x62\xc3\xaa\x13\x45\x1d\x58\xe6\xab\x11\x6e\x5a\x35\xff\x58\x5d\x55\x95\xcd\x4b\xdd\xb8\x2e\x71\x94\x56\x4d\x3f\x56\x38\xb2\xeb\xe5\x46\x8e\xfc\xee\x79\x93\x23\x3f\x73\x9e\x35\x70\x14\x07\xd2\x1a\x1b\x3b\x7b\xd6\x96\xaa\x1c\x4c\xcb\x3c\x4f\x15\x19\xc2\x7f\x02\x00\x00\xff\xff\x83\x32\x0a\xf3\x35\x26\x00\x00")
 
 func cmakeTarsToolsCmakeBytes() ([]byte, error) {
 	return bindataRead(
@@ -201,47 +201,47 @@ func cmakeTarsToolsCmake() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "cmake/tars-tools.cmake", size: 9754, mode: os.FileMode(493), modTime: time.Unix(1661308924, 0)}
+	info := bindataFileInfo{name: "cmake/tars-tools.cmake", size: 9781, mode: os.FileMode(493), modTime: time.Unix(1682498124, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
 
-var _scriptsMakefileTars = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xb4\x57\x6d\x73\xa3\x38\x12\xfe\x0c\xbf\xa2\x8f\x25\x17\x2b\x63\xf0\x24\xa9\xba\xba\xb2\x8b\x9b\x21\x1e\xe2\xcc\x2e\x6b\x5c\xb6\xb3\x57\x57\xf1\x14\x4b\x40\xd8\xaa\x05\xc4\x81\xec\x64\x92\xf5\x7f\xbf\x92\x04\xd8\x24\x4e\x6e\xa7\x2a\xc9\x97\xc8\xad\x7e\x79\xba\xa5\x7e\xd4\xfc\x64\xbc\xed\x9f\xfa\x53\x4c\xee\x21\x5c\x52\x08\x69\x9a\x93\x04\x03\x2e\x0a\x5a\xa8\xf8\x3e\xa7\x05\x03\x77\xe8\xdb\xae\x0b\x00\x16\xe0\xcc\xbf\x9e\x99\xd7\xf3\x4b\xe3\x9f\xcd\xae\x3d\x1e\x01\x3c\xdb\x7d\x73\x8c\xea\xc8\x9b\xd8\xf3\x2b\xf8\x64\x81\xde\x29\x57\x38\x49\x60\x49\x01\x67\x1b\x90\x1b\x48\x1d\x79\x53\xcf\x9b\x1f\x54\xe0\x1b\x5c\x01\x2a\xa4\xfa\xa3\x94\x6d\x7b\xb7\x24\xeb\x2d\xa9\xfa\xf6\x78\x13\x72\x9b\x07\x6c\x65\xe9\x8f\xe5\xfa\xb6\x64\xd0\xef\x42\x57\xef\x54\x58\xb7\xea\xdc\x9e\xce\xce\x2a\x40\x7d\x8e\x38\x26\x45\xc9\xee\x68\x11\x71\xf4\x8d\x05\x34\x26\x08\x09\xac\x2c\x28\xca\x33\x0e\x98\xc4\xf8\xbf\xd0\xe9\x72\xc3\x2c\x2a\x59\x41\xb2\x25\x18\x74\xcd\x22\x52\x74\xf5\xce\xcf\x67\x23\xcf\xbf\x74\xed\x11\x42\x48\xe5\x31\x1a\x01\x00\x7c\xb0\x6a\x4d\x6b\x83\xb3\x88\x9f\x74\x16\x91\x58\x55\x27\x17\x35\x24\xe5\x47\x30\xe5\x05\x65\x34\x7c\x87\x1a\x8a\x2a\xf9\xb3\xe9\x70\x57\xa6\x3b\x92\x44\x61\x50\x44\x70\x62\xf2\x52\x20\x75\x32\xf5\x1a\x8d\x0f\x4f\x34\x04\x30\x7e\xee\x8d\xc6\x33\x27\x4b\x8a\xde\x1e\xb8\xc0\x1e\xd2\xfc\x7b\xcc\x9b\xc9\x02\x12\x03\xc3\x25\x03\xe3\x01\x34\xbd\x63\x4f\x26\x48\x83\x3f\xff\xdc\x97\xcd\xed\xe9\xc8\x99\x23\x6d\x00\x6c\x85\x33\x58\xa8\xd0\xfe\xc3\xe1\x8a\x82\x76\x73\x6c\x4f\x26\xc7\x40\x0b\x38\x96\x06\xc7\x40\x73\x46\x68\x06\xa4\x04\x9c\xe6\xec\xbb\xf9\x4d\x1b\x00\xbe\x27\x0c\x4e\x07\x6d\x2f\x0a\x4e\x4a\x0c\x0b\x55\x51\x6a\x41\x8d\xea\x6f\x60\xf0\x03\x3e\x43\x2f\x04\x57\x94\x2a\xfa\x98\x42\xb9\x0e\x57\x10\x91\xa2\xcf\xf5\xbb\x90\xd1\x3b\xb8\xc3\x10\x16\x38\x60\x18\x08\xe3\xe1\xa5\xb9\xa2\x28\x4a\xfa\x47\x44\x0a\x30\x72\xe9\x9c\x87\x56\x94\x98\xb4\x70\x35\xbe\x87\x34\xff\x0e\xa2\x5c\x7a\xe7\x14\x81\xf1\x2f\x61\xc4\xb3\x69\x69\x87\x39\x18\x1b\xa9\x22\x11\x0b\xaf\x31\x19\xa8\xaa\xed\xba\x7d\x68\x2a\xf9\x2e\x67\xda\x78\xef\x03\xbf\x99\x17\xd7\x5f\xdd\x2f\xa2\x1b\xf8\xfd\x42\xaa\xc2\x97\x08\x6e\xd7\x24\x89\xa4\x58\x68\xc8\x3e\x04\x83\x82\xfe\xf9\x7d\xee\x1a\x89\x33\x4e\x06\x7a\x87\xf3\x40\x2e\x8b\x20\xfa\x06\xa1\x2e\x52\x1b\xa8\xfd\xfd\x1d\x55\xf9\x2c\x4a\x6f\x60\xd0\x16\xf8\xe6\xfc\x7c\x70\x9a\xca\xed\xb3\x91\x87\x16\xf8\xe6\x63\x0a\x5c\xfe\x8f\xc1\x69\x0a\xfa\x63\x6d\xb7\x05\xb1\x65\x9a\xa6\xc6\xf3\xad\xf5\x61\x9f\x72\x5a\x61\xf8\xb5\xfb\x3f\x10\xb4\x8c\x02\x6f\x66\x71\xfe\x5a\x4d\x46\x4d\x52\x55\x87\xf3\x54\x26\x53\xaf\x71\x53\x8b\x0f\x25\x22\x19\xe9\x59\x12\x95\x49\x2b\x07\x69\xac\x77\x04\xf7\x21\xae\x74\x51\x27\xb2\x05\xbd\x13\x44\x51\x5e\x60\xfe\x3e\x1a\x86\xf0\xea\x0b\x56\xe7\x24\x58\xf2\xe7\x4f\xef\xf0\x3b\xbe\x43\x83\x10\x6a\x61\xd3\x3b\x31\x2d\x70\x10\xae\x44\x72\xdd\xbd\x6c\xf4\x0e\xde\x04\x09\xbc\x16\xbe\x15\xb2\x8a\xc4\xdd\xa0\xea\x1f\x52\x95\x98\x16\xb2\x6d\x48\xb6\x83\xd4\x80\xe1\x4d\x17\x51\xd1\x24\x11\x29\xb2\x20\xc5\x96\x2e\xfc\xf0\x25\xe8\xd2\x8b\xe8\xcc\x17\x11\x2c\xa9\x4f\xd7\xcc\xca\x93\xf5\x92\x64\xa5\xc5\xcf\xa9\xc8\xc3\xbe\xae\xd7\x5e\xda\x20\xf5\xb6\x73\x09\x20\xc3\xf2\x1a\xbc\x76\x7c\xfc\x12\x08\x47\xed\x5b\xf0\x1e\xed\xc2\x82\x62\x8f\x29\x40\xef\x0c\xbd\xf1\xe5\xd7\x11\x87\x42\x62\xb8\x91\x74\x58\xed\xfa\x2c\xcd\x7d\x5e\xf8\x6f\x0d\x3b\xd6\x9c\xc5\xa5\xab\xa0\xe4\x74\x5b\xb2\xfe\x33\x8b\x2e\x04\xb7\xb4\x60\xa6\x26\x49\x6a\x47\xca\x3b\x26\x96\x04\xf9\x3c\xd6\x33\x49\xaf\x91\x88\xc3\xe2\x2c\x58\xc4\x87\x52\x78\xcd\xb4\x27\x81\x84\x07\xb2\x1b\xf0\xfe\x03\xc3\xc0\xf7\x61\xb2\x8e\xb0\xa5\x99\xe5\x26\xd3\xf6\x05\xbe\x14\x84\x0f\x9b\xbd\xb8\x26\x5b\x3e\xc0\xbe\xff\x30\x02\xd3\x94\x61\x64\x25\xe3\xbd\xc7\x8d\x6b\x6b\xad\x3a\x2a\xe9\x06\x8c\x97\x1d\x9a\xbf\x47\xfc\x59\xf9\x70\xf4\x9f\xa3\xf4\x28\x3a\xba\x3a\xfa\xf5\x68\xf6\x3b\x57\x91\x11\xe4\x53\x22\x9c\xbc\x92\xb5\x70\x69\x56\xb9\x17\x69\xbb\x70\xbb\xfc\x77\xc7\xca\x2b\x11\x6e\xe2\x87\x27\xa8\x38\x5b\x08\xb5\x98\x54\x73\xc9\xbf\x9d\x0b\xff\xca\x9b\xcd\x01\xf8\xcc\xb9\x62\x2c\xef\xf7\x7a\x09\x0d\x83\x64\x45\x4b\xd6\x3f\xff\xf8\xf1\xe3\x4e\x71\xee\xfd\xe2\x8c\x85\xa2\xa6\xa9\xd7\x13\xd7\xb3\xbf\xf8\xd7\x33\x67\x2a\xde\xb5\xbd\x91\xf5\x6e\x45\x83\x94\xa0\x5a\xc5\x9b\x55\x4f\xdf\x27\x0b\x12\x92\xad\xef\xd5\x75\x9e\xd0\x20\xea\x43\x35\x7e\x8f\x3c\x6f\x66\xe9\x8f\x8d\xfa\xb6\x51\x60\x41\xd1\xf4\x56\x3b\x15\xc3\x30\x40\x8e\x21\x66\x7b\x03\xbc\x59\x5f\xcc\xc6\xde\x6c\xab\xa9\x4a\xb8\x2e\x92\x9a\xfa\xeb\x64\xb7\xbd\x20\x27\x3d\x19\xc3\x0f\xb2\xc8\xcf\xd7\xb7\x09\x29\x57\x9f\x18\x09\xff\xc0\xcc\xda\x53\x17\x29\x6f\xc1\xb8\x2c\xd7\x25\xb6\x3e\x8b\x9d\x91\x33\xdf\x4a\x0c\x97\x41\x9e\x27\x24\x0c\xf8\xf8\x62\xe9\x8f\xf6\x64\xc2\x55\x53\x1a\xad\x13\xec\x4b\xa6\xaa\x0d\xc0\xb8\x0c\x69\x9a\xe2\x8c\x59\x32\x30\x8e\x8c\xdb\xef\x46\x93\x35\xaf\xe3\xb6\xc9\x75\x91\xbd\x21\x57\x2c\x32\x4d\x55\xaf\x1c\x77\x22\x67\x4b\xbe\xf2\xe7\xf6\x14\xed\x3d\x4f\xcd\x73\xd6\x45\xaa\x3a\x73\xa6\xbf\x39\x53\x7f\x6c\xff\xea\xc8\x51\xb3\x1e\x47\x2a\x3e\x7b\x9f\x49\x33\xc1\x41\xd6\x57\xc5\xed\x16\xdd\xf4\xc5\x99\x38\xe3\x2f\xbe\x40\xe6\x5d\xfc\xcc\x59\xe1\xeb\xf8\x37\xef\x17\xc7\x3f\xb0\xe3\x7a\x43\xdb\xad\x7f\xec\x08\x45\xae\x66\x48\xba\xf3\x2f\xbf\xba\x0e\x7f\x1c\x86\xae\x63\x8f\xf9\x8f\x2d\x98\x27\x66\x64\xb2\x34\x87\x65\x4a\x33\x93\xae\x59\x83\xe0\xa4\x71\x74\xc2\x0f\xbb\x42\x18\x24\xc9\x4b\x20\xaf\xd0\x13\xc1\x70\x32\x79\x2a\x3a\x04\xf7\x6a\xe8\x48\x5b\xbe\xa8\x6c\x7e\x24\x85\x13\x93\xee\xd2\xe0\x8b\xbf\x90\xcb\x93\xdb\x20\xbc\xa1\x83\x52\xdb\x75\x91\xd4\x97\xbf\xe5\xb7\xa7\xb6\xc8\x16\xf8\xe6\x74\x70\x7e\x9e\x8a\xb2\x88\x59\xa4\xbf\x60\x0b\x76\x53\xe0\x94\x6e\xf0\x0b\x67\xf2\x4d\x53\x5b\xae\x0f\xb8\x0a\x92\xa4\xf6\x76\x23\x04\xf0\x77\x28\x52\x91\x58\x6d\x3d\xb7\xa7\xf5\x47\xf0\x9e\x35\x0b\x8a\x3d\x18\x77\x24\x49\x20\xa2\x70\xcc\xa9\xf0\x29\x37\x4f\x1d\xd7\xb1\x67\x8e\xac\xe6\xf1\x37\x4d\x55\x57\x38\xc9\xfb\x7b\x53\x98\x2c\x01\xd2\x44\xef\xbc\xf9\x37\xe1\xff\x02\x00\x00\xff\xff\xfc\xf1\x74\x31\xfd\x10\x00\x00")
+var _scriptsMakefileTarsGomodMk = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xcc\x58\xff\x73\x9b\x38\x16\xff\xd9\xfc\x15\xef\xbc\xe4\x6c\xda\x80\xe3\xe4\xa6\xb3\xe3\x8c\xaf\xa1\x0e\x71\xd2\x75\x8c\x07\x3b\xbd\xb9\x69\x3a\x2c\x01\x81\x35\x05\xc4\x81\xec\x24\x4d\xfd\xbf\xdf\x48\x02\x0c\x0e\xc9\xb6\xbb\xc9\xcc\xe6\x87\x0c\x92\x9e\x3e\xef\xab\x3e\x7a\xf2\x2f\xea\xcb\xfe\x49\xbf\xf8\xf8\x0e\xdc\x80\x80\x4b\xa2\x04\x87\x08\x50\x9a\x92\x54\x42\x77\x09\x49\x29\x4c\x46\xb6\x3e\x99\x00\xc0\x10\x50\x6c\x5f\xcd\xb5\xab\xc5\x99\xfa\x6b\xb9\xaa\x4f\xc7\x00\x8f\x56\x5f\xdc\x46\x69\x6c\xce\xf4\xc5\x39\xbc\x1f\x82\xdc\xcd\x96\x28\x0c\x21\x20\x80\xe2\x35\x88\x05\x45\x1a\x9b\x96\x69\x2e\x1a\x05\xd8\x02\x13\x80\xdc\x52\xf9\x41\xcc\x6d\x7a\x37\x38\xee\x05\x44\x1a\x9b\xf6\xe5\xc5\xd4\xb4\xec\x4f\x86\x35\xbf\x30\xa7\xb0\x05\x91\xbb\x63\x53\x81\x35\x4a\x33\x4c\x62\xf8\x0e\xee\x8a\x82\xea\x42\xff\x5f\x6a\x31\xf0\x3a\xd0\x01\xd5\xef\x6f\xc7\x1a\x1b\x1f\x2a\x8f\x60\xfb\xef\x60\x08\xfd\x77\xd2\xcb\x87\x27\xc4\x37\x89\x43\x97\x43\xf9\x21\x5b\xdd\x64\x14\x06\xfb\xb0\xcf\x2c\xe7\xa1\xd9\x48\x63\xf3\xd2\x3c\xbd\x9a\x18\x53\xfd\xd2\x18\x6c\x7d\x5b\x22\xc7\x03\x35\xee\x43\x40\xb4\x88\x78\xf0\x1d\x9c\xdb\xaf\xd0\x79\x48\x52\x1c\x53\x90\xe5\xc3\x4d\x47\x91\x24\xec\xa3\xff\x41\x77\x5f\xee\xfa\x38\xf6\x32\x9a\xe2\x38\x00\x95\xac\xa8\x87\xd3\x7d\xb9\xfb\xf1\x70\x6c\xda\x67\x13\x7d\xac\x28\x8a\xc4\xc2\x5b\x4e\x00\xc0\xdb\x61\x21\x39\xa4\x4e\x9a\xa9\x49\x4a\x28\x71\x49\x28\xa1\xd8\xc3\xfe\x13\xd0\x11\xf1\x56\x21\xfa\x11\x68\x21\x39\x64\xe9\xdc\xfa\xb7\x29\xc0\x67\x1f\x0e\xf3\x94\xb7\xb8\xcf\x3e\x4e\x33\x7a\x4b\x52\x8f\xf9\x5f\x46\x09\xca\x30\x29\x0a\x2f\x07\x61\xe3\x2b\xe4\x48\x5a\xe8\xd6\xdc\x9e\x5b\x23\x6e\x13\x37\xe9\x16\x87\x9e\xeb\xa4\x1e\xbc\xd1\x58\x7c\x14\x69\x66\x99\xa5\xc4\xdb\x1d\x09\x6e\x18\x2f\xaa\x42\xe2\x11\x48\x40\x94\x97\x37\x9c\xdb\xee\x92\xe4\xde\x67\xdc\x30\x04\xec\x03\x45\x19\x05\xf5\x1b\xb4\xe5\xae\x3e\x9b\x29\x6d\xf8\xfe\xbd\x3a\xb7\xd0\xad\xb1\xb1\x50\xda\xc7\x40\x97\x28\x86\x6b\x09\xea\x7f\xc8\x5d\x12\x68\x7f\xee\xe8\xb3\x59\x07\x48\x0a\x1d\xb1\xa1\x03\x24\xa1\xec\x94\xe1\x0c\x50\x94\xd0\x7b\xed\x4b\xfb\x18\xd0\x1d\xa6\xd0\x3f\xae\xa3\xb4\x50\x98\x21\xb8\x96\x5a\xad\x62\xa2\xb0\xea\x1f\xa0\xb2\x04\x1f\x2a\x4f\x28\x6f\xb5\x72\xed\x53\x02\xd9\xca\x5d\x82\x87\xd3\x01\x93\xdf\x87\x98\xdc\xc2\x2d\x02\x37\x45\x0e\x45\x80\x29\x53\x2f\xb6\xb7\x5a\xad\x56\xf4\xd5\xc3\x29\xa8\x89\x00\x67\xaa\x5b\x2d\x1f\xd7\xec\x2a\xb1\x47\x24\xb9\x07\x1e\x2e\xb9\xdb\x57\x40\xfd\x37\xdf\xc4\xbc\xa9\x49\xbb\x09\xa8\x6b\x21\x22\x2c\xe6\xa8\x3e\x3e\x96\x24\x7d\x32\x19\x40\x19\xc9\x57\xc9\x69\x89\x3e\x00\x56\x99\x1f\xae\x2e\x26\xa7\x30\xb3\x4c\xf1\xc1\x8e\x05\x2b\x34\x45\x6a\x09\x0a\x64\x04\x41\xb1\x77\x5f\x8c\x6f\x56\x38\xf4\x84\x18\xdf\x21\x4e\x2a\xa8\x04\xe4\x93\xd7\x29\x42\xec\xc7\x8c\x2e\xe4\x2e\x63\x8a\x44\x44\x87\x1f\x28\x45\xd9\x57\xa4\xd2\x87\x41\x75\x05\xd8\xb9\x3a\x0c\x88\xd4\x3a\x11\xb9\xb9\x3e\x38\x3a\xfa\x7c\x74\x74\xdc\x8f\x84\xd4\xe1\xd8\x54\xf8\xdc\x41\x04\x62\xed\xdd\x71\x3f\x02\xf9\xa1\x80\xd8\x40\xbe\xac\x69\x5a\x9b\x39\x5f\xec\x82\x2a\x43\x55\x75\x4a\xac\x38\x9f\xb0\xa7\xb4\x23\x26\xdc\x34\x5e\x25\xed\x2d\x1f\xe6\x1e\xe6\x3c\xc0\xfc\x2a\x32\xc2\x60\x8a\x69\x10\x24\xa5\x06\x28\x56\x03\x52\x1f\xa9\x0c\x36\x4d\xdc\x06\x8f\x85\x5c\xa3\xb7\x39\xf2\x8e\xb3\x02\x41\xee\x72\x2a\x55\x98\xd8\x87\xc2\xe3\x0d\xc8\x5d\xc7\xf3\x92\x14\xb1\xee\x41\x15\xd4\x6e\xf3\x4b\x88\x71\x6a\xc6\x9a\x03\xb9\xcb\x8e\xcc\xd6\x6c\x45\x51\x2a\x23\x16\x4b\x9f\xa4\xc8\x71\x97\x3c\x0a\xfb\x15\xb7\xe5\x2e\x5a\x3b\x21\x3c\xa7\xbe\xa6\xb2\xd0\xc4\x70\xb8\x12\xf1\xa1\x48\x2d\x9f\xa4\xe2\x24\xe2\x78\x6b\x56\x69\x10\x3b\xc7\x1e\xe1\xe7\xce\xc3\x69\xec\x44\x68\x28\x73\x2c\xf6\x09\xb2\xcc\x61\xf8\x61\x7f\xd2\x8a\x80\xd8\x64\x45\x87\xb2\x5c\xec\x52\x2b\x49\x78\xb4\x54\xb5\x59\xae\xeb\x11\xb6\xc4\x48\x94\x4f\x53\xda\xab\xc5\xc3\x81\xea\xd5\xf3\x1a\x67\x8e\x3a\x69\x85\x87\x40\xee\x8e\xcc\xe9\xd9\xc5\x98\x99\x82\x7d\xf8\x2c\xc8\x36\x5f\xb5\x69\x94\xd8\x2c\x0f\x5f\x4a\xee\x2d\x18\x91\xcd\x2e\x9d\x8c\x91\x79\x46\x07\x8f\x76\xec\x83\x73\x43\x52\xaa\xb5\x05\x05\x6e\x29\x7f\xcb\xf3\x15\xfa\xdd\xd9\xdc\x2b\x67\x78\xa2\x18\xa9\xa6\x7e\x93\xcd\xcf\x6d\xed\x09\xcd\x6e\x83\x3b\xc7\xec\xa0\x82\xaa\xa2\x3b\x37\x5c\x79\x68\xd8\xd6\xb2\x75\xdc\xae\x4e\xd8\x62\xc2\xfd\xb6\xae\xe8\xd5\x68\xf0\x0d\xaa\xf8\xae\x07\x9a\x26\xd4\x88\xd0\xf9\x95\xbb\x92\x49\xb7\x6b\x81\x6b\x45\x6b\x50\x9f\x06\xd4\x7e\xf7\xd8\x2d\xf5\x76\xef\xbf\x7b\xd1\x9e\xb7\x77\xbe\x77\xb9\x37\xff\x9d\x89\x08\x0d\xe2\x66\xe2\x20\xcf\x78\xcd\x21\xb5\xdc\xf7\x34\xaa\x07\x6e\xeb\xff\x36\x8f\x2c\x12\xee\xda\xff\xb6\x63\x15\x63\x0b\x2e\xe6\xe3\xbc\xcd\xf9\x8f\xf1\xc1\x3e\x37\xe7\x0b\x00\xd6\x91\x2f\x29\x4d\x06\xbd\x5e\x48\x5c\x27\x5c\x92\x8c\x0e\x8e\x0e\x0e\x0e\xb6\x82\x0b\xf3\x37\x63\xca\x05\xdb\x6d\xe9\x6a\x36\x31\xf5\x53\xfb\x6a\x6e\x58\xfc\x9a\xac\x34\xf4\xb7\x4b\xe2\x44\x58\x29\x44\xcc\x79\x7e\x93\xbe\x1f\x42\x88\xe3\xd5\x9d\xb4\x4a\x42\xe2\x78\x03\xc8\x1f\x27\x63\xd3\x9c\x0f\xe5\x87\x52\x7c\x53\x0a\x50\x27\x2d\x0f\x53\xdd\x15\x55\x55\x41\x74\x35\x5a\x7d\x01\xcc\xf9\x80\xbf\x1c\xcc\xf9\xa6\x2d\xb5\xdc\x55\x1a\x16\xb7\x44\xe1\xec\xa6\xe7\x24\xb8\x27\x74\xd8\x4e\xec\xd9\xc9\xea\x26\xc4\xd9\xf2\x3d\xc5\xee\x57\x44\x87\x15\x71\xee\xf2\x06\xd4\xb3\x6c\x95\xa1\xe1\x09\x5f\x19\x1b\x8b\x8d\xb0\xe1\xcc\x49\x92\x10\xbb\x0e\xeb\x86\x86\xf2\x83\x3e\x9b\x31\x51\xd1\xee\xda\x82\xa5\x8a\x0d\xa0\x9e\xb9\x24\x8a\x50\x4c\x87\x42\x31\xf2\xd4\x9b\x7b\xb5\xf4\x9a\xc5\x71\xb3\xbd\x0b\xe2\x17\x24\x87\xeb\xb8\x2d\x49\xd2\xb9\x31\x99\x89\x5e\x95\x7d\xd9\x0b\xdd\x52\x2a\x17\x59\x79\xf1\xed\x2b\x92\x34\x37\xac\x4f\x86\x65\xb3\x2e\x5d\xb4\xae\x45\x7b\x93\x33\xd8\xeb\x74\xae\x21\x72\xe2\x81\xc4\xcb\x9b\x1f\xa7\x53\x63\x66\x4c\x4f\x6d\x6e\x99\xf9\xe1\x23\xa3\x85\x8b\xe9\x27\xf3\x37\xc3\x6e\x58\x99\x98\x23\x7d\x52\x0c\xb6\x8c\x22\xbe\xe6\x8a\x80\xb3\xcf\x2e\x26\x06\xbb\x19\x46\x13\x43\x9f\xb2\xc1\x06\xb4\x37\x9a\xa7\xd1\x28\x81\x20\x22\xb1\x46\x56\xb4\xb4\xe0\x4d\x09\xf4\x86\x65\x3b\xb7\xd0\x09\xc3\xa7\x8c\x3c\x57\x76\x26\x46\xb3\xd9\xee\x54\x93\xb9\xe7\x23\x43\xec\x65\x1f\xf9\x9e\x9f\x71\xe1\x8d\x46\xb6\x6e\xb0\x8f\x1f\xf0\x65\xa7\x1a\x38\x9a\xd2\x38\xab\x4f\x26\x8a\x90\x17\x63\xf1\x34\x6f\x5f\xc7\xbc\x03\xe9\x1f\x1f\x1d\x45\x3c\x30\x79\x43\x32\xb8\xa6\xd7\xf4\x73\x8a\x22\xb2\x46\x4f\xe4\xe5\x4b\x5b\xaa\xc1\x37\xc2\x39\x61\xb8\x45\xfc\xcc\xa7\xe0\x9f\x90\x46\xdc\xc1\x02\x61\xa1\x5b\xc5\x6f\x05\x35\x04\xea\xa4\x35\x73\x6e\x71\x18\x82\x47\xa0\xc3\x78\x71\x97\xa8\x2d\x63\x62\xe8\x73\x43\x44\xb6\xf3\xa5\x2d\x49\x4b\x14\x26\x83\x6d\x57\xc5\x54\x29\x6d\x7e\x8a\x5e\xfc\xb1\x99\xb7\xbd\x83\xe2\x89\x5d\xe1\x50\xec\x2e\x8b\xae\x98\x35\x48\x27\x0f\x8c\xb8\x33\x44\x41\x45\xc0\x6e\xd0\x92\x3c\xfb\xfd\xbe\x78\x56\x0f\x49\xcc\xd9\x3d\xef\x7c\xed\xc5\xe5\xcc\x3e\xbd\xb0\x58\xab\x14\x7d\xa5\x28\x4a\x40\xf5\xf8\xe5\xcb\xae\x4f\xb9\x90\x3a\xbd\xb0\xd8\x5c\x40\xf8\xd3\x01\xc7\x98\x02\x8d\x12\x36\xc5\xaf\x3f\xfe\x72\xb8\xd4\x3f\x6e\x7f\x1e\x51\x40\x0d\x68\xc3\x7c\xff\x9d\x52\xde\x8c\x20\x10\x71\x9c\x51\x27\x0c\x21\xc0\x74\xb9\xba\xd1\x5c\x12\xf5\x16\x4e\x9a\x8d\x42\xb2\xf2\xf8\xd7\x98\xf4\x98\x8f\x3d\x4a\x48\x98\xf5\x72\x77\x4f\x42\x87\xbd\x11\x8f\x2b\x7d\x45\x40\x20\x40\xf4\xcf\xe3\xf8\x98\xfd\x2f\xae\x4e\x79\x27\x42\x3c\x9e\x1b\x29\x9f\x1d\x36\xff\x72\xc5\x7f\x75\x28\x9e\x29\xe5\xd3\xa1\x2a\x5f\xcf\x59\xd1\xee\xd5\xba\xfe\xa7\x12\x5d\x13\xfa\xc9\x74\xcf\x2c\x73\x61\x8e\xec\xb1\x31\xb5\xff\x30\xe9\x8d\xb2\xaf\x9f\x7e\x42\x82\x10\x69\x01\x09\x9d\x38\xd0\x48\x1a\x88\x5f\x6f\x6e\x56\x7e\xcf\x8d\xbc\x5e\xcd\xf9\x93\x75\x5f\x3b\xfc\x55\x3b\x68\xca\xfe\x9f\x84\xd9\x49\x7e\x63\x0c\xf2\x12\xa8\xad\x3d\x57\x08\x35\x65\xe5\x53\xa0\x61\x6f\x53\x7e\x1b\x4b\xa3\x78\x8b\xfc\x50\x89\x14\xc2\x7f\xad\x54\x74\x6b\x6e\xcd\x46\x3f\x57\x32\xf5\x3d\x7f\x17\xe6\x68\x0c\xce\x5f\xe6\x91\x67\x51\xe1\x0f\x2b\xab\x1e\xaa\xa6\x0a\x2b\x64\x7e\xb8\xd2\xca\x5f\x0d\x1e\x57\xdc\x23\xac\xe7\xca\x26\xaf\xc0\xff\x07\x00\x00\xff\xff\x23\x82\x90\x14\xa4\x18\x00\x00")
 
-func scriptsMakefileTarsBytes() ([]byte, error) {
+func scriptsMakefileTarsGomodMkBytes() ([]byte, error) {
 	return bindataRead(
-		_scriptsMakefileTars,
-		"scripts/makefile.tars",
+		_scriptsMakefileTarsGomodMk,
+		"scripts/makefile.tars.gomod.mk",
 	)
 }
 
-func scriptsMakefileTars() (*asset, error) {
-	bytes, err := scriptsMakefileTarsBytes()
+func scriptsMakefileTarsGomodMk() (*asset, error) {
+	bytes, err := scriptsMakefileTarsGomodMkBytes()
 	if err != nil {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "scripts/makefile.tars", size: 4349, mode: os.FileMode(493), modTime: time.Unix(1661308924, 0)}
+	info := bindataFileInfo{name: "scripts/makefile.tars.gomod.mk", size: 6308, mode: os.FileMode(493), modTime: time.Unix(1687749623, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
 
-var _scriptsMakefileTarsGomod = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xcc\x58\xff\x73\x9b\x38\x16\xff\xd9\xfc\x15\xef\xbc\xe4\x6c\x65\x0d\xae\x93\x9b\xce\x8e\x3d\xbe\x0b\x75\x5c\x27\x3d\xc7\x78\xb0\xdb\x99\x9b\xa4\xc3\x12\x10\x58\x53\x40\x1c\xc8\x4e\xdb\x34\xff\xfb\x8d\x24\xc0\xe0\x90\x6c\xbb\xdb\xcc\x9c\x7f\x68\xd1\xd3\xd3\x47\xef\x9b\x3e\x7a\xca\x2f\xda\xcf\xfd\x29\xbf\xf8\xe4\x33\xb8\x01\x05\x97\x46\x09\x09\x31\xe0\x34\xa5\xa9\x82\x3f\x27\x34\x65\x30\x9f\xd8\xc6\x7c\x0e\x00\x63\xc0\xb1\xfd\x7e\xa5\xbf\x5f\xbf\xd5\x7e\x2b\x67\x8d\xc5\x0c\xe0\xd1\xec\x4f\xb7\x51\x99\x99\x4b\x63\x7d\x01\xff\x1a\x83\xda\xcd\x36\x38\x0c\x21\xa0\x80\xe3\x1d\xc8\x09\xa4\xcc\x4c\xcb\x34\xd7\x8d\x0a\x7c\x82\x2b\x40\x6e\xa9\x7a\x2f\x65\x0f\xfd\x5b\x12\xf7\x03\xaa\xcc\x4c\xfb\xea\x72\x61\x5a\xf6\x87\xa9\xb5\xba\x34\x17\xb0\x07\x51\xbb\x33\x13\xc1\x0e\xa7\x19\xa1\x31\x7c\x03\x77\xcb\x40\x73\x61\xf0\x0f\xad\x18\x78\x1d\xe8\x80\xe6\x0f\xf6\x63\x9d\x8f\x4f\xd0\x23\xd8\xc1\x6b\x18\xc3\xe0\xb5\xf2\xf3\xc3\x13\x92\xdb\xc4\x61\x9b\xb1\x7a\x9f\x6d\x6f\x33\x06\xc3\x1e\xf4\xb8\xe5\x22\x34\x0f\xca\xcc\xbc\x32\xcf\xdf\xcf\xa7\x0b\xe3\x6a\x3a\xdc\xfb\xb6\xc1\x8e\x07\x5a\x3c\x80\x80\xea\x11\xf5\xe0\x1b\x38\x77\x9f\xa0\x73\x9f\xa4\x24\x66\xa0\xaa\x27\x0f\x1d\xa4\x28\xc4\xc7\xff\x85\x6e\x4f\xed\xfa\x24\xf6\x32\x96\x92\x38\x00\x8d\x6e\x99\x47\xd2\x9e\xda\x7d\x77\x32\x33\xed\xb7\x73\x63\x86\x10\x52\x78\x78\x4b\x01\x00\xfc\x3a\x2e\x34\xc7\xcc\x49\x33\x2d\x49\x29\xa3\x2e\x0d\x15\x1c\x7b\xc4\x7f\x02\x3a\xa2\xde\x36\xc4\xdf\x03\x2d\x35\xc7\x3c\x9d\x7b\xff\x1e\x0a\xf0\xe5\x9b\x93\x3c\xe5\x2d\xe1\xb3\x4f\xd2\x8c\xdd\xd1\xd4\xe3\xfe\x97\x51\x82\x32\x4c\x08\x89\x72\x90\x36\xbe\x40\x8e\x94\xb5\x61\xad\xec\x95\x35\x11\x36\x09\x93\xee\x48\xe8\xb9\x4e\xea\xc1\xb1\xce\xe3\x83\x94\xa5\x65\x96\x1a\xbf\x1e\x68\x08\xc3\x44\x51\x15\x1a\x8f\x40\x02\x8a\x7e\xbe\xe1\xc2\x76\x97\x26\x5f\x7c\xce\x0d\x63\x20\x3e\x30\x9c\x31\xd0\xbe\x42\x5b\xed\x1a\xcb\x25\x6a\xc3\xb7\x6f\x55\xd9\xda\xb0\x66\xd3\x35\x6a\x8f\x80\x6d\x70\x0c\x37\x0a\xd4\x7f\xd8\xdd\x50\x68\x5f\x77\x8c\xe5\xb2\x03\x34\x85\x8e\x5c\xd0\x01\x9a\x30\x7e\xca\x48\x06\x38\x4a\xd8\x17\xfd\x63\x7b\x04\xf8\x33\x61\x30\x18\xd5\x51\x5a\x38\xcc\x30\xdc\x28\xad\x56\x21\x28\xac\xfa\x1b\x68\x3c\xc1\x27\xe8\x89\xcd\x5b\xad\x7c\xf7\x05\x85\x6c\xeb\x6e\xc0\x23\xe9\x90\xeb\xf7\x20\xa6\x77\x70\x87\xc1\x4d\xb1\xc3\x30\x10\xc6\xb7\x97\xcb\x5b\xad\x56\x2b\xfa\xe4\x91\x14\xb4\x44\x82\xf3\xad\x5b\x2d\x9f\xd4\xec\x2a\xb1\x27\x34\xf9\x02\x22\x5c\x6a\x77\x80\x40\xfb\xa7\x58\xc4\xbd\xa9\x69\xbb\x09\x68\x3b\xa9\x22\x2d\x16\xa8\x3e\x19\x29\x8a\x31\x9f\x0f\xa1\x8c\xe4\x8b\xe4\xb4\x44\x1f\x02\xaf\xcc\x37\xef\x2f\xe7\xe7\xb0\xb4\x4c\xf9\xc1\x8f\x05\x2f\x34\xa4\xb4\x24\x05\x72\x82\x60\xc4\xfb\x52\x8c\x6f\xb7\x24\xf4\xa4\x9a\x58\x21\x4f\x2a\x68\x14\xd4\xb3\x97\x29\x42\xe2\xc7\x9c\x2e\xd4\x2e\x67\x8a\x44\x46\x47\x1c\x28\x84\x7a\x48\x29\x7d\x18\x56\x67\x80\x9f\xab\x93\x80\x2a\xad\x33\x99\x9b\x9b\x57\xa7\xa7\xd7\xa7\xa7\xa3\x41\x24\xb5\x4e\x66\x26\x12\xb2\x57\x11\xc8\xb9\xd7\xa3\x41\x04\xea\x7d\x01\xf1\x00\xf9\xb4\xae\xeb\x6d\xee\x7c\xb1\x0a\xaa\x0c\x55\xdd\x53\xe1\xc5\xf9\x84\x3d\xa5\x1d\x31\x15\xa6\x89\x2a\x69\xef\xf9\x30\xf7\x30\xe7\x01\xee\x57\x91\x11\x0e\x53\x88\x41\x92\x94\x16\xe0\x58\x0b\x68\x7d\xa4\x71\xd8\x34\x71\x1b\x3c\x96\x7a\x8d\xde\xe6\xc8\x07\xce\x4a\x04\xb5\x2b\xa8\x14\x71\xb5\x37\x85\xc7\x0f\xa0\x76\x1d\xcf\x4b\x52\xcc\xbb\x07\x4d\x52\xbb\x2d\x2e\x21\xce\xa9\x19\x6f\x0e\xd4\x2e\x3f\x32\x7b\xb3\x11\x42\x95\x11\x8f\xa5\x4f\x53\xec\xb8\x1b\x11\x85\x5e\xc5\x6d\xb5\x8b\x77\x4e\x08\xcf\x6d\x5f\xdb\x32\xdf\x89\xc3\xa0\xfc\x3f\xa4\xb4\x7c\x9a\xca\x53\x48\xe2\xbd\x49\xa5\x31\xfc\x0c\x7b\x54\x9c\x39\x8f\xa4\xb1\x13\xe1\xb1\x2a\x70\xf8\x27\xa8\x12\x45\x1c\xf4\x27\x2d\x08\xa8\x4d\xb7\x6c\xac\xaa\xc5\x2a\xad\x92\x80\x47\x53\x55\x7b\xd5\xfa\x3e\xd2\x96\x18\xcb\xd2\x69\x4a\x79\xb5\x70\x04\x50\xbd\x72\x5e\xe2\xbc\x31\x27\xad\x70\x10\xa8\xdd\x89\xb9\x78\x7b\x39\xe3\xa6\x10\x1f\xae\x25\xd1\xe6\xb3\x36\x8b\x12\x9b\xe7\xe0\x63\xc9\xbb\x05\x1b\x72\xe9\xc6\xc9\x38\x91\x67\x6c\xf8\x68\x45\x0f\x9c\x5b\x9a\x32\xbd\x2d\xe9\x6f\x4f\xf7\x7b\x8e\xaf\x50\xef\xc1\xe2\x7e\x29\x11\x89\xe2\x84\x9a\xfa\x4d\x36\x3f\xb7\xb4\x2f\x77\x76\x1b\xdc\x19\xf1\x43\x0a\x9a\x86\x3f\xbb\xe1\xd6\xc3\xe3\xb6\x9e\xed\xe2\x76\x55\x60\x4b\x81\xfb\x75\x57\xd9\x57\x67\xc1\x57\xa8\xe2\xbb\x1e\xe8\xba\xdc\x46\x86\xce\xaf\xdc\x93\x5c\xbb\x5d\x0b\x5c\x2b\xda\x81\xf6\x34\xa0\xfe\xbb\xc7\x6f\xa8\x5f\x8f\xfe\x73\x14\x1d\x79\x47\x17\x47\x57\x47\xab\xdf\xb9\x8a\xdc\x41\xde\x4a\x02\xe4\x19\xaf\x05\xa4\x9e\xfb\x9e\x46\xf5\xc0\xed\xfd\xdf\xe7\x91\x47\xc2\xdd\xf9\x5f\x0f\xac\xe2\x4c\x21\xd4\x7c\xa2\x28\xca\xc5\x74\xbe\x94\x9d\x0b\xff\xb2\xd7\x86\x85\x2a\xb4\x56\xd2\x60\x0f\x29\xca\x6a\x6a\x7d\x98\x5a\x36\xef\xd9\x64\x23\x53\x5c\x76\x79\x4d\xbf\x4c\x1f\x13\x62\x27\x1e\x2a\xc2\x61\x11\xe0\xf3\xe9\x72\xba\x38\xb7\x85\x65\xe6\x9b\x77\xbc\x50\x2e\x17\x1f\xcc\x7f\x4f\xed\x86\x99\xb9\x39\x31\xe6\xc5\x60\x5f\x63\xf2\x6b\x85\x24\x9c\xfd\xf6\x72\x3e\xe5\x5c\x31\x99\x4f\x8d\x05\x1f\x3c\x80\x7e\xac\x7b\x3a\x8b\x12\x08\x22\x1a\xeb\x74\xcb\x4a\x0b\x8e\x4b\xa0\x63\x1e\xce\xdc\x42\x27\x0c\x9f\x32\xf2\x02\x1d\x08\x26\xcb\xe5\xa1\xa8\xc9\xdc\x8b\xc9\x54\xae\xe5\x1f\xf9\x9a\x1f\x71\xe1\x58\xa7\x7b\x37\xf8\xc7\x77\xf8\x72\x50\x0d\x02\x0d\x35\x4a\x8d\xf9\x1c\x49\x7d\x39\x96\x0f\xb5\xf6\x4d\x2c\xee\xa3\xc1\xe8\xf4\x34\x12\x81\xc9\xaf\xa7\xe1\x0d\xbb\x61\xd7\x29\x8e\xe8\x0e\x3f\x91\x97\x8f\x6d\xa5\x06\xdf\x08\xe7\x84\xe1\x1e\xf1\x5a\x88\xe0\xef\x90\x46\xc2\xc1\x02\x61\x6d\x58\xc5\xcb\xb1\x86\xc0\x9c\xb4\x66\xce\x1d\x09\x43\xf0\x28\x74\xf8\x49\x39\x3c\xba\xd6\x74\x3e\x35\x56\x53\x19\xd9\xce\xc7\xb6\xa2\x6c\x70\x98\x0c\x0b\x5a\xd7\x70\x1e\x0e\xd4\xbe\x89\xdb\x2f\xf0\xfa\xc8\xfb\xa0\x61\xf1\xe6\x2a\x1f\x81\x77\x1b\xe2\x6e\x8a\x36\x89\xdf\x9a\x67\xf7\xfc\x34\x67\x98\x71\xa3\x38\xad\xe6\x4f\xfd\x99\x39\x18\x0c\xe4\x3b\x6b\x4c\x63\x71\xe4\xf3\x56\xc8\x5e\x5f\x2d\xed\xf3\x4b\x8b\xdf\x9f\xd1\x27\x86\xa3\x04\x34\x4f\x30\x32\xe7\x54\xb5\xd0\x3a\xbf\xb4\xb8\x2c\xa0\xa2\x97\x24\x31\x61\xc0\xa2\x84\x8b\x04\x27\x8a\x56\xf2\xca\x78\xb7\x7f\x2f\x23\xd0\x02\xd6\x20\x1f\xbc\x46\x25\x5d\x82\x44\x24\x71\xc6\x1c\xfe\xe6\x27\x6c\xb3\xbd\xd5\x5d\x1a\xf5\xd7\x4e\x9a\x4d\x42\xba\xf5\xc4\xd7\x8c\xf6\xb9\x8f\x7d\x46\x69\x98\xf5\x73\x77\xcf\x42\x87\x3f\x1a\x46\x95\xcb\x26\xa0\x10\x60\xf6\xe7\x71\x7c\xc2\xff\x2d\xf8\x54\x3d\x88\x90\x88\xe7\x83\x92\x4b\xc7\xcd\x7f\xca\x10\xcf\xd0\xa2\x6f\x2d\x7b\xc9\xaa\x7e\x3d\x67\x45\x0f\x50\x6b\x03\x9f\x4a\x74\x4d\xe9\x07\xd3\xbd\xb4\xcc\xb5\x39\xb1\x67\xd3\x85\xfd\x87\x49\x6f\xd4\x7d\xf9\xf4\x53\x1a\x84\x58\x0f\x68\xe8\xc4\x81\x4e\xd3\x40\x3e\xe7\x6f\xb7\x7e\xdf\x8d\xbc\x7e\xcd\xf9\xb3\xdd\x40\x3f\xf9\x4d\x7f\xd5\x94\xfd\x3f\x09\x73\x90\xfc\xc6\x18\xe4\x25\x50\x9b\x7b\xae\x10\x6a\x9b\x95\xfd\x61\xc3\xda\xa6\xfc\x36\x96\x46\xd1\xa0\x7e\x57\x89\x14\xca\x7f\xad\x54\x0c\x6b\x65\x2d\x27\x3f\x56\x32\xf5\x35\xff\x2f\xcc\xd1\x18\x9c\xbf\xcc\x23\xcf\xa2\xc2\x1f\x56\x56\x3d\x54\x4d\x15\x56\xe8\x7c\x77\xa5\x95\xcf\xc8\xc7\x15\xf7\x08\xeb\xb9\xb2\xc9\x2b\xf0\x7f\x01\x00\x00\xff\xff\xf9\x62\xf0\x5d\xb5\x16\x00\x00")
+var _scriptsMakefileTarsMk = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xb4\x57\x6d\x6f\xab\x38\x16\xfe\x0c\xbf\xe2\x2c\x43\xb7\x71\x6f\x20\xb7\xad\xb4\x5a\x25\x62\xef\xa5\x29\x4d\x3b\xc3\x84\x28\x49\x67\xb5\x6a\x2a\x86\x82\x49\xac\x01\xcc\x82\x93\xbe\x4d\xfe\xfb\xca\x76\x20\xa1\x4d\xbb\x73\xa5\x36\x9f\x9c\xe3\xf3\xf2\x9c\x63\x9f\xc7\x87\x9f\x8c\x8f\xfd\xa9\x3f\xc5\xe4\x01\xc2\x39\x85\x90\xa6\x39\x49\x30\xe0\xa2\xa0\x85\x8a\x1f\x72\x5a\x30\x70\xfb\xbe\xed\xba\x00\x60\x01\xce\xfc\xeb\x89\x79\x3d\xbd\x30\xfe\x59\xef\xda\xc3\x01\xc0\xab\xdd\x0f\xc7\xa8\x0e\xbc\x91\x3d\xbd\x84\x6f\x16\xe8\xad\x72\x81\x93\x04\xe6\x14\x70\xb6\x02\xb9\x81\xd4\x81\x37\xf6\xbc\xe9\x5e\x05\xbe\xc1\x15\x60\x83\x54\x7f\x96\xb2\x75\xe7\x8e\x64\x9d\x39\x55\x3f\x1e\x6f\x42\xee\xf2\x80\x2d\x2c\xfd\xb9\x5c\xde\x95\x0c\xba\x6d\x68\xeb\xad\x0d\xd6\xb5\x3a\xb5\xc7\x93\x93\x0d\xa0\x2e\x47\x1c\x93\xa2\x64\xf7\xb4\x88\x38\xfa\xda\x02\x6a\x13\x84\x04\x56\x16\x14\xe5\x09\x07\x4c\x62\xfc\x5f\x68\xb5\xb9\x61\x16\x95\xac\x20\xd9\x1c\x0c\xba\x64\x11\x29\xda\x7a\xeb\xe7\x93\x81\xe7\x5f\xb8\xf6\x00\x21\xa4\xf2\x18\xb5\x00\x00\xbe\x58\x95\xa6\xb5\xc2\x59\xc4\x4f\x3a\x8b\x48\xac\xaa\xa3\xb3\x0a\x92\xf2\x23\x98\xf2\x82\x32\x1a\x7e\x42\x0d\x45\x95\xfc\xc9\xb8\xbf\x2d\xd3\x3d\x49\xa2\x30\x28\x22\x38\x32\x79\x29\x90\x3a\x1a\x7b\xb5\xc6\x97\x17\x1a\x02\x18\x3f\xf7\x5a\xe3\x95\x93\x39\x45\x1f\x0f\x5c\x60\x0f\x69\xfe\x18\xf3\x66\xb2\x80\xc4\xc0\x70\xc9\xc0\x78\x02\x4d\x6f\xd9\xa3\x11\xd2\xe0\xcf\x3f\x77\x65\x53\x7b\x3c\x70\xa6\x48\xeb\x01\x5b\xe0\x0c\x66\x2a\x34\x7f\x38\x5c\x50\xd0\x6e\x0e\xed\xd1\xe8\x10\x68\x01\x87\xd2\xe0\x10\x68\xce\x08\xcd\x80\x94\x80\xd3\x9c\x3d\x9a\xb7\x5a\x0f\xf0\x03\x61\x70\xdc\x6b\x7a\x51\x70\x52\x62\x98\xa9\x8a\x52\x09\x2a\x54\x7f\x03\x83\x1f\xf0\x09\x7a\x23\xb8\xa2\x6c\xa2\x0f\x29\x94\xcb\x70\x01\x11\x29\xba\x5c\xbf\x0d\x19\xbd\x87\x7b\x0c\x61\x81\x03\x86\x81\x30\x1e\x5e\x9a\x2b\x8a\xa2\xa4\x7f\x44\xa4\x00\x23\x97\xce\x79\x68\x45\x89\x49\x03\x57\xed\xbb\x4f\xf3\x47\x10\xe5\xd2\x5b\xc7\x08\x8c\x7f\x09\x23\x9e\x4d\x43\x3b\xcc\xc1\x58\x49\x15\x89\x58\x78\x8d\x49\x4f\x55\x6d\xd7\xed\x42\x5d\xc9\x4f\x39\xd3\xda\x7b\x17\xf8\xcd\x3c\xbb\xbe\x72\xcf\x45\x37\xf0\xfb\x85\x54\x85\x2f\x11\xdc\x2d\x49\x12\x49\xb1\xd0\x90\x7d\x08\x06\x05\xfd\xfb\xe7\xdc\x35\x12\x67\x9c\x0c\xf4\x16\xe7\x81\x5c\x16\x41\xf4\x0d\x42\x6d\xa4\xd6\x50\xbb\xbb\x3b\xaa\xf2\x5d\x94\xde\xc0\xa0\xcd\xf0\xcd\xe9\x69\xef\x38\x95\xdb\x27\x03\x0f\xcd\xf0\xcd\xd7\x14\xb8\xfc\x1f\xbd\xe3\x14\xf4\xe7\xca\x6e\x0d\x62\xcb\x34\x4d\x8d\xe7\x5b\xe9\xc3\x2e\xe5\x34\xc2\xf0\x6b\xf7\x7f\x20\x68\x19\x05\xde\xcc\xe2\xfc\xb5\x8a\x8c\xea\xa4\x36\x1d\xce\x53\x19\x8d\xbd\xda\x4d\x25\xde\x97\x88\x64\xa4\x57\x49\x6c\x4c\x1a\x39\x48\x63\xbd\x25\xb8\x0f\x71\xa5\xb3\x2a\x91\x35\xe8\xad\x20\x8a\xf2\x02\xf3\xf7\xd1\x30\x84\x57\x5f\xb0\x3a\x27\xc1\x92\x3f\x7f\x7a\x8b\xdf\xf1\x2d\x1a\x84\x50\x03\x9b\xde\x8a\x69\x81\x83\x70\x21\x92\x6b\xef\x64\xa3\xb7\xf0\x2a\x48\xe0\xbd\xf0\x8d\x90\x55\x24\xee\x47\x04\x91\x0b\xa4\x2a\x31\x2d\x64\xeb\x90\x6c\x0b\xab\x06\xc4\x1b\x2f\xa2\xa2\x51\x22\x52\x64\x41\x8a\x2d\x5d\xf8\xe2\x4b\xd0\x75\xe1\x46\x74\xe7\x9b\x28\xe6\xd4\xa7\x4b\x66\xe5\xc9\x72\x4e\xb2\xd2\xe2\x67\x55\xe4\x61\x57\xd7\x2b\x2f\x4d\xa0\x7a\xd3\xb9\x04\x90\x61\x79\x15\xde\x3b\x42\x7e\x11\x84\xa3\xe6\x4d\xf8\x8c\x96\x61\x41\xb1\xc3\x16\xa0\xb7\xfa\xde\xf0\xe2\x6a\xc0\xa1\x90\x18\x6e\x24\x25\x6e\x76\x7d\x96\xe6\x3e\x2f\xfe\x6d\xcd\x90\x15\x6f\x71\xe9\x22\x28\x39\xe5\x96\xac\xfb\xca\xa2\x0d\xc1\x1d\x2d\x98\xa9\x49\xa2\xda\x12\xf3\x96\x8d\x25\x49\xbe\x8e\xf5\x4a\xd2\xa9\x25\xe2\xb0\x38\x13\x16\xf1\xbe\x14\xde\x33\xed\x48\x20\xe1\x9e\xec\x7a\xbc\x07\xc1\x30\xf0\x43\x98\x2c\x23\x6c\x69\x66\xb9\xca\xb4\x5d\x81\x2f\x05\xe1\xd3\x6a\x27\xae\xc9\xe6\x4f\xb0\xeb\x3f\x8c\xc0\x34\x65\x18\x59\xc9\x78\xe7\x81\xe3\xda\x5a\xa3\x8e\x4a\xba\x02\xe3\x6d\x87\xe6\xef\x11\x7f\x5a\xbe\x1c\xfc\xe7\x20\x3d\x88\x0e\x2e\x0f\x7e\x3d\x98\xfc\xce\x55\x64\x04\xf9\x9c\x08\x27\xef\x64\x2d\x5c\x9a\x9b\xdc\x8b\xb4\x59\xb8\x6d\xfe\xdb\x63\xe5\x95\x08\x57\xf1\xd3\x0b\x54\x9c\x31\x84\x5a\x4c\x36\xb3\xc9\xbf\x9d\x33\xff\xd2\x9b\x4c\x01\xf8\xdc\xb9\x60\x2c\xef\x76\x3a\x09\x0d\x83\x64\x41\x4b\xd6\x3d\xfd\xfa\xf5\xeb\x56\x71\xea\xfd\xe2\x0c\x85\xa2\xa6\xa9\xd7\x23\xd7\xb3\xcf\xfd\xeb\x89\x33\x16\x6f\xdb\xce\xd8\x7a\xbf\xa0\x41\x4a\x50\xa5\xe2\x4d\x36\xcf\xdf\x37\x0b\x12\x92\x2d\x1f\xd4\x65\x9e\xd0\x20\xea\xc2\x66\x04\x1f\x78\xde\xc4\xd2\x9f\x6b\xf5\x75\xad\xc0\x82\xa2\xee\xad\x66\x2a\x86\x61\x80\x1c\x45\xcc\xe6\x06\x78\x93\xae\x98\x8f\xbd\xc9\x5a\x53\x95\x70\x59\x24\x15\xfd\x57\xc9\xae\x3b\x41\x4e\x3a\x32\x86\x1f\x64\x91\x9f\x2f\xef\x12\x52\x2e\xbe\x31\x12\xfe\x81\x99\xb5\xa3\x2e\x52\x5e\x83\x71\x51\x2e\x4b\x6c\x7d\x17\x3b\x03\x67\xba\x96\x18\x2e\x82\x3c\x4f\x48\x18\xf0\x11\xc6\xd2\x9f\xed\xd1\x88\xab\xa6\x34\x5a\x26\xd8\x97\x4c\x55\x19\x80\x71\x11\xd2\x34\xc5\x19\xb3\x64\x60\x1c\x19\x77\x8f\x46\x9d\x35\xaf\xe3\xba\xce\x75\x96\x7d\x20\x57\xcc\x32\x4d\x55\x2f\x1d\x77\x24\xe7\x4b\xbe\xf2\xa7\xf6\x18\xed\x3c\x51\xf5\x93\xd6\x46\xaa\x3a\x71\xc6\xbf\x39\x63\x7f\x68\xff\xea\xc8\x71\xb3\x1a\x49\x36\x7c\xf6\x39\xd3\x66\x82\x83\xac\xab\x8a\xdb\x2d\xba\xe9\xdc\x19\x39\xc3\x73\x5f\x20\xf3\xce\x7e\xe6\xac\x70\x35\xfc\xcd\xfb\xc5\xf1\xf7\xec\xb8\x5e\xdf\x76\xab\x3f\x5b\x42\x91\xab\x09\x92\xee\xfc\x8b\x2b\xd7\xe1\x8f\x43\xdf\x75\xec\x21\xff\xb3\x06\xf3\xc8\x8c\x4c\x96\xe6\x30\x4f\x69\x66\xd2\x25\xab\x11\x1c\xd5\x8e\x8e\xf8\x61\x6f\x10\x06\x49\xf2\x16\xc8\x4b\xf4\x42\xd0\x1f\x8d\x5e\x8a\xf6\xc1\xbd\xec\x3b\xd2\x96\x2f\x36\x36\x3f\x92\xc2\x91\x49\xb7\x69\xf0\xc5\x5f\xc8\xe5\xc5\x6d\x10\xde\xd0\x5e\xa9\xed\xba\x48\xea\xcb\xff\xf2\xfb\x53\x9b\x65\x33\x7c\x73\xdc\x3b\x3d\x4d\x45\x59\xc4\x3c\xd2\x9d\xb1\x19\xbb\x29\x70\x4a\x57\xf8\x8d\x33\xb9\xd5\xd4\x86\xeb\x3d\xae\x82\x24\xa9\xbc\xdd\x08\x01\xfc\x1d\x8a\x54\x24\x56\x59\x4f\xed\x71\xf5\x21\xbc\x63\xcd\x82\x62\x07\xc6\x3d\x49\x12\x88\x28\x1c\x72\x2a\x7c\xc9\xcd\x63\xc7\x75\xec\x89\x23\xab\x79\x78\xab\xa9\xea\x02\x27\x79\x77\x67\x12\x93\x25\x40\x9a\xe8\x9d\x0f\xff\x2e\xfc\x5f\x00\x00\x00\xff\xff\xf4\xe3\xf5\x30\x01\x11\x00\x00")
 
-func scriptsMakefileTarsGomodBytes() ([]byte, error) {
+func scriptsMakefileTarsMkBytes() ([]byte, error) {
 	return bindataRead(
-		_scriptsMakefileTarsGomod,
-		"scripts/makefile.tars.gomod",
+		_scriptsMakefileTarsMk,
+		"scripts/makefile.tars.mk",
 	)
 }
 
-func scriptsMakefileTarsGomod() (*asset, error) {
-	bytes, err := scriptsMakefileTarsGomodBytes()
+func scriptsMakefileTarsMk() (*asset, error) {
+	bytes, err := scriptsMakefileTarsMkBytes()
 	if err != nil {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "scripts/makefile.tars.gomod", size: 5813, mode: os.FileMode(493), modTime: time.Unix(1661769677, 0)}
+	info := bindataFileInfo{name: "scripts/makefile.tars.mk", size: 4353, mode: os.FileMode(493), modTime: time.Unix(1686929967, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
@@ -304,19 +304,21 @@ var _bindata = map[string]func() (*asset, error){
 	"cmake/CMakeTestGoCompiler.cmake":      cmakeCmaketestgocompilerCmake,
 	"cmake/golang.cmake":                   cmakeGolangCmake,
 	"cmake/tars-tools.cmake":               cmakeTarsToolsCmake,
-	"scripts/makefile.tars":                scriptsMakefileTars,
-	"scripts/makefile.tars.gomod":          scriptsMakefileTarsGomod,
+	"scripts/makefile.tars.gomod.mk":       scriptsMakefileTarsGomodMk,
+	"scripts/makefile.tars.mk":             scriptsMakefileTarsMk,
 }
 
 // AssetDir returns the file names below a certain
 // directory embedded in the file by go-bindata.
 // For example if you run go-bindata on data/... and data contains the
 // following hierarchy:
-//     data/
-//       foo.txt
-//       img/
-//         a.png
-//         b.png
+//
+//	data/
+//	  foo.txt
+//	  img/
+//	    a.png
+//	    b.png
+//
 // then AssetDir("data") would return []string{"foo.txt", "img"}
 // AssetDir("data/img") would return []string{"a.png", "b.png"}
 // AssetDir("foo.txt") and AssetDir("nonexistent") would return an error
@@ -358,8 +360,8 @@ var _bintree = &bintree{nil, map[string]*bintree{
 		"tars-tools.cmake":               &bintree{cmakeTarsToolsCmake, map[string]*bintree{}},
 	}},
 	"scripts": &bintree{nil, map[string]*bintree{
-		"makefile.tars":       &bintree{scriptsMakefileTars, map[string]*bintree{}},
-		"makefile.tars.gomod": &bintree{scriptsMakefileTarsGomod, map[string]*bintree{}},
+		"makefile.tars.gomod.mk": &bintree{scriptsMakefileTarsGomodMk, map[string]*bintree{}},
+		"makefile.tars.mk":       &bintree{scriptsMakefileTarsMk, map[string]*bintree{}},
 	}},
 }}
 
