@@ -2,10 +2,10 @@ package consts
 
 // Release is the current tarsgo tool version.
 const (
-	RepoURL      = "https://github.com/TarsCloud/TarsGo.git"
-	Release      = "v1.3.4"
-	MakeDemoDir  = "Demo4GoMod"
-	CMakeDemoDir = "Demo4Cmake"
-	Make         = "make"
-	CMake        = "cmake"
+	RepoURL     = "https://github.com/TarsCloud/TarsGo.git"
+	Release     = "v1.3.6"
+	MakeDemoDir = "Demo4GoMod"
+	//CMakeDemoDir = "Demo4Cmake"
+	Make  = "make"
+	CMake = "cmake"
 )
