@@ -7,95 +7,113 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/TarsCloud/TarsGo/tars/util/gpool"
+
 	"github.com/TarsCloud/TarsGo/tars/protocol/res/basef"
 	"github.com/TarsCloud/TarsGo/tars/util/current"
 	"github.com/TarsCloud/TarsGo/tars/util/grace"
 )
 
 type udpHandler struct {
-	conf *TarsServerConf
-	ts   *TarsServer
+	config *TarsServerConf
+	server *TarsServer
 
 	conn *net.UDPConn
+	pool *gpool.Pool
 }
 
-func (h *udpHandler) Listen() (err error) {
-	cfg := h.conf
-	h.conn, err = grace.CreateUDPConn(cfg.Address)
+func (u *udpHandler) Listen() (err error) {
+	cfg := u.config
+	u.conn, err = grace.CreateUDPConn(cfg.Address)
 	if err != nil {
 		return err
 	}
-	TLOG.Info("UDP listen", h.conn.LocalAddr())
+	TLOG.Info("UDP listen", u.conn.LocalAddr())
+
+	// init goroutine pool
+	if cfg.MaxInvoke > 0 {
+		u.pool = gpool.NewPool(int(cfg.MaxInvoke), cfg.QueueCap)
+	}
 	return nil
 }
 
-func (h *udpHandler) getConnContext(udpAddr *net.UDPAddr) context.Context {
+func (u *udpHandler) getConnContext(udpAddr *net.UDPAddr) context.Context {
 	ctx := current.ContextWithTarsCurrent(context.Background())
 	current.SetClientIPWithContext(ctx, udpAddr.IP.String())
 	current.SetClientPortWithContext(ctx, strconv.Itoa(udpAddr.Port))
 	current.SetRecvPkgTsFromContext(ctx, time.Now().UnixNano()/1e6)
-	current.SetRawConnWithContext(ctx, h.conn, udpAddr)
+	current.SetRawConnWithContext(ctx, u.conn, udpAddr)
 	return ctx
 }
 
-func (h *udpHandler) Handle() error {
-	atomic.AddInt32(&h.ts.numConn, 1)
+func (u *udpHandler) handleUDPAddr(udpAddr *net.UDPAddr, pkg []byte) {
+	ctx := u.getConnContext(udpAddr)
+	atomic.AddInt32(&u.server.numInvoke, 1)
+	handler := func() {
+		defer atomic.AddInt32(&u.server.numInvoke, -1)
+		rsp := u.server.invoke(ctx, pkg) // no need to check package
+
+		cPacketType, ok := current.GetPacketTypeFromContext(ctx)
+		if !ok {
+			TLOG.Error("Failed to GetPacketTypeFromContext")
+		}
+
+		if cPacketType == basef.TARSONEWAY {
+			return
+		}
+
+		if _, err := u.conn.WriteToUDP(rsp, udpAddr); err != nil {
+			TLOG.Errorf("send pkg to %v failed %v", udpAddr, err)
+		}
+	}
+
+	cfg := u.config
+	if cfg.MaxInvoke > 0 { // use goroutine pool
+		u.pool.JobQueue <- handler
+	} else {
+		go handler()
+	}
+}
+
+func (u *udpHandler) Handle() error {
+	atomic.AddInt32(&u.server.numConn, 1)
 	// wait invoke done
 	defer func() {
 		tick := time.NewTicker(time.Second)
 		defer tick.Stop()
-		for atomic.LoadInt32(&h.ts.numInvoke) > 0 {
+		for atomic.LoadInt32(&u.server.numInvoke) > 0 {
 			<-tick.C
 		}
-		atomic.AddInt32(&h.ts.numConn, -1)
+		atomic.AddInt32(&u.server.numConn, -1)
 	}()
 	buffer := make([]byte, 65535)
 	for {
-		if atomic.LoadInt32(&h.ts.isClosed) == 1 {
+		if atomic.LoadInt32(&u.server.isClosed) == 1 {
 			return nil
 		}
-		n, udpAddr, err := h.conn.ReadFromUDP(buffer)
+		n, udpAddr, err := u.conn.ReadFromUDP(buffer)
 		if err != nil {
-			if atomic.LoadInt32(&h.ts.isClosed) == 1 {
+			if atomic.LoadInt32(&u.server.isClosed) == 1 {
 				return nil
 			}
 			if isNoDataError(err) {
 				continue
-			} else {
-				TLOG.Errorf("Close connection %s: %v", h.conf.Address, err)
-				return err // TODO: check if necessary
 			}
+			TLOG.Errorf("Close connection %s: %v", u.config.Address, err)
+			return err // TODO: check if necessary
 		}
 		pkg := make([]byte, n)
 		copy(pkg, buffer[0:n])
-		ctx := h.getConnContext(udpAddr)
-		go func() {
-			atomic.AddInt32(&h.ts.numInvoke, 1)
-			defer atomic.AddInt32(&h.ts.numInvoke, -1)
-			rsp := h.ts.invoke(ctx, pkg) // no need to check package
-
-			cPacketType, ok := current.GetPacketTypeFromContext(ctx)
-			if !ok {
-				TLOG.Error("Failed to GetPacketTypeFromContext")
-			}
-
-			if cPacketType == basef.TARSONEWAY {
-				return
-			}
-
-			if _, err := h.conn.WriteToUDP(rsp, udpAddr); err != nil {
-				TLOG.Errorf("send pkg to %v failed %v", udpAddr, err)
-			}
-		}()
+		u.handleUDPAddr(udpAddr, pkg)
 	}
 }
 
-func (h *udpHandler) OnShutdown() {
+func (u *udpHandler) OnShutdown() {
 }
 
-func (h *udpHandler) CloseIdles(n int64) bool {
-	if h.ts.numInvoke == 0 {
-		h.conn.Close()
+func (u *udpHandler) CloseIdles(_ int64) bool {
+	if u.server.numInvoke == 0 {
+		u.conn.Close()
 		return true
 	}
 	return false
